add_edges {igraph}	R Documentation

Add edges to a graph

Description

The new edges are given as a vertex sequence, e.g. internal numeric vertex ids, or vertex names. The first edge points from edges[1] to edges[2], the second from edges[3] to edges[4], etc.

Usage

add_edges(graph, edges, ..., attr = list())

Arguments

graph
The input graph

edges
The edges to add, a vertex sequence with even number of vertices.

...
Additional arguments, they must be named, and they will be added as edge attributes, for the newly added edges. See also details below.

attr
A named list, its elements will be added as edge attributes, for the newly added edges. See also details below.

Details

If attributes are supplied, and they are not present in the graph, their values for the original edges of the graph are set to NA.

Value

The graph, with the edges (and attributes) added.

See Also

Other functions for manipulating graph structure: +.igraph, add_vertices, delete_edges, delete_vertices, edge, igraph-minus, path, vertex

Examples

g <- make_empty_graph(n = 5) %>%
  add_edges(c(1,2, 2,3, 3,4, 4,5)) %>%
  set_edge_attr("color", value = "red") %>%
  add_edges(c(5,1), color = "green")
E(g)[[]]
plot(g)

[Package igraph version 1.2.2 Index]