--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.2 --- {% raw %}

add_layout_ {igraph}	R Documentation

Add layout to graph

Description

Add layout to graph

Usage

add_layout_(graph, ..., overwrite = TRUE)

Arguments

graph
The input graph.

...
Additional arguments are passed to layout_.

overwrite
Whether to overwrite the layout of the graph, if it already has one.

Value

The input graph, with the layout added.

See Also

layout_ for a description of the layout API.

Other graph layouts: component_wise(), layout_as_bipartite(), layout_as_star(), layout_as_tree(), layout_in_circle(), layout_nicely(), layout_on_grid(), layout_on_sphere(), layout_randomly(), layout_with_dh(), layout_with_fr(), layout_with_gem(), layout_with_graphopt(), layout_with_kk(), layout_with_lgl(), layout_with_mds(), layout_with_sugiyama(), layout_(), merge_coords(), norm_coords(), normalize()

Examples

(make_star(11) + make_star(11)) %>%
  add_layout_(as_star(), component_wise()) %>%
  plot()

[Package igraph version 1.3.2 Index]
{% endraw %}