articulation.points {igraph}	R Documentation

Articulation points of a graph

Description

Articuation points or cut vertices are vertices whose removal increases the number of connected components in a graph.

Usage

articulation.points(graph)

Arguments

graph 	 The input graph. It is treated as an undirected graph, even if it is directed.

Details

Articuation points or cut vertices are vertices whose removal increases the number of connected components in a graph.

Value

A numeric vector giving the vertex ids of the articulation points of the input graph.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

biconnected.components, clusters, is.connected, vertex.connectivity

Examples

g <- graph.disjoint.union( graph.full(5), graph.full(5) )
clu <- clusters(g)$membership
g <- add.edges(g, c(which(clu==0), which(clu==1))-1)
articulation.points(g)

[Package igraph version 0.5.2-2 Index]