articulation_points {igraph}	R Documentation

Articulation points of a graph

Description

Articuation points or cut vertices are vertices whose removal increases the number of connected components in a graph.

Usage

articulation_points(graph)

Arguments

graph
The input graph. It is treated as an undirected graph, even if it is directed.

Details

Articuation points or cut vertices are vertices whose removal increases the number of connected components in a graph. If the original graph was connected, then the removal of a single articulation point makes it undirected. If a graph contains no articulation points, then its vertex connectivity is at least two.

Value

A numeric vector giving the vertex ids of the articulation points of the input graph.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

biconnected_components, components, is_connected, vertex_connectivity

Examples

g <- disjoint_union( make_full_graph(5), make_full_graph(5) )
clu <- components(g)$membership
g <- add_edges(g, c(match(1, clu), match(2, clu)) )
articulation_points(g)

[Package igraph version 1.2.2 Index]