Find Bonacich Power Centrality Scores of Network Positions

bonpow() was renamed to power_centrality() to create a more consistent API.

Usage

bonpow(
  graph,
  nodes = V(graph),
  loops = FALSE,
  exponent = 1,
  rescale = FALSE,
  tol = 1e-07,
  sparse = TRUE
)

Arguments

graph

the input graph.

nodes

vertex sequence indicating which vertices are to be included in the calculation. By default, all vertices are included.

loops

boolean indicating whether or not the diagonal should be treated as valid data. Set this true if and only if the data can contain loops. loops is FALSE by default.

exponent

exponent (decay rate) for the Bonacich power centrality score; can be negative

rescale

if true, centrality scores are rescaled such that they sum to 1.

tol

tolerance for near-singularities during matrix inversion (see solve())

sparse

Logical scalar, whether to use sparse matrices for the calculation. The ‘Matrix’ package is required for sparse matrix support