categorical_pal {igraph}	R Documentation

Palette for categories

Description

This is a color blind friendly palette from http://jfly.iam.u-tokyo.ac.jp/color. It has 8 colors.

Usage

categorical_pal(n)

Arguments

n
The number of colors in the palette. We simply take the first n colors from the total 8.

Details

This is the suggested palette for visualizations where vertex colors mark categories, e.g. community membership.

Value

A character vector of RGB color codes.

Examples

library(igraphdata)
data(karate)
karate <- karate 
  add_layout_(with_fr()) 
  set_vertex_attr("size", value = 10)

cl_k <- cluster_optimal(karate)

V(karate)$color <- membership(cl_k)
karate$palette <- categorical_pal(length(cl_k))
plot(karate)

See Also

Other palettes: diverging_pal, r_pal, sequential_pal

[Package igraph version 1.2.2 Index]