--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.4 --- {% raw %}

centr_betw {igraph}	R Documentation

Centralize a graph according to the betweenness of vertices

Description

See centralize for a summary of graph centralization.

Usage

centr_betw(graph, directed = TRUE, nobigint = TRUE, normalized = TRUE)

Arguments

graph
The input graph.

directed
logical scalar, whether to use directed shortest paths for calculating betweenness.

nobigint
Logical scalar, whether to use big integers for the betweenness calculation. This argument is deprecated in igraph 1.3 and will be removed in igraph 1.4.

normalized
Logical scalar. Whether to normalize the graph level centrality score by dividing by the theoretical maximum.

Value

A named list with the following components:

res
The node-level centrality scores.

centralization
The graph level centrality index.

theoretical_max
The maximum theoretical graph level centralization score for a graph with the given number of vertices, using the same parameters. If the normalized argument was TRUE, then the result was divided by this number.

See Also

Other centralization related: centr_betw_tmax(), centr_clo_tmax(), centr_clo(), centr_degree_tmax(), centr_degree(), centr_eigen_tmax(), centr_eigen(), centralize()

Examples

# A BA graph is quite centralized
g <- sample_pa(1000, m = 4)
centr_degree(g)$centralization
centr_clo(g, mode = "all")$centralization
centr_betw(g, directed = FALSE)$centralization
centr_eigen(g, directed = FALSE)$centralization

[Package igraph version 1.3.4 Index]
{% endraw %}