--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.1 --- {% raw %}

centr_betw_tmax {igraph}	R Documentation

Theoretical maximum for betweenness centralization

Description

See centralize for a summary of graph centralization.

Usage

centr_betw_tmax(graph = NULL, nodes = 0, directed = TRUE)

Arguments

graph
The input graph. It can also be NULL, if nodes is given.

nodes
The number of vertices. This is ignored if the graph is given.

directed
logical scalar, whether to use directed shortest paths for calculating betweenness.

Value

Real scalar, the theoretical maximum (unnormalized) graph betweenness centrality score for graphs with given order and other parameters.

See Also

Other centralization related: centr_betw(), centr_clo_tmax(), centr_clo(), centr_degree_tmax(), centr_degree(), centr_eigen_tmax(), centr_eigen(), centralize()

Examples

# A BA graph is quite centralized
g <- sample_pa(1000, m = 4)
centr_betw(g, normalized = FALSE)$centralization %>%
 `/`(centr_betw_tmax(g))
centr_betw(g, normalized = TRUE)$centralization

[Package igraph version 1.3.1 Index]
{% endraw %}