--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.2 --- {% raw %}

centr_clo_tmax {igraph}	R Documentation

Theoretical maximum for closeness centralization

Description

See centralize for a summary of graph centralization.

Usage

centr_clo_tmax(graph = NULL, nodes = 0, mode = c("out", "in", "all", "total"))

Arguments

graph
The input graph. It can also be NULL, if nodes is given.

nodes
The number of vertices. This is ignored if the graph is given.

mode
This is the same as the mode argument of closeness.

Value

Real scalar, the theoretical maximum (unnormalized) graph closeness centrality score for graphs with given order and other parameters.

See Also

Other centralization related: centr_betw_tmax(), centr_betw(), centr_clo(), centr_degree_tmax(), centr_degree(), centr_eigen_tmax(), centr_eigen(), centralize()

Examples

# A BA graph is quite centralized
g <- sample_pa(1000, m = 4)
centr_clo(g, normalized = FALSE)$centralization %>%
 `/`(centr_clo_tmax(g))
centr_clo(g, normalized = TRUE)$centralization

[Package igraph version 1.3.2 Index]
{% endraw %}