centr_degree_tmax {igraph}	R Documentation

Theoretical maximum for degree centralization

Description

See centralize for a summary of graph centralization.

Usage

centr_degree_tmax(graph = NULL, nodes = 0, mode = c("all", "out", "in",
  "total"), loops = FALSE)

Arguments

graph
The input graph. It can also be NULL, if nodes, mode and loops are all given.

nodes
The number of vertices. This is ignored if the graph is given.

mode
This is the same as the mode argument of degree.

loops
Logical scalar, whether to consider loops edges when calculating the degree.

Value

Real scalar, the theoratical maximum (unnormalized) graph degree centrality score for graphs with given order and other parameters.

See Also

Other centralization related: centr_betw_tmax, centr_betw, centr_clo_tmax, centr_clo, centr_degree, centr_eigen_tmax, centr_eigen, centralize

Examples

# A BA graph is quite centralized
g <- sample_pa(1000, m = 4)
centr_degree(g, normalized = FALSE)$centralization %>%
 `/`(centr_degree_tmax(g))
centr_degree(g, normalized = TRUE)$centralization

[Package igraph version 1.2.2 Index]