Random citation graphs

citing.cited.type.game() was renamed to sample_cit_cit_types() to create a more consistent API.

Usage

citing.cited.type.game(
  n,
  edges = 1,
  types = rep(0, n),
  pref = matrix(1, nrow = length(types), ncol = length(types)),
  directed = TRUE,
  attr = TRUE
)

Arguments

n

Number of vertices.

edges

Number of edges per step.

types

Vector of length ‘n’, the types of the vertices. Types are numbered from zero.

pref

Vector (sample_last_cit() and sample_cit_types() or matrix (sample_cit_cit_types()) giving the (unnormalized) citation probabilities for the different vertex types.

directed

Logical scalar, whether to generate directed networks.

attr

Logical scalar, whether to add the vertex types to the generated graph as a vertex attribute called ‘type’.