complementer {igraph}	R Documentation

Complementer of a graph

Description

A complementer graph contains all edges that were not present in the input graph.

Usage

complementer(graph, loops = FALSE)

Arguments

graph
The input graph, can be directed or undirected.

loops
Logical constant, whether to generate loop edges.

Details

complementer creates the complementer of a graph. Only edges which are not present in the original graph will be included in the new graph.

complementer keeps graph and vertex attriubutes, edge attributes are lost.

Value

A new graph object.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

Examples

## Complementer of a ring
g <- make_ring(10)
complementer(g)

## A graph and its complementer give together the full graph
g <- make_ring(10)
gc <- complementer(g)
gu <- union(g, gc)
gu
graph.isomorphic(gu, make_full_graph(vcount(g)))

[Package igraph version 1.2.2 Index]