--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.1 --- {% raw %}

component_wise {igraph}	R Documentation

Component-wise layout

Description

This is a layout modifier function, and it can be used to calculate the layout separately for each component of the graph.

Usage

component_wise(merge_method = "dla")

Arguments

merge_method
Merging algorithm, the method argument of merge_coords.

See Also

merge_coords, layout_.

Other layout modifiers: normalize()

Other graph layouts: add_layout_(), layout_as_bipartite(), layout_as_star(), layout_as_tree(), layout_in_circle(), layout_nicely(), layout_on_grid(), layout_on_sphere(), layout_randomly(), layout_with_dh(), layout_with_fr(), layout_with_gem(), layout_with_graphopt(), layout_with_kk(), layout_with_lgl(), layout_with_mds(), layout_with_sugiyama(), layout_(), merge_coords(), norm_coords(), normalize()

Examples

g <- make_ring(10) + make_ring(10)
g %>%
  add_layout_(in_circle(), component_wise()) %>%
  plot()

[Package igraph version 1.3.1 Index]
{% endraw %}