count_motifs {igraph}	R Documentation

Graph motifs

Description

Graph motifs are small connected subgraphs with a well-defined structure. These functions search a graph for various motifs.

Usage

count_motifs(graph, size = 3, cut.prob = rep(0, size))

Arguments

graph
Graph object, the input graph.

size
The size of the motif, currently 3 and 4 are supported only.

cut.prob
Numeric vector giving the probabilities that the search graph is cut at a certain level. Its length should be the same as the size of the motif (the size argument). By default no cuts are made.

Details

count_motifs calculates the total number of motifs of a given size in graph.

Value

count_motifs returns a numeric scalar.

See Also

isomorphism_class

Other graph motifs: motifs, sample_motifs

Examples

g <- barabasi.game(100)
motifs(g, 3)
count_motifs(g, 3)
sample_motifs(g, 3)

[Package igraph version 1.2.2 Index]