decompose.graph {igraph}	R Documentation

Decompose a graph into components

Description

Creates a separate graph for each component of a graph.

Usage

decompose.graph(graph, mode = c("weak", "strong"),
       max.comps = NA, min.vertices = 0) 

Arguments

graph 	 The original graph.

mode 	 Character constant giving the type of the components, wither weak for weakly connected components or strong for strongly connected components.

max.comps 	 The maximum number of components to return. The first max.comps components will be returned (which hold at least min.vertices vertices, see the next parameter), the others will be ignored. Supply NA here if you don't want to limit the number of components.

min.vertices 	 The minimum number of vertices a component should contain in order to place it in the result list. Eg. supply 2 here to ignore isolate vertices.

Value

A list of graph objects.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

is.connected to decide whether a graph is connected, clusters to calculate the connected components of a graph.

Examples

# the diameter of each component in a random graph
g <- erdos.renyi.game(1000, 1/1000)
comps <- decompose.graph(g, min.vertices=2)
sapply(comps, diameter)

[Package igraph version 0.5.2-2 Index]