--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.5 --- {% raw %}

delete_edges {igraph}	R Documentation

Delete edges from a graph

Description

Delete edges from a graph

Usage

delete_edges(graph, edges)

Arguments

graph
The input graph.

edges
The edges to remove, specified as an edge sequence. Typically this is either a numeric vector containing edge IDs, or a character vector containing the IDs or names of the source and target vertices, separated by |

Value

The graph, with the edges removed.

See Also

Other functions for manipulating graph structure: +.igraph(), add_edges(), add_vertices(), delete_vertices(), edge(), igraph-minus, path(), vertex()

Examples

g <- make_ring(10) %>%
  delete_edges(seq(1, 9, by = 2))
g

g <- make_ring(10) %>%
  delete_edges("10|1")
g

g <- make_ring(5)
g <- delete_edges(g, get.edge.ids(g, c(1,5, 4,5)))
g

[Package igraph version 1.3.5 Index]
{% endraw %}