--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.5 --- {% raw %}

.apply_modifiers {igraph}	R Documentation

Applies a set of constructor modifiers to an already constructed graph.

Description

This is a helper function for the common parts of make_ and sample_.

Usage

.apply_modifiers(graph, mods)

Arguments

graph
The graph to apply the modifiers to

mods
The modifiers to apply

Value

The modified graph

[Package igraph version 1.3.5 Index]
{% endraw %}