--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.4 --- {% raw %}

.extract_constructor_and_modifiers {igraph}	R Documentation

Takes an argument list and extracts the constructor specification and constructor modifiers from it.

Description

This is a helper function for the common parts of make_ and sample_.

Usage

.extract_constructor_and_modifiers(..., .operation, .variant)

Arguments

...
Parameters to extract from

.operation
Human-readable description of the operation that this helper is a part of

.variant
Constructor variant; must be one of ‘make’, ‘graph’ or ‘sample’. Used in cases when the same constructor specification has deterministic and random variants.

Value

A named list with three items: ‘cons’ for the constructor function, ‘mods’ for the modifiers and ‘args’ for the remaining, unparsed arguments.

[Package igraph version 1.3.4 Index]
{% endraw %}