each_edge {igraph}	R Documentation

Rewires the endpoints of the edges of a graph to a random vertex

Description

This function can be used together with rewire. This method rewires the endpoints of the edges with a constant probability uniformly randomly to a new vertex in a graph.

Usage

each_edge(prob, loops = FALSE, multiple = FALSE)

Arguments

prob
The rewiring probability, a real number between zero and one.

loops
Logical scalar, whether loop edges are allowed in the rewired graph.

multiple
Logical scalar, whether multiple edges are allowed int the generated graph.

Details

Note that this method might create graphs with multiple and/or loop edges.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

Other rewiring functions: keeping_degseq, rewire

Examples

# Some random shortcuts shorten the distances on a lattice
g <- make_lattice(length = 100, dim = 1, nei = 5)
mean_distance(g)
g <- rewire(g, each_edge(prob = 0.05))
mean_distance(g)

[Package igraph version 1.2.2 Index]