eccentricity {igraph}	R Documentation

Eccentricity of the vertices in a graph

Description

The eccentricity of a vertex is its shortest path distance from the farthest other node in the graph.

Usage

eccentricity(graph, vids = V(graph), mode = c("all", "out", "in", "total"))

Arguments

graph
The input graph, it can be directed or undirected.

vids
The vertices for which the eccentricity is calculated.

mode
Character constant, gives whether the shortest paths to or from the given vertices should be calculated for directed graphs. If out then the shortest paths from the vertex, if in then to it will be considered. If all, the default, then the corresponding undirected graph will be used, edge directions will be ignored. This argument is ignored for undirected graphs.

Details

The eccentricity of a vertex is calculated by measuring the shortest distance from (or to) the vertex, to (or from) all vertices in the graph, and taking the maximum.

This implementation ignores vertex pairs that are in different components. Isolate vertices have eccentricity zero.

Value

eccentricity returns a numeric vector, containing the eccentricity score of each given vertex.

References

Harary, F. Graph Theory. Reading, MA: Addison-Wesley, p. 35, 1994.

See Also

radius for a related concept, distances for general shortest path calculations.

Examples

g <- make_star(10, mode="undirected")
eccentricity(g)

[Package igraph version 1.2.2 Index]