edge {igraph}	R Documentation

Helper function for adding and deleting edges

Description

This is a helper function that simplifies adding and deleting edges to/from graphs.

Usage

edge(...)

edges(...)

Arguments

...
See details below.

Details

edges is an alias for edge.

When adding edges via +, all unnamed arguments of edge (or edges) are concatenated, and then passed to add_edges. They are interpreted as pairs of vertex ids, and an edge will added between each pair. Named arguments will be used as edge attributes for the new edges.

When deleting edges via -, all arguments of edge (or edges) are concatenated via c() and passed to delete_edges.

Value

A special object that can be used with together with igraph graphs and the plus and minus operators.

See Also

Other functions for manipulating graph structure: +.igraph, add_edges, add_vertices, delete_edges, delete_vertices, igraph-minus, path, vertex

Examples

g <- make_ring(10) %>%
  set_edge_attr("color", value = "red")

g <- g + edge(1, 5, color = "green") +
  edge(2, 6, color = "blue") -
  edge("8|9")

E(g)[[]]

g %>%
  add_layout_(in_circle()) %>%
  plot()

g <- make_ring(10) + edges(1:10)
plot(g)

[Package igraph version 1.2.2 Index]