edge_attr {igraph}	R Documentation

Query edge attributes of a graph

Description

Query edge attributes of a graph

Usage

edge_attr(graph, name, index = E(graph))

Arguments

graph
The graph

name
The name of the attribute to query. If missing, then all edge attributes are returned in a list.

index
An optional edge sequence, to query edge attributes for a subset of edges.

Value

The value of the edge attribute, or the list of all edge attributes if name is missing.

See Also

Other graph attributes: delete_edge_attr, delete_graph_attr, delete_vertex_attr, edge_attr<-, edge_attr_names, graph_attr<-, graph_attr_names, graph_attr, igraph-dollar, igraph-vs-attributes, set_edge_attr, set_graph_attr, set_vertex_attr, vertex_attr<-, vertex_attr_names, vertex_attr

Examples

g <- make_ring(10) %>%
  set_edge_attr("weight", value = 1:10) %>%
  set_edge_attr("color", value = "red")
g
plot(g, edge.width = E(g)$weight)

[Package igraph version 1.2.2 Index]