edge_attr<- {igraph}	R Documentation

Set one or more edge attributes

Description

Set one or more edge attributes

Usage

edge_attr(graph, name, index = E(graph)) <- value

Arguments

graph
The graph.

name
The name of the edge attribute to set. If missing, then value must be a named list, and its entries are set as edge attributes.

index
An optional edge sequence to set the attributes of a subset of edges.

value
The new value of the attribute(s) for all (or index) edges.

Value

The graph, with the edge attribute(s) added or set.

See Also

Other graph attributes: $.igraph.vs, $<-.igraph.vs, V<-, [<-.igraph.vs, [[<-.igraph.vs, igraph-vs-attributes, igraph-vs-attributes, igraph-vs-attributes, igraph-vs-attributes, igraph-vs-attributes; $.igraph, $<-.igraph, igraph-dollar, igraph-dollar; attributes, graph_attr_names, list.graph.attributes; delete_edge_attr, remove.edge.attribute; delete_graph_attr, remove.graph.attribute; delete_vertex_attr, remove.vertex.attribute; edge.attributes, edge_attr, get.edge.attribute; edge_attr_names, list.edge.attributes; get.graph.attribute, graph.attributes, graph_attr; get.vertex.attribute, vertex.attributes, vertex_attr; graph.attributes<-, graph_attr<-; list.vertex.attributes, vertex_attr_names; set.edge.attribute, set_edge_attr; set.graph.attribute, set_graph_attr; set.vertex.attribute, set_vertex_attr; vertex.attributes<-, vertex_attr<-

Examples

g <- make_ring(10)
edge_attr(g) <- list(name = LETTERS[1:10],
                      color = rep("green", gsize(g)))
edge_attr(g, "label") <- E(g)$name
g
plot(g)

[Package igraph version 1.0.1 Index]