--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.1 --- {% raw %}

ends {igraph}	R Documentation

Incident vertices of some graph edges

Description

Incident vertices of some graph edges

Usage

ends(graph, es, names = TRUE)

Arguments

graph
The input graph

es
The sequence of edges to query

names
Whether to return vertex names or numeric vertex ids. By default vertex names are used.

Value

A two column matrix of vertex names or vertex ids.

See Also

Other structural queries: [.igraph(), [[.igraph(), adjacent_vertices(), are_adjacent(), get.edge.ids(), gorder(), gsize(), head_of(), incident_edges(), incident(), is_directed(), neighbors(), tail_of()

Examples

g <- make_ring(5)
ends(g, E(g))

[Package igraph version 1.3.1 Index]
{% endraw %}