erdos.renyi.game {igraph}	R Documentation

Generate random graphs according to the Erdos-Renyi model

Description

This model is very simple, every possible edge is created with the same constant probability.

Usage

erdos.renyi.game(n, p.or.m, type=c("gnp", "gnm"),
                 directed = FALSE, loops = FALSE, ...)

Arguments

n 	 The number of vertices in the graph.

p.or.m 	 Either the probability for drawing an edge between two arbitrary vertices (G(n,p) graph), or the number of edges in the graph (for G(n,m) graphs).

type 	 The type of the random graph to create, either gnp (G(n,p) graph) or gnm (G(n,m) graph).

directed 	 Logical, whether the graph will be directed, defaults to FALSE.

loops 	 Logical, whether to add loop edges, defaults to FALSE.

... 	 Additional arguments, ignored.

Details

In G(n,p) graphs, the graph has ‘n’ vertices and for each edge the probability that it is present in the graph is ‘p’.

In G(n,m) graphs, the graph has ‘n’ vertices and ‘m’ edges, and the ‘m’ edges are chosen uniformly randomly from the set of all possible edges. This set includes loop edges as well if the loops parameter is TRUE.

random.graph.game is an alias to this function.

Value

A graph object.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

References

Erdos, P. and Renyi, A., On random graphs, Publicationes Mathematicae 6, 290–297 (1959).

See Also

barabasi.game

Examples

g <- erdos.renyi.game(1000, 1/1000)
degree.distribution(g)

[Package igraph version 0.5.2-2 Index]