Find Eigenvector Centrality Scores of Network Positions

evcent() was renamed to eigen_centrality() to create a more consistent API.

Usage

evcent(
  graph,
  directed = FALSE,
  scale = TRUE,
  weights = NULL,
  options = arpack_defaults()
)

Arguments

graph

Graph to be analyzed.

directed

Logical scalar, whether to consider direction of the edges in directed graphs. It is ignored for undirected graphs.

scale

Logical scalar, whether to scale the result to have a maximum score of one. If no scaling is used then the result vector has unit length in the Euclidean norm.

weights

A numerical vector or NULL. This argument can be used to give edge weights for calculating the weighted eigenvector centrality of vertices. If this is NULL and the graph has a weight edge attribute then that is used. If weights is a numerical vector then it is used, even if the graph has a weight edge attribute. If this is NA, then no edge weights are used (even if the graph has a weight edge attribute). Note that if there are negative edge weights and the direction of the edges is considered, then the eigenvector might be complex. In this case only the real part is reported. This function interprets weights as connection strength. Higher weights spread the centrality better.

options

A named list, to override some ARPACK options. See arpack() for details.