gclust.app {igraph}	R Documentation

Graph Clustering Using NMF (and no SVT) – Apparent Clusters

Description

Performs Clustering of Graphs using Non-negative Factorization.

Usage

    gclust.app(glist, r=1,nmfout=FALSE,maxit=10000, nmfmethod='lee')

Arguments

glist
List of igraph Objects

r
Maximum Number of Clusters Allowed

nmfout
T/F indicating if the output from nmf should be returned; Default is FALSE

maxit
A number passed to irlba or svd function limiting the number of iteration; Default is 10000

nmfmethod
A number passed to irlba or svd, limiting the number of iterations; Default is 'lee'

Value

nmf
An NMF object

W
Basis Graphs

H
Probability Vector for Cluster Weights timeseries of graphs

Xorigin
Input Data in the matrix form

Author(s)

Nam Lee nhlee@jhu.edu

[Package igraph version 1.0.1 Index]