getAICc {igraph}	R Documentation

Compute AIC based on a Poisson Approximation using the output from gclust

Description

Compute and Extract information Criteria Value from gclust using a Poisson approximation, where the penality term is adjusted for small sample cases.

Usage

 getAICc(gfit) 

Arguments

gfit
The output from a call to gclust.rsvt or gclust.app

Value

nclust
Number of clusters being considered

negloglikpart
Negative log likelihood part

parampart
Parameter penalty term

AIC
AIC Value Using a Possion approximation

Author(s)

Nam Lee nhlee@jhu.edu

Examples

require(NMF)
require(irlba)
require(igraph)

[Package igraph version 1.0.1 Index]