graph.bipartite {igraph}	R Documentation

Create a bipartite graph

Description

A bipartite graph has two kinds of vertices and connections are only allowed between different kinds.

Usage

graph.bipartite(types, edges, directed=FALSE)

Arguments

types 	 A vector giving the vertex types. It will be coerced into boolean. The length of the vector gives the number of vertices in the graph.

edges 	 A vector giving the edges of the graph, the same way as for the regular graph function. It is checked that the edges indeed connect vertices of different kind, accoding to the supplied types vector.

directed 	 Whether to create a directed graph, boolean constant. Note that by default undirected graphs are created, as this is more common for bipartite graphs.

Details

Bipartite graphs have a ‘type’ vertex attribute in igraph, this is boolean and FALSE for the vertices of the first kind and TRUE for vertices of the second kind.

graph.bipartite basically does three things. First it checks tha edges vector against the vertex types. Then it creates a graph using the edges vector and finally it adds the types vector as a vertex attribute called type.

Value

A bipartite igraph graph. In other words, an igraph graph that has a vertex attribute type.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

graph to create one-mode networks

Examples

g <- graph.bipartite( rep(0:1,length=10), c(0:9))
print(g, v=TRUE)

[Package igraph version 0.5.2-2 Index]