graph.full.bipartite {igraph}	R Documentation

Create a full bipartite graph

Description

Bipartite graphs are also called two-mode by some. This function creates a bipartite graph in which every possible edge is present.

Usage

graph.full.bipartite (n1, n2, directed = FALSE, mode = c("all", "out", "in")) 

Arguments

n1 	 The number of vertices of the first kind.

n2 	 The number of vertices of the second kind.

directed 	 Logical scalar, whether the graphs is directed.

mode 	 Scalar giving the kind of edges to create for directed graphs. If this is ‘out’ then all vertices of the first kind are connected to the others; ‘in’ specifies the opposite direction; ‘all’ creates mutual edges. This argument is ignored for undirected graphs.x

Details

Bipartite graphs have a ‘type’ vertex attribute in igraph, this is boolean and FALSE for the vertices of the first kind and TRUE for vertices of the second kind.

Value

An igraph graph, with the ‘type’ vertex attribute set.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

graph.full for creating one-mode full graphs

Examples

g <- graph.full.bipartite(2, 3)
g2 <- graph.full.bipartite(2, 3, dir=TRUE)
g3 <- graph.full.bipartite(2, 3, dir=TRUE, mode="in")
g4 <- graph.full.bipartite(2, 3, dir=TRUE, mode="all")

[Package igraph version 0.5.2-2 Index]