graph.laplacian {igraph}	R Documentation

Graph Laplacian

Description

The Laplacian of a graph.

Usage

graph.laplacian(graph, normalized=FALSE)

Arguments

graph 	 The input graph.

normalized 	 Whether to calculate the normalized Laplacian. See definitions below.

Details

The Laplacian Matrix of a graph is a symmetric matrix having the same number of rows and columns as the number of vertices in the graph and element (i,j) is d[i], the degree of vertex i if if i==j, -1 if i!=j and there is an edge between vertices i and j and 0 otherwise.

A normalized version of the Laplacian Matrix is similar: element (i,j) is 1 if i==j, -1/sqrt(d[i] d[j]) if i!=j and there is an edge between vertices i and j and 0 otherwise.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

Examples

g <- graph.ring(10)
graph.laplacian(g)
graph.laplacian(g, norm=TRUE)

[Package igraph version 0.5.2-2 Index]