graph.strength {igraph}	R Documentation

Strength or weighted vertex degree

Description

Summing up the edge weights of the adjacent edges for each vertex.

Usage

graph.strength (graph, vids = V(graph), mode = c("all", "out", "in", "total"),
      loops = TRUE, weights = NULL)

Arguments

graph 	 The input graph.

vids 	 The vertices for which the strength will be calculated.

mode 	 Character string, “out” for out-degree, “in” for in-degree or “all” for the sum of the two. For undirected graphs this argument is ignored.

loops 	 Logical; whether the loop edges are also counted.

weights 	 Weight vector. If the graph has a weight edge attribute, then this is used by default. If the graph does not have a weight edge attribute and this argument is NULL, then a warning is given and degree is called.

Value

A numeric vector giving the strength of the vertices.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

References

Alain Barrat, Marc Barthelemy, Romualdo Pastor-Satorras, Alessandro Vespignani: The architecture of complex weighted networks, Proc. Natl. Acad. Sci. USA 101, 3747 (2004)

See Also

degree for the unweighted version.

Examples

g <- graph.star(10)
E(g)$weight <- seq(ecount(g))
graph.strength(g)
graph.strength(g, mode="out")
graph.strength(g, mode="in")

# No weights, a warning is given
g <- graph.ring(10)
graph.strength(g)

[Package igraph version 0.5.2-2 Index]