graph.structure {igraph}	R Documentation

Method for structural manipulation of graphs

Description

These are the methods for simple manipulation of graphs: adding and deleting edges and vertices.

Usage

add.edges(graph, edges, ..., attr=list())
add.vertices(graph, nv, ..., attr=list())
delete.edges(graph, edges)
delete.vertices(graph, v)

Arguments

graph 	 The graph to work on.

edges 	 Edge sequence, the edges to remove.

... 	 Additional parameters will be added as edge/vertex attributes. Note that these arguments have to be named.

attr 	 Additional edge/vertex attributes to add. This will be concatenated to the other supplied attributes.

nv 	 Numeric constant, the number of vertices to add.

v 	 Vector sequence, the vertices to remove.

Details

add.edges adds the specified edges to the graph. The ids of the vertices are preserved. The additionally supplied named arguments will be added as edge attributes for the new edges. If an attribute was not present in the original graph, its value for the original edges will be NA.

add.vertices adds the specified number of isolate vertices to the graph. The ids of the old vertices are preserved. The additionally supplied named arguments will be added as vertex attributes for the new vertices. If an attribute was not present in the original graph, its value is set to NA for the original vertices.

delete.edges removes the specified edges from the graph. If a specified edge is not present, the function gives an error message, and the original graph remains unchanged. The ids of the vertices are preserved.

delete.vertices removes the specified vertices from the graph together with their adjacent edges. The ids of the vertices are not preserved.

Value

The new graph.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

Examples

g <- graph.ring(10)
add.edges(g, c(1,5,2,6) )
delete.edges(g, E(g, P=c(0,9, 1,2)) )
delete.vertices(g, c(1,6,7) )

[Package igraph version 0.5.2-2 Index]