graph_attr {igraph}	R Documentation

Graph attributes of a graph

Description

Graph attributes of a graph

Usage

graph_attr(graph, name)

Arguments

graph
Input graph.

name
The name of attribute to query. If missing, then all attributes are returned in a list.

Value

A list of graph attributes, or a single graph attribute.

See Also

Other graph attributes: delete_edge_attr, delete_graph_attr, delete_vertex_attr, edge_attr<-, edge_attr_names, edge_attr, graph_attr<-, graph_attr_names, igraph-dollar, igraph-vs-attributes, set_edge_attr, set_graph_attr, set_vertex_attr, vertex_attr<-, vertex_attr_names, vertex_attr

Examples

g <- make_ring(10)
graph_attr(g)
graph_attr(g, "name")

[Package igraph version 1.2.2 Index]