graph_attr<- {igraph}	R Documentation

Set all or some graph attributes

Description

Set all or some graph attributes

Usage

graph_attr(graph, name) <- value

Arguments

graph
The graph.

name
The name of the attribute to set. If missing, then value should be a named list, and all list members are set as attributes.

value
The value of the attribute to set

Value

The graph, with the attribute(s) added.

See Also

Other graph attributes: delete_edge_attr, delete_graph_attr, delete_vertex_attr, edge_attr<-, edge_attr_names, edge_attr, graph_attr_names, graph_attr, igraph-dollar, igraph-vs-attributes, set_edge_attr, set_graph_attr, set_vertex_attr, vertex_attr<-, vertex_attr_names, vertex_attr

Examples

g <- make_graph(~ A - B:C:D)
graph_attr(g, "name") <- "4-star"
g

graph_attr(g) <- list(layout = layout_with_fr(g),
                       name = "4-star layed out")
plot(g)

[Package igraph version 1.2.2 Index]