Create a graph from the Graph Atlas

graph_from_atlas() creates graphs from the book ‘An Atlas of Graphs’ by Roland C. Read and Robin J. Wilson. The atlas contains all undirected graphs with up to seven vertices, numbered from 0 up to 1252. The graphs are listed:

in increasing order of number of nodes;

for a fixed number of nodes, in increasing order of the number of edges;

for fixed numbers of nodes and edges, in increasing order of the degree sequence, for example 111223 < 112222;

for fixed degree sequence, in increasing number of automorphisms.

Usage

graph_from_atlas(n)

atlas(...)

Arguments

n

The id of the graph to create.

...

Passed to graph_from_atlas().

Value

An igraph graph.

See also

Examples

## Some randomly picked graphs from the atlas
graph_from_atlas(sample(0:1252, 1))
#> IGRAPH f5f0839 U--- 7 6 -- Graph from the Atlas #282
#> + attr: name (g/c), n (g/n)
#> + edges from f5f0839:
#> [1] 1--5 1--4 2--3 3--6 6--7 3--7
graph_from_atlas(sample(0:1252, 1))
#> IGRAPH 89dea73 U--- 7 13 -- Graph from the Atlas #1073
#> + attr: name (g/c), n (g/n)
#> + edges from 89dea73:
#>  [1] 1--2 1--7 2--5 2--6 3--4 3--5 3--6 3--7 4--5 4--6 4--7 5--7 6--7