graph_from_isomorphism_class {igraph}	R Documentation

Create a graph from an isomorphism class

Description

The isomorphism class is a non-negative integer number. Graphs (with the same number of vertices) having the same isomorphism class are isomorphic and isomorphic graphs always have the same isomorphism class. Currently it can handle only graphs with 3 or 4 vertices.

Usage

graph_from_isomorphism_class(size, number, directed = TRUE)

Arguments

size
The number of vertices in the graph.

number
The isomorphism class.

directed
Whether to create a directed graph (the default).

Value

An igraph object, the graph of the given size, directedness and isomorphism class.

See Also

Other graph isomorphism: count_isomorphisms, count_subgraph_isomorphisms, isomorphic, isomorphism_class, isomorphisms, subgraph_isomorphic, subgraph_isomorphisms

[Package igraph version 1.2.2 Index]