graph_version {igraph}	R Documentation

Igraph data structure versions

Description

Igraph's internal data representation changes sometimes between versions. This means that it is not possible to use igraph objects that were created (and possibly saved to a file) with an older igraph version.

Usage

graph_version(graph)

Arguments

graph
The input graph. If it is missing, then the version number of the current data format is returned.

Details

graph_version queries the current data format, or the data format of a possibly older igraph graph.

upgrade_graph can convert an older data format to the current one.

Value

A character scalar.

See Also

upgrade_graph to convert the data format of a graph.

[Package igraph version 1.2.2 Index]