grg.game {igraph}	R Documentation

Geometric random graphs

Description

Generate a random graph based on the distance of random point on a unit square

Usage

grg.game(nodes, radius, torus=FALSE, coords=FALSE)

Arguments

nodes 	 The number of vertices in the graph.

radius 	 The radius within which the vertices will be connected by an edge.

torus 	 Logical constant, whether to use a torus instead of a square.

coords 	 Logical scalar, whether to add the positions of the vertices as vertex attributes called ‘x’ and ‘y’.

Details

First a number of points are dropped on a unit square, these points correspond to the vertices of the graph to create. Two points will be connected with an undirected edge if they are closer to each other in Euclidean norm than a given radius. If the torus argument is TRUE then a unit area torus is used instead of a square.

Value

A graph object. If coords is TRUE then with vertex attributes ‘x’ and ‘y’.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu, first version was written by Keith Briggs keith.briggs@bt.com

See Also

random.graph.game

Examples

g <- grg.game(1000, 0.05, torus=FALSE)
g2 <- grg.game(1000, 0.05, torus=TRUE)

[Package igraph version 0.5.2-2 Index]