groups {igraph}	R Documentation

Groups of a vertex partitioning

Description

Create a list of vertex groups from some graph clustering or community structure.

Usage

groups(x)

Arguments

x
Some object that represents a grouping of the vertices. See details below.

Details

Currently two methods are defined for this function. The default method works on the output of components. (In fact it works on any object that is a list with an entry called membership.)

The second method works on communities objects.

Value

A named list of numeric or character vectors. The names are just numbers that refer to the groups. The vectors themselves are numeric or symbolic vertex ids.

See Also

components and the various community finding functions.

Examples

g <- make_graph("Zachary")
fgc <- cluster_fast_greedy(g)
groups(fgc)

g2 <- make_ring(10) + make_full_graph(5)
groups(components(g2))

[Package igraph version 1.2.2 Index]