growing.random.game {igraph}	R Documentation

Growing random graph generation

Description

This function creates a random graph by simulating its stochastic evolution.

Usage

growing.random.game(n, m=1, directed=TRUE, citation=FALSE)

Arguments

n 	 Numeric constant, number of vertices in the graph.

m 	 Numeric constant, number of edges added in each time step.

directed 	 Logical, whether to create a directed graph.

citation 	 Logical. If TRUE a citation graph is created, ie. in each time step the added edges are originating from the new vertex.

Details

This is discrete time step model, in each time step a new vertex is added to the graph and m new edges are created. If citation is FALSE these edges are connecting two uniformly randomly chosen vertices, otherwise the edges are connecting new vertex to uniformly randomly chosen old vertices.

Value

A new graph object.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

barabasi.game, erdos.renyi.game

Examples

g <- growing.random.game(500, citation=FALSE)
g2 <- growing.random.game(500, citation=TRUE)

[Package igraph version 0.5.2-2 Index]