hrg {igraph}	R Documentation

Create a hierarchical random graph from an igraph graph

Description

hrg creates a HRG from an igraph graph. The igraph graph must be a directed binary tree, with n-1 internal and n leaf vertices. The prob argument contains the HRG probability labels for each vertex; these are ignored for leaf vertices.

Usage

hrg(graph, prob)

Arguments

graph
The igraph graph to create the HRG from.

prob
A vector of probabilities, one for each vertex, in the order of vertex ids.

Value

hrg returns an igraphHRG object.

See Also

Other hierarchical random graph functions: consensus_tree, fit_hrg, hrg-methods, hrg_tree, predict_edges, print.igraphHRGConsensus, print.igraphHRG, sample_hrg

[Package igraph version 1.2.2 Index]