hrg-methods {igraph}	R Documentation

Hierarchical random graphs

Description

Fitting and sampling hierarchical random graph models.

Details

A hierarchical random graph is an ensemble of undirected graphs with n vertices. It is defined via a binary tree with n leaf and n-1 internal vertices, where the internal vertices are labeled with probabilities. The probability that two vertices are connected in the random graph is given by the probability label at their closest common ancestor.

Please see references below for more about hierarchical random graphs.

igraph contains functions for fitting HRG models to a given network (fit_hrg, for generating networks from a given HRG ensemble (sample_hrg), converting an igraph graph to a HRG and back (hrg, hrg_tree), for calculating a consensus tree from a set of sampled HRGs (consensus_tree) and for predicting missing edges in a network based on its HRG models (predict_edges).

The igraph HRG implementation is heavily based on the code published by Aaron Clauset, at his website (not functional any more).

See Also

Other hierarchical random graph functions: consensus_tree, fit_hrg, hrg_tree, hrg, predict_edges, print.igraphHRGConsensus, print.igraphHRG, sample_hrg

[Package igraph version 1.2.2 Index]