hrg_tree {igraph}	R Documentation

Create an igraph graph from a hierarchical random graph model

Description

hrg_tree creates the corresponsing igraph tree of a hierarchical random graph model.

Usage

hrg_tree(hrg)

Arguments

hrg
A hierarchical random graph model.

Value

An igraph graph.

See Also

Other hierarchical random graph functions: consensus_tree, fit_hrg, hrg-methods, hrg, predict_edges, print.igraphHRGConsensus, print.igraphHRG, sample_hrg

[Package igraph version 1.2.2 Index]