identical_graphs {igraph}	R Documentation

Decide if two graphs are identical

Description

This is similar to identical in the base package, but ignores the mutable piece of igraph objects, that might be different, even if the two graphs are identical.

Usage

identical_graphs(g1, g2)

Arguments

g1, g2
The two graphs

Value

Logical scalar

[Package igraph version 1.2.2 Index]