igraph-dollar {igraph}	R Documentation

Getting and setting graph attributes, shortcut

Description

The $ operator is a shortcut to get and and set graph attributes. It is shorter and just as readable as graph_attr and set_graph_attr.

Usage

## S3 method for class 'igraph'
x$name

## S3 replacement method for class 'igraph'
x$name <- value

Arguments

x
An igraph graph

name
Name of the attribute to get/set.

value
New value of the graph attribute.

See Also

Other graph attributes: delete_edge_attr, delete_graph_attr, delete_vertex_attr, edge_attr<-, edge_attr_names, edge_attr, graph_attr<-, graph_attr_names, graph_attr, igraph-vs-attributes, set_edge_attr, set_graph_attr, set_vertex_attr, vertex_attr<-, vertex_attr_names, vertex_attr

Examples

g <- make_ring(10)
g$name
g$name <- "10-ring"
g$name

[Package igraph version 1.2.2 Index]