igraph-es-indexing2 {igraph}	R Documentation

Select edges and show their metadata

Description

The double bracket operator can be used on edge sequences, to print the meta-data (edge attributes) of the edges in the sequence.

Usage

## S3 method for class 'igraph.es'
x[[...]]

Arguments

x
An edge sequence.

...
Additional arguments, passed to [.

Details

Technically, when used with edge sequences, the double bracket operator does exactly the same as the single bracket operator, but the resulting edge sequence is printed differently: all attributes of the edges in the sequence are printed as well.

See [.igraph.es for more about indexing edge sequences.

Value

Another edge sequence, with metadata printing turned on. See details below.

See Also

Other vertex and edge sequences: E, V, igraph-es-attributes, igraph-es-indexing, igraph-vs-attributes, igraph-vs-indexing2, igraph-vs-indexing, print.igraph.es, print.igraph.vs

Other vertex and edge sequence operations: c.igraph.es, c.igraph.vs, difference.igraph.es, difference.igraph.vs, igraph-es-indexing, igraph-vs-indexing2, igraph-vs-indexing, intersection.igraph.es, intersection.igraph.vs, rev.igraph.es, rev.igraph.vs, union.igraph.es, union.igraph.vs, unique.igraph.es, unique.igraph.vs

Examples

g <- make_(ring(10),
  with_vertex_(name = LETTERS[1:10]),
  with_edge_(weight = 1:10, color = "green"))
E(g)
E(g)[[]]
E(g)[[.inc('A')]]

[Package igraph version 1.2.2 Index]