igraph_test {igraph}	R Documentation

Run package tests

Description

Runs all package tests.

Usage

igraph_test()

Details

The testthat package is needed to run all tests. The location tests themselves can be extracted from the package via system.file("tests", package="igraph").

This function simply calls the test_dir function from the testthat package on the test directory.

Value

Whatever is returned by test_dir from the testthat package.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

[Package igraph version 1.2.2 Index]