--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.1 --- {% raw %}

igraph_version {igraph}	R Documentation

Query igraph's version string

Description

Queries igraph's original version string. See details below.

Usage

igraph_version()

Details

The igraph version string is the same as the version of the R package for all released igraph versions. For development versions and nightly builds, they might differ however.

The reason for this is, that R package version numbers are not flexible enough to cover in-between releases versions, e.g. alpha and beta versions, release candidates, etc.

Value

A character scalar, the igraph version string.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

Examples

## Compare to the package version
packageDescription("igraph")$Version
igraph_version()

[Package igraph version 1.3.1 Index]
{% endraw %}