incident {igraph}	R Documentation

Incident edges of a vertex in a graph

Description

Incident edges of a vertex in a graph

Usage

incident(graph, v, mode = c("all", "out", "in", "total"))

Arguments

graph
The input graph.

v
The vertex of which the indicent edges are queried.

mode
Whether to query outgoing (‘out’), incoming (‘in’) edges, or both types (‘all’). This is ignored for undirected graphs.

Value

An edge sequence containing the incident edges of the input vertex.

See Also

Other structural queries: [.igraph, [[.igraph, adjacent_vertices, are_adjacent, ends, get.edge.ids, gorder, gsize, head_of, incident_edges, is_directed, neighbors, tail_of

Examples

g <- make_graph("Zachary")
incident(g, 1)
incident(g, 34)

[Package igraph version 1.2.2 Index]