intersection.igraph.es {igraph}	R Documentation

Intersection of edge sequences

Description

Intersection of edge sequences

Usage

## S3 method for class 'igraph.es'
intersection(...)

Arguments

...
The edge sequences to take the intersection of.

Details

They must belong to the same graph. Note that this function has ‘set’ semantics and the multiplicity of edges is lost in the result.

Value

An edge sequence that contains edges that appear in all given sequences, each edge exactly once.

See Also

Other vertex and edge sequence operations: c.igraph.es, c.igraph.vs, difference.igraph.es, difference.igraph.vs, igraph-es-indexing2, igraph-es-indexing, igraph-vs-indexing2, igraph-vs-indexing, intersection.igraph.vs, rev.igraph.es, rev.igraph.vs, union.igraph.es, union.igraph.vs, unique.igraph.es, unique.igraph.vs

Examples

g <- make_(ring(10), with_vertex_(name = LETTERS[1:10]))
intersection(E(g)[1:6], E(g)[5:9])

[Package igraph version 1.2.2 Index]