--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.3 --- {% raw %}

intersection.igraph.vs {igraph}	R Documentation

Intersection of vertex sequences

Description

Intersection of vertex sequences

Usage

## S3 method for class 'igraph.vs'
intersection(...)

Arguments

...
The vertex sequences to take the intersection of.

Details

They must belong to the same graph. Note that this function has ‘set’ semantics and the multiplicity of vertices is lost in the result.

Value

A vertex sequence that contains vertices that appear in all given sequences, each vertex exactly once.

See Also

Other vertex and edge sequence operations: c.igraph.es(), c.igraph.vs(), difference.igraph.es(), difference.igraph.vs(), igraph-es-indexing2, igraph-es-indexing, igraph-vs-indexing2, igraph-vs-indexing, intersection.igraph.es(), rev.igraph.es(), rev.igraph.vs(), union.igraph.es(), union.igraph.vs(), unique.igraph.es(), unique.igraph.vs()

Examples

g <- make_(ring(10), with_vertex_(name = LETTERS[1:10]))
intersection(E(g)[1:6], E(g)[5:9])

[Package igraph version 1.3.3 Index]
{% endraw %}