is.mutual {igraph}	R Documentation

Find mutual edges in a directed graph

Description

This function checks the reciproc pair of the supplied edges.

Usage

is.mutual(graph, es = E(graph))

Arguments

graph 	 The input graph.

es 	 Edge sequence, the edges that will be probed. By default is includes all edges in the order of their ids.

Details

In a directed graph an (A,B) edge is mutual if the graph also includes a (B,A) directed edge.

Note that multi-graphs are not handled properly, i.e. if the graph contains two copies of (A,B) and one copy of (B,A), then these three edges are considered to be mutual.

Undirected graphs contain only mutual edges by definition.

Value

A logical vector of the same length as the number of edges supplied.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

reciprocity, dyad.census if you just want some statistics about mutual edges.

Examples

g <- erdos.renyi.game(10,50,type="gnm",directed=TRUE)
reciprocity(g)
dyad.census(g)
is.mutual(g)
sum(is.mutual(g))/2 == dyad.census(g)$mut

[Package igraph version 0.5.2-2 Index]