is_named {igraph}	R Documentation

Named graphs

Description

An igraph graph is named, if there is a symbolic name associated with its vertices.

Usage

is_named(graph)

Arguments

graph
The input graph.

Details

In igraph vertices can always be identified and specified via their numeric vertex ids. This is, however, not always convenient, and in many cases there exist symbolic ids that correspond to the vertices. To allow this more flexible identification of vertices, one can assign a vertex attribute called ‘name’ to an igraph graph. After doing this, the symbolic vertex names can be used in all igraph functions, instead of the numeric ids.

Note that the uniqueness of vertex names are currently not enforced in igraph, you have to check that for yourself, when assigning the vertex names.

Value

A logical scalar.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

Examples

g <- make_ring(10)
is_named(g)
V(g)$name <- letters[1:10]
is_named(g)
neighbors(g, "a")

[Package igraph version 1.2.2 Index]