is_weighted {igraph}	R Documentation

Weighted graphs

Description

In weighted graphs, a real number is assigned to each (directed or undirected) edge.

Usage

is_weighted(graph)

Arguments

graph
The input graph.

Details

In igraph edge weights are represented via an edge attribute, called ‘weight’. The is_weighted function only checks that such an attribute exists. (It does not even checks that it is a numeric edge attribute.)

Edge weights are used for different purposes by the different functions. E.g. shortest path functions use it as the cost of the path; community finding methods use it as the strength of the relationship between two vertices, etc. Check the manual pages of the functions working with weighted graphs for details.

Value

A logical scalar.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

Examples

g <- make_ring(10)
shortest_paths(g, 8, 2)
E(g)$weight <- seq_len(ecount(g))
shortest_paths(g, 8, 2)

[Package igraph version 1.2.2 Index]