kleinberg {igraph}	R Documentation

Kleinberg's centrality scores.

Description

Kleinberg's hub and authority scores.

Usage

authority.score (graph, scale = TRUE, options = igraph.arpack.default) 
hub.score (graph, scale = TRUE, options = igraph.arpack.default) 

Arguments

graph 	 The input graph.

scale 	 Logical scalar, whether to scale the result to have a maximum score of one. If no scaling is used then the result vector has unit length in the Euclidean norm.

options 	 A named list, to override some ARPACK options. See arpack for details.

Details

The authority scores of the vertices are defined as the principal eigenvector of t(A)*A, where A is the adjacency matrix of the graph.

The hub scores of the vertices are defined as the principal eigenvector of A*t(A), where A is the adjacency matrix of the graph.

Obviously, for undirected matrices the adjacency matrix is symmetric and the two scores are the same.

Value

A named list with members:

vector 	 The authority/hub scores of the vertices.

value 	 The corresponding eigenvalue of the calculated principal eigenvector.

options 	 Some information about the ARPACK computation, it has the same members as the options member returned by arpack, see that for documentation.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

References

J. Kleinberg. Authoritative sources in a hyperlinked environment. Proc. 9th ACM-SIAM Symposium on Discrete Algorithms, 1998. Extended version in Journal of the ACM 46(1999). Also appears as IBM Research Report RJ 10076, May 1997.

See Also

evcent for eigenvector centrality, page.rank for the Page Rank scores. arpack for the underlining machinery of the computation.

Examples

## An in-star
g <- graph.star(10)
hub.score(g)$vector
authority.score(g)$vector

## A ring
g2 <- graph.ring(10)
hub.score(g2)$vector
authority.score(g2)$vector

[Package igraph version 0.5.2-2 Index]