Average nearest neighbor degree

Calculate the average nearest neighbor degree of the given vertices and the same quantity in the function of vertex degree

Usage

knn(
  graph,
  vids = V(graph),
  mode = c("all", "out", "in", "total"),
  neighbor.degree.mode = c("all", "out", "in", "total"),
  weights = NULL
)

Arguments

graph

The input graph. It may be directed.

vids

The vertices for which the calculation is performed. Normally it includes all vertices. Note, that if not all vertices are given here, then both ‘knn’ and ‘knnk’ will be calculated based on the given vertices only.

mode

Character constant to indicate the type of neighbors to consider in directed graphs. out considers out-neighbors, in considers in-neighbors and all ignores edge directions.

neighbor.degree.mode

The type of degree to average in directed graphs. out averages out-degrees, in averages in-degrees and all ignores edge directions for the degree calculation.

weights

Weight vector. If the graph has a weight edge attribute, then this is used by default. If this argument is given, then vertex strength (see strength()) is used instead of vertex degree. But note that knnk is still given in the function of the normal vertex degree. Weights are are used to calculate a weighted degree (also called strength()) instead of the degree.

Value

A list with two members:

knn

A numeric vector giving the average nearest neighbor degree for all vertices in vids.

knnk

A numeric vector, its length is the maximum (total) vertex degree in the graph. The first element is the average nearest neighbor degree of vertices with degree one, etc.

Details

Note that for zero degree vertices the answer in ‘knn’ is NaN (zero divided by zero), the same is true for ‘knnk’ if a given degree never appears in the network.

The weighted version computes a weighted average of the neighbor degrees as

$$k_{nn,u} = \frac{1}{s_u} \sum_v w_{uv} k_v,$$

where \(s_u = \sum_v w_{uv}\) is the sum of the incident edge weights of vertex u, i.e. its strength. The sum runs over the neighbors v of vertex u as indicated by mode. \(w_{uv}\) denotes the weighted adjacency matrix and \(k_v\) is the neighbors' degree, specified by neighbor_degree_mode.

References

Alain Barrat, Marc Barthelemy, Romualdo Pastor-Satorras, Alessandro Vespignani: The architecture of complex weighted networks, Proc. Natl. Acad. Sci. USA 101, 3747 (2004)

See also

Author

Gabor Csardi csardi.gabor@gmail.com

Related documentation in the C library

avg_nearest_neighbor_degree().

Examples

# Some trivial ones
g <- make_ring(10)
knn(g)
#> $knn
#>  [1] 2 2 2 2 2 2 2 2 2 2
#> 
#> $knnk
#> [1] NaN   2
#> 
g2 <- make_star(10)
knn(g2)
#> $knn
#>  [1] 1 9 9 9 9 9 9 9 9 9
#> 
#> $knnk
#> [1]   9 NaN NaN NaN NaN NaN NaN NaN   1
#> 

# A scale-free one, try to plot 'knnk'
g3 <- sample_pa(1000, m = 5)
knn(g3)
#> $knn
#>    [1]  14.61677  14.82555  13.26648  15.27240  16.80723  18.09722  28.00000
#>    [8]  31.88333  15.85950  50.07895  48.20588  15.64362  26.44928  40.33333
#>   [15]  20.93000 106.91667  14.37415  24.18824 132.85714  26.91071  56.26087
#>   [22] 111.00000  33.42424  57.85714  67.75000  29.67500  54.92593  27.40385
#>   [29] 160.00000 148.20000 274.20000  64.47368  27.89286  25.72840  42.13043
#>   [36]  24.46429  36.16000  17.92063  43.52632  39.86207  44.48387 128.62500
#>   [43]  83.81818  66.00000  76.33333  65.31250  47.05882  90.60000  22.08571
#>   [50]  74.00000  87.87500  53.62500  63.30000 128.00000  72.27778  42.70370
#>   [57] 126.20000  43.50000 157.00000 108.69231  40.00000 202.66667  90.53333
#>   [64] 108.57143  98.50000 106.72727  90.70000  17.19643  87.77778 104.25000
#>   [71] 118.77778 128.85714 116.37500 102.55556  93.50000  65.11111  54.25000
#>   [78] 173.00000  91.63636  55.31250 124.83333 183.00000  18.43333  54.66667
#>   [85]  90.70000  53.00000  97.21429 104.10000 142.00000  44.75000  61.08333
#>   [92]  86.80000 118.80000  59.88235  78.00000  64.57143  79.93750  42.32143
#>   [99]  53.80000  60.84615  93.00000  62.15385 212.20000 146.50000 142.37500
#>  [106]  70.94737 121.45455 144.50000  98.85714 175.50000 184.00000 125.71429
#>  [113] 237.00000 142.12500 107.14286 126.33333  96.58333 155.50000  69.25000
#>  [120] 163.66667 226.60000  91.75000 111.50000 119.87500 132.12500  91.12500
#>  [127]  39.62500  58.87500  90.42857 153.00000 172.80000 225.66667 111.00000
#>  [134] 177.83333 158.87500  69.86667 179.14286 138.20000  70.41667 118.00000
#>  [141]  75.92857 111.12500 156.66667  78.68750 101.00000  43.38889  84.60000
#>  [148] 123.66667 123.16667 150.80000  41.17647 147.71429  62.00000  33.15385
#>  [155]  59.86667  90.40000 107.14286  89.60000  50.60000  53.61111  90.88889
#>  [162]  55.50000 180.00000 205.00000 137.83333  66.66667  98.00000  67.78571
#>  [169] 140.60000 215.80000  56.33333 195.16667 161.85714 179.80000  88.50000
#>  [176]  92.40000 109.88889 130.28571  67.15385 150.80000  90.87500 216.66667
#>  [183] 129.57143 103.16667  67.22222 148.62500  37.66667 137.00000 137.60000
#>  [190] 160.71429 188.40000 208.60000  35.80952 222.40000  94.14286 216.00000
#>  [197] 143.22222  71.50000  17.18182 110.16667 192.60000  37.37500  72.11111
#>  [204] 110.00000 161.85714  34.18182 111.63636 159.16667  91.12500 237.20000
#>  [211]  83.00000 208.20000 161.20000 131.66667  81.25000  98.33333  68.60000
#>  [218] 167.60000 151.71429 132.42857 134.40000  79.00000 125.80000 135.80000
#>  [225] 191.66667 106.00000 184.66667 178.00000  58.40000 286.80000 180.20000
#>  [232]  68.85714  85.57143  31.76923  98.66667 155.60000 147.57143  42.00000
#>  [239] 138.00000 145.20000 108.50000  51.10000 100.76923 151.80000  85.00000
#>  [246] 164.16667  61.66667  27.92308  44.85714  68.10000 122.60000 130.20000
#>  [253]  47.66667  58.40000 119.66667 100.16667 126.22222  53.50000  96.16667
#>  [260]  99.69231 104.80000 117.14286 135.00000 165.83333  94.28571 100.16667
#>  [267]  54.83333  58.42857 104.40000  62.08333  62.36364 288.20000 118.57143
#>  [274] 258.60000  97.33333 119.85714  39.71429  94.00000  89.40000 121.66667
#>  [281]  95.60000 207.00000  38.60000  47.76923 164.60000 184.28571 177.20000
#>  [288]  82.33333 147.40000 159.33333 195.00000 103.00000 100.40000 138.66667
#>  [295]  99.00000 149.60000 109.40000 160.16667 179.60000  47.00000 108.12500
#>  [302] 132.50000  25.75000 164.16667  70.62500  27.41667 153.20000  72.22222
#>  [309] 120.14286  77.00000  85.80000 143.80000 143.20000 106.75000  80.16667
#>  [316] 120.00000 180.00000  64.63636 194.60000 125.50000  68.63636 125.57143
#>  [323]  37.66667  31.85714 114.80000 145.80000  87.72727 134.20000 175.83333
#>  [330] 176.60000 118.00000 148.80000 168.50000 167.40000 166.20000  80.33333
#>  [337] 117.16667 176.80000 116.60000 166.33333  90.70000  67.20000 122.14286
#>  [344] 104.87500 177.42857  86.00000  47.60000  83.00000  80.00000  97.40000
#>  [351]  56.40000  83.62500 152.60000 221.80000  64.00000 127.50000 135.80000
#>  [358]  60.12500 101.37500  91.85714 232.00000  31.10000 193.80000  92.00000
#>  [365] 138.40000  46.12500  40.50000  84.20000 182.80000 101.40000 127.80000
#>  [372] 107.42857 130.60000 257.80000  41.18182 217.20000 180.80000 164.50000
#>  [379] 164.50000 148.57143 142.80000 129.50000  93.71429 193.80000  50.55556
#>  [386] 145.40000  86.00000 118.83333 160.20000 185.60000  65.00000 118.71429
#>  [393]  67.83333 110.00000  60.40000 192.40000  83.81818  50.40000 129.71429
#>  [400]  73.83333 241.60000 138.20000 102.66667 126.28571 148.00000 157.60000
#>  [407]  26.45455 189.20000  56.14286 134.80000  87.33333  84.50000  49.50000
#>  [414] 210.00000  73.50000 197.00000 105.30000  84.40000 128.40000 143.16667
#>  [421] 201.00000  32.83333 165.16667  81.00000 177.33333  59.62500 127.50000
#>  [428] 158.00000 167.50000 178.80000  62.66667  76.60000 152.20000 198.80000
#>  [435]  86.50000 121.66667 105.83333  93.66667  68.62500 115.83333 123.40000
#>  [442]  57.33333 204.00000 165.80000 108.16667 232.40000  38.00000 105.85714
#>  [449]  48.50000 128.83333 115.16667 151.00000 124.40000 158.85714  85.60000
#>  [456] 118.80000 204.40000 178.20000  56.28571  98.12500 108.87500 125.28571
#>  [463] 173.00000  76.00000 152.00000  81.16667 120.66667  87.62500  59.12500
#>  [470] 205.00000  82.85714  94.50000 151.33333  33.10000 106.66667 114.57143
#>  [477] 151.80000  73.62500 106.28571 190.00000 117.00000 147.80000  80.00000
#>  [484] 173.60000 188.83333 121.83333 218.16667 112.60000 113.42857 205.40000
#>  [491] 183.40000  71.57143  93.83333  88.00000 116.00000 154.80000 142.62500
#>  [498] 126.60000 240.40000 147.20000  97.80000 112.40000 168.00000  52.00000
#>  [505]  81.20000 130.83333  84.71429 175.20000 123.60000  98.33333 154.20000
#>  [512] 197.20000  69.00000  65.66667  52.00000 144.00000 163.80000 131.00000
#>  [519]  92.71429  79.20000 163.00000 166.40000 122.60000  93.85714 172.33333
#>  [526]  88.20000  90.14286 118.00000  85.50000  31.00000  54.33333 126.80000
#>  [533] 164.80000  69.88889 143.80000 244.40000  78.16667 113.16667 175.20000
#>  [540]  66.20000  64.28571  87.55556 173.60000 102.85714 137.50000  61.80000
#>  [547] 148.83333 264.20000 201.66667 172.60000 136.60000  98.60000 110.50000
#>  [554]  87.00000 134.28571  59.00000 104.00000  73.10000 116.83333 108.00000
#>  [561] 137.60000 126.40000 179.00000 120.00000 133.40000 107.00000 202.20000
#>  [568] 234.60000 135.20000 189.60000  96.40000  90.00000 125.66667 139.66667
#>  [575] 138.60000  99.00000 146.60000 181.57143 178.00000 218.40000 116.20000
#>  [582] 167.00000  77.60000 163.20000 125.20000 139.00000  61.20000 196.40000
#>  [589]  63.40000  92.60000 147.50000 198.60000 147.20000  90.80000  78.20000
#>  [596] 131.20000  99.00000 154.40000 105.00000 112.60000 113.40000 199.40000
#>  [603] 152.60000 196.66667 278.00000 163.83333 154.40000 101.16667 181.00000
#>  [610] 199.00000  44.40000 232.00000  78.00000  50.28571 161.80000 165.60000
#>  [617] 129.00000 194.00000 133.42857 138.80000  71.60000 195.20000 288.00000
#>  [624] 113.33333  68.80000  55.57143  73.83333 248.60000  93.33333 103.83333
#>  [631] 240.60000 140.66667 188.80000 149.80000  65.00000 157.40000 135.80000
#>  [638] 165.80000 103.20000 139.60000  69.40000 134.00000 285.00000 171.80000
#>  [645]  62.66667  61.80000 250.40000  75.40000 170.80000 144.40000 136.80000
#>  [652]  87.66667 166.40000 141.20000  61.20000  96.12500 111.50000 128.60000
#>  [659] 111.60000 136.00000 122.83333  73.20000 186.20000 149.80000 149.00000
#>  [666] 194.00000 154.60000  91.80000 136.00000  38.00000 124.60000 120.20000
#>  [673]  16.83333 167.00000  47.77778 167.33333 172.80000 101.40000  32.50000
#>  [680]  90.28571  94.40000 142.00000  67.16667 172.00000 199.00000 122.00000
#>  [687] 185.60000  93.60000 171.60000 258.60000  73.00000 141.50000  26.33333
#>  [694]  33.80000 215.00000  85.28571 159.40000  65.28571 137.40000 122.66667
#>  [701] 215.80000  31.00000  18.71429  66.20000  43.16667  65.62500 121.60000
#>  [708]  66.50000 127.00000 179.40000 227.60000  68.00000 133.20000  88.00000
#>  [715]  20.83333  76.80000 105.50000 133.40000  73.33333 188.80000  64.00000
#>  [722] 207.66667 204.00000 181.00000 167.40000 124.66667 169.80000  15.20000
#>  [729] 141.80000  86.00000  64.60000  74.50000 189.80000  49.85714  65.50000
#>  [736]  48.33333 146.20000  86.33333 121.20000  13.40000 166.60000 174.40000
#>  [743]  79.60000 133.00000  38.40000 188.40000  63.20000 120.00000  59.33333
#>  [750]  98.20000  40.40000 132.60000 119.60000  87.00000 184.80000  11.00000
#>  [757]  22.60000 193.40000 104.83333 137.20000 137.60000 181.00000 121.80000
#>  [764]  76.80000  96.00000 183.60000 156.40000 226.80000 100.66667  62.00000
#>  [771]  24.60000  72.00000 185.20000 148.40000  44.40000 131.40000  84.83333
#>  [778]  63.80000  79.40000 111.16667 162.40000 181.00000  80.20000  97.33333
#>  [785] 173.20000  23.33333 184.80000 163.20000 218.40000 161.20000 135.80000
#>  [792] 108.00000  41.20000  81.40000  70.60000 220.00000  79.20000  84.80000
#>  [799] 112.20000  74.00000 128.40000 180.60000  54.80000  49.00000 136.40000
#>  [806]  87.60000  20.60000  54.16667  75.33333 130.60000  89.80000  72.00000
#>  [813]  79.40000 171.60000 247.60000 218.00000  19.00000  91.71429 115.60000
#>  [820] 176.40000  72.50000 175.20000 197.80000 164.80000 180.40000 118.80000
#>  [827] 121.00000 113.00000  72.40000  70.66667 159.20000 123.60000 195.20000
#>  [834]  85.00000 164.20000 147.20000 140.00000 105.20000  79.60000  64.66667
#>  [841] 112.00000 151.60000 115.20000 107.40000  96.00000 150.00000  29.71429
#>  [848]  61.40000 143.20000  55.00000 131.20000 131.60000  76.60000  85.60000
#>  [855]  72.80000 165.60000  69.60000 204.20000  40.50000 118.80000 182.20000
#>  [862] 108.80000 158.00000 191.00000 148.66667  89.00000 160.80000  94.00000
#>  [869] 136.40000  91.20000 140.80000 130.20000 198.16667  41.20000 137.60000
#>  [876]  11.80000 117.00000 100.60000 125.00000  25.40000  97.00000  93.00000
#>  [883]  70.80000  72.00000 258.00000  80.80000  62.60000 117.60000 142.40000
#>  [890]  85.20000 153.00000 114.80000  51.20000  73.40000  81.60000 124.40000
#>  [897] 130.80000 108.00000 108.80000 123.20000 102.40000 133.20000  62.60000
#>  [904]  87.60000 127.20000  97.33333  47.60000 110.20000  76.40000  73.00000
#>  [911]  60.80000 159.20000 111.60000 231.00000 161.80000  47.80000 162.80000
#>  [918] 121.20000  28.60000  50.40000  87.00000 146.20000 129.80000 167.60000
#>  [925] 205.20000 124.40000 231.60000  99.16667 158.00000 130.20000  91.20000
#>  [932] 215.40000  74.40000 171.20000 181.00000  23.20000 143.80000 157.20000
#>  [939] 150.80000 137.20000 202.20000 188.80000  71.33333 212.40000 145.60000
#>  [946] 103.20000 161.60000  65.40000  97.60000 228.20000  66.40000 139.66667
#>  [953]  39.00000 209.00000  22.33333  85.20000 163.40000  98.60000  75.40000
#>  [960] 132.80000  90.00000 136.40000  90.80000 119.60000  72.00000 139.40000
#>  [967]   9.20000 119.00000  58.16667  80.40000 194.00000 141.00000 145.00000
#>  [974] 205.80000 171.60000 177.00000 147.40000 173.20000 143.80000  64.00000
#>  [981]  43.40000  76.80000 165.60000 113.60000  66.60000 147.40000  96.00000
#>  [988] 104.16667 104.20000 155.40000 139.00000  56.00000 114.00000  90.80000
#>  [995]  77.20000  26.60000  75.20000 132.20000  22.40000 195.40000
#> 
#> $knnk
#>   [1]       NaN       NaN       NaN       NaN 135.13004 113.97401 105.72857
#>   [8] 102.56562  83.30370  84.17273  74.67380  75.38542  67.41880  76.37500
#>  [15]  65.74667  60.76042  50.52941  56.42593  67.71053  53.00000  48.90476
#>  [22]       NaN  49.19565  46.62500  36.16000  33.15385  48.81481  42.69048
#>  [29]  39.86207  18.43333  44.48387       NaN  33.42424  47.63235  40.00000
#>  [36]       NaN       NaN  46.80263       NaN  29.67500       NaN       NaN
#>  [43]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#>  [50]       NaN  40.33333  27.40385       NaN       NaN       NaN  22.05357
#>  [57]       NaN       NaN       NaN  31.88333       NaN       NaN  17.92063
#>  [64]       NaN       NaN       NaN       NaN       NaN  26.44928  25.04286
#>  [71]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#>  [78]       NaN       NaN       NaN  25.72840       NaN       NaN  24.46429
#>  [85]  24.18824       NaN       NaN       NaN       NaN       NaN       NaN
#>  [92]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#>  [99]       NaN  20.93000       NaN       NaN       NaN       NaN       NaN
#> [106]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [113]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [120]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [127]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [134]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [141]       NaN       NaN       NaN       NaN       NaN       NaN  14.37415
#> [148]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [155]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [162]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [169]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [176]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [183]       NaN       NaN       NaN       NaN       NaN  15.64362       NaN
#> [190]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [197]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [204]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [211]       NaN       NaN       NaN       NaN       NaN  18.09722       NaN
#> [218]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [225]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [232]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [239]       NaN       NaN       NaN  15.85950       NaN       NaN       NaN
#> [246]       NaN       NaN       NaN  16.80723       NaN       NaN       NaN
#> [253]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [260]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [267]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [274]       NaN       NaN       NaN       NaN       NaN  15.27240       NaN
#> [281]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [288]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [295]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [302]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [309]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [316]       NaN       NaN       NaN       NaN       NaN  14.82555       NaN
#> [323]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [330]       NaN       NaN       NaN       NaN  14.61677       NaN       NaN
#> [337]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [344]       NaN       NaN       NaN       NaN       NaN  13.26648
#> 

# A random graph
g4 <- sample_gnp(1000, p = 5 / 1000)
knn(g4)
#> $knn
#>    [1]  6.666667  5.000000  7.428571  8.250000  5.750000  6.400000  6.666667
#>    [8]  5.500000  5.333333  7.500000  8.333333  7.333333  3.750000  6.000000
#>   [15]  5.000000  5.285714  6.200000  4.250000  6.416667  7.400000  4.571429
#>   [22]  5.166667  6.142857  7.444444  6.000000  5.200000  7.500000  6.666667
#>   [29]  5.857143  7.818182  5.500000  7.000000  7.166667  6.833333  5.857143
#>   [36]  6.000000  6.000000  6.000000  3.500000  7.500000  4.750000  5.000000
#>   [43]  6.428571  7.333333  4.000000  4.000000  5.222222  7.250000  6.000000
#>   [50]  7.750000  6.444444  3.750000  6.166667  5.166667  4.857143  4.000000
#>   [57] 10.000000  4.500000  5.000000  4.000000  6.000000  3.000000  6.250000
#>   [64]  4.333333  8.000000  7.333333  8.666667  3.000000  6.666667  5.000000
#>   [71]  5.000000  4.750000  7.500000  7.166667  6.090909  5.888889  5.200000
#>   [78]  7.000000  5.500000  6.000000  6.666667  5.250000  5.777778  6.600000
#>   [85]  6.666667  6.250000  5.500000  4.250000  6.000000  4.875000  5.500000
#>   [92]  6.428571  3.333333  6.666667  7.125000  7.500000  6.000000  7.166667
#>   [99]  7.500000  7.333333  5.727273  5.272727  7.333333  4.333333  7.166667
#>  [106]  5.846154  7.200000  6.636364  8.000000  7.500000  5.000000  5.500000
#>  [113]  5.000000  5.875000  6.500000  6.833333  6.333333  5.714286  4.666667
#>  [120]  5.000000  7.750000  5.166667  6.750000  7.600000  5.857143  5.000000
#>  [127]  7.250000  6.285714  7.000000  4.600000  6.000000  5.000000  6.166667
#>  [134]  4.666667  6.285714  7.500000  4.800000  5.857143  5.000000  6.333333
#>  [141]  6.600000  7.333333  8.200000  6.250000  5.800000  9.500000  6.500000
#>  [148]  6.000000  6.333333  5.000000  7.500000  6.375000  5.250000  5.666667
#>  [155]  8.000000  6.500000  6.333333  5.428571  5.111111  5.600000  6.666667
#>  [162] 11.000000  4.600000  5.750000  5.250000  6.142857  5.000000  6.800000
#>  [169]  4.200000  5.400000  7.000000  7.000000  5.250000  5.666667  6.000000
#>  [176]  7.000000  6.333333  8.400000  5.125000  6.714286  7.000000  6.777778
#>  [183]  6.750000  5.666667  5.750000  5.777778  5.200000  6.100000  6.000000
#>  [190]  6.666667  7.000000  6.600000  4.000000  7.142857  6.125000  4.500000
#>  [197]  6.000000  6.000000  6.400000  4.333333  6.250000  7.000000  4.666667
#>  [204]  3.400000  7.000000  6.833333  6.000000  7.333333  7.750000  5.250000
#>  [211]  5.400000  4.666667  6.666667  6.000000  7.000000  7.166667  7.000000
#>  [218]  5.875000  5.500000  4.666667  6.250000  5.000000  4.000000  5.750000
#>  [225]  5.800000  4.333333  4.666667  6.200000  6.750000  6.333333  5.500000
#>  [232] 10.000000  5.500000  5.250000  5.666667  5.600000  9.000000  3.666667
#>  [239]  6.500000  6.500000  5.500000  6.200000  7.500000  6.000000  3.666667
#>  [246]  6.250000  4.500000  7.428571  6.000000  7.000000  7.333333  4.000000
#>  [253]  6.250000  7.000000  5.333333  7.000000  7.000000  4.000000  6.857143
#>  [260]  6.250000  6.200000  5.000000  7.100000  7.000000  7.600000  5.333333
#>  [267]  6.250000  5.250000  5.000000  7.000000  6.666667  8.166667  6.333333
#>  [274]  5.400000  7.333333  6.375000  5.444444  6.000000  9.000000  5.000000
#>  [281]  5.250000  4.500000  5.000000  6.200000  5.857143  6.000000  6.300000
#>  [288]  5.000000  7.000000  5.600000  5.500000  7.000000  6.500000  7.142857
#>  [295]  6.250000  8.666667  6.600000  7.571429  6.000000  6.333333  7.111111
#>  [302]  6.600000  7.000000  7.000000  5.333333  6.000000  6.750000  8.500000
#>  [309]  7.600000  7.600000  7.000000  6.000000  6.750000  5.900000  6.800000
#>  [316]  5.000000  5.250000  4.500000  5.666667  6.000000  6.200000  7.666667
#>  [323]  6.000000  5.400000  5.000000  8.600000  5.875000  7.666667  5.666667
#>  [330]  6.000000  8.000000  5.666667  4.000000  6.666667  6.333333  6.285714
#>  [337]  6.272727  5.000000  7.000000  6.166667  6.333333  7.333333  5.600000
#>  [344]  4.750000  5.250000  6.750000  5.000000  5.750000  7.000000  5.500000
#>  [351]  5.500000  6.000000  6.200000  5.800000  5.571429  4.833333  4.600000
#>  [358]  5.166667  6.615385  6.250000  6.333333  7.666667  5.166667  8.333333
#>  [365]  5.000000  6.750000  5.000000  5.250000  6.333333  4.333333  5.666667
#>  [372]  7.777778  7.333333  6.888889  6.166667  5.272727  6.333333  6.714286
#>  [379]  6.000000  3.400000  4.750000  6.142857  7.500000  5.400000  6.818182
#>  [386]  5.500000  7.166667  7.555556  5.833333  7.666667  6.250000  5.000000
#>  [393]  7.333333  6.000000  6.500000  5.400000  3.333333  5.000000  6.000000
#>  [400]  5.000000  6.416667  7.750000  5.200000  5.083333  6.555556  7.800000
#>  [407]  7.333333  8.250000  6.750000  7.333333  4.750000 11.000000  5.000000
#>  [414]  5.750000  6.285714  6.833333  5.600000  5.285714  6.375000  5.833333
#>  [421]  5.000000  5.363636  8.000000  6.142857  7.444444  5.000000  4.000000
#>  [428]  5.200000  8.500000  5.625000  5.800000  6.875000  5.833333  6.333333
#>  [435]  6.666667  4.000000  5.500000  5.250000  5.222222  5.375000  6.000000
#>  [442]  5.000000  5.750000  4.750000  6.800000  7.666667  5.666667  5.888889
#>  [449]  7.200000  5.250000  5.750000  3.857143  7.666667  5.166667  6.833333
#>  [456]  5.500000  5.500000  7.166667  6.600000  6.800000  5.400000  5.000000
#>  [463]  6.666667  7.000000  6.000000  4.875000  4.000000  7.000000  7.500000
#>  [470]  5.200000  6.250000  5.200000  6.142857  7.750000  5.000000  6.857143
#>  [477]  6.500000  6.500000  5.666667  7.400000  4.200000  4.000000  7.666667
#>  [484]  6.000000  6.400000  6.857143  4.333333  5.833333  9.000000  5.500000
#>  [491]  5.666667  6.666667  7.400000  6.750000  5.400000  7.400000  6.600000
#>  [498]  6.250000  4.833333  6.000000 12.000000  5.500000  5.250000  6.400000
#>  [505]  5.166667  4.000000  7.166667  4.714286  6.500000  5.000000  8.000000
#>  [512]  8.400000  9.000000  5.250000  4.500000  4.500000  5.000000  6.833333
#>  [519]  6.285714  5.857143  7.250000  6.666667  5.166667  7.166667  7.333333
#>  [526]  5.750000  5.333333  4.666667  5.000000  6.666667  7.500000  4.333333
#>  [533]  7.000000  7.166667  6.000000  6.333333  6.000000  5.454545  5.400000
#>  [540]  5.571429  5.714286  5.500000  5.333333  5.666667  8.000000  6.750000
#>  [547]  4.200000  5.250000  8.142857  6.111111  5.500000  7.000000  8.250000
#>  [554]  7.000000  6.500000  5.250000  5.833333  5.000000  4.636364       NaN
#>  [561]  4.000000  6.000000  5.200000  8.500000  6.666667  6.000000 11.000000
#>  [568]  6.000000  7.111111  7.000000  7.500000  3.000000  7.285714  7.000000
#>  [575]  6.000000  6.500000  4.000000  4.000000       NaN  5.750000  5.750000
#>  [582]  5.166667  8.400000  5.200000 10.333333  6.666667  5.500000  6.000000
#>  [589]  5.500000  7.500000  6.666667  6.333333  6.000000  4.777778  5.500000
#>  [596]  5.857143  5.285714  5.666667  7.777778       NaN  5.500000  4.800000
#>  [603]  6.750000  5.666667  5.222222  7.000000  6.000000  6.000000  5.888889
#>  [610]  8.500000  4.000000  6.500000  6.000000  4.750000  5.800000  7.500000
#>  [617]  5.000000  5.500000  5.666667  5.200000  5.666667  7.000000  5.800000
#>  [624]  6.800000  3.666667  6.555556  6.500000  5.500000  6.400000  7.166667
#>  [631]  4.777778  6.200000  6.111111  5.750000  6.500000  5.500000  6.200000
#>  [638]  6.428571  3.750000  6.571429  8.285714  4.250000  6.750000  6.800000
#>  [645]  4.500000  5.333333  6.428571  6.500000  6.250000  7.333333  5.142857
#>  [652]  5.000000  5.250000  6.000000  6.000000  5.333333  4.714286  5.000000
#>  [659]  6.000000  4.333333  4.400000  8.250000  5.800000  5.833333  4.800000
#>  [666]  7.666667  6.000000  5.777778  4.800000  6.142857  6.333333  6.875000
#>  [673]  6.400000  5.857143  5.250000  5.000000  4.800000  7.333333  7.000000
#>  [680]  5.600000  4.333333  4.800000  4.500000  6.000000  5.000000  4.000000
#>  [687]  7.166667  5.000000  7.000000  7.666667  3.500000  6.833333  4.000000
#>  [694]  5.833333  7.000000  6.000000  8.500000  6.000000  5.250000  7.500000
#>  [701] 10.000000  6.666667  8.000000  5.500000  7.000000  5.625000  6.500000
#>  [708]  5.857143  4.750000  7.090909  6.750000  6.000000  5.900000  5.833333
#>  [715]  5.750000  7.285714  6.000000  6.625000  7.500000  6.000000  5.000000
#>  [722]  7.600000  5.250000  7.000000  7.250000  7.166667  7.800000  8.200000
#>  [729]  6.400000  8.250000  7.166667  3.666667  6.500000  7.800000  4.000000
#>  [736]  6.333333  7.000000  5.000000  4.333333  6.142857  5.600000  4.000000
#>  [743]  5.600000  6.666667  3.750000  5.750000  5.000000  6.166667  6.200000
#>  [750]  6.000000  7.000000  7.750000  5.571429  6.300000  5.250000  6.000000
#>  [757]  7.666667  7.250000  6.333333  5.666667  8.000000  5.000000  6.200000
#>  [764]  6.000000  6.250000  4.800000  4.285714  6.166667  5.000000  5.666667
#>  [771]  4.800000  6.500000  5.400000  6.750000  5.285714  6.625000  8.500000
#>  [778]  4.833333  6.500000  5.750000  6.666667  5.545455  6.833333  7.250000
#>  [785]  4.666667  5.818182  6.666667  5.000000  6.666667  5.500000  6.666667
#>  [792]  7.333333  5.333333  3.000000  7.000000  5.285714  5.444444  5.250000
#>  [799]  6.571429  6.500000  6.000000  6.600000  8.000000  5.500000  7.400000
#>  [806]  5.000000  5.500000  5.166667  4.666667  6.500000  5.250000  4.500000
#>  [813]  4.000000  6.833333  5.500000  5.666667  5.666667  4.333333  6.333333
#>  [820]  4.000000  6.500000  6.125000  5.666667  5.666667  6.500000  5.000000
#>  [827]  8.500000  5.500000  7.250000  7.000000  6.000000  7.000000  5.000000
#>  [834]  6.000000  6.111111  6.857143  6.636364  5.166667       NaN  5.800000
#>  [841]  6.666667  8.571429  6.000000  6.142857  6.142857  6.625000  8.000000
#>  [848]  7.666667  5.333333  6.000000  6.375000  5.000000  6.500000  5.000000
#>  [855]  7.166667  4.600000  6.100000  5.166667  6.750000  7.250000  8.000000
#>  [862]  5.500000  9.000000  4.000000  5.000000  4.600000  6.400000  6.000000
#>  [869]  6.125000  8.000000  5.857143  5.666667  6.500000  6.000000  6.500000
#>  [876]  4.000000  5.333333  5.600000  7.500000  5.000000  5.833333  6.285714
#>  [883]  5.571429  5.000000  4.666667  8.500000  5.000000  8.333333  6.250000
#>  [890]  4.250000  6.600000  9.000000  5.111111  5.200000  6.250000  6.333333
#>  [897]  5.000000  6.833333  6.142857  5.500000  6.800000  5.857143  6.666667
#>  [904]  3.750000  5.000000  5.500000  6.200000  6.833333  5.000000  6.333333
#>  [911]  8.428571  8.500000  6.250000  5.000000  6.285714  6.750000  7.750000
#>  [918]  4.625000  6.500000  6.400000  4.666667  7.428571  5.833333  5.333333
#>  [925]  6.571429  3.000000  4.000000  6.666667  5.833333  5.333333  7.777778
#>  [932]  4.750000  7.600000  6.875000  4.333333  5.600000  7.600000  3.333333
#>  [939]  5.666667  5.000000  5.000000  5.500000  8.500000  7.000000  7.500000
#>  [946]  7.166667       NaN  2.000000  8.000000  6.200000  6.111111  4.666667
#>  [953]  6.800000  5.111111  3.000000  5.800000  7.000000  6.333333  4.000000
#>  [960]  5.600000  5.285714  6.500000  8.000000  5.750000  7.000000  7.428571
#>  [967]  6.500000  6.000000  7.500000  5.714286  7.857143  5.833333  5.250000
#>  [974]  6.000000  5.800000  6.000000  6.285714  7.750000  7.000000  3.666667
#>  [981]  6.666667  6.000000  6.800000  5.666667  5.000000  6.666667  6.500000
#>  [988]  6.000000  4.000000  6.166667  6.400000  5.000000  6.400000  8.500000
#>  [995]  7.333333  5.000000  6.500000  5.625000  6.666667  5.500000
#> 
#> $knnk
#>  [1] 5.906250 6.351852 5.942605 6.116216 6.049682 6.180208 6.152276 5.907328
#>  [9] 6.219444 5.813333 5.858586 5.972222 6.153846 5.714286
#> 

# A weighted graph
g5 <- make_star(10)
E(g5)$weight <- seq(ecount(g5))
knn(g5)
#> $knn
#>  [1] 1 9 9 9 9 9 9 9 9 9
#> 
#> $knnk
#> [1]   9 NaN NaN NaN NaN NaN NaN NaN   1
#>