layout_as_star {igraph}	R Documentation

Generate coordinates to place the vertices of a graph in a star-shape

Description

A simple layout generator, that places one vertex in the center of a circle and the rest of the vertices equidistantly on the perimeter.

Usage

layout_as_star(graph, center = V(graph)[1], order = NULL)

as_star(...)

Arguments

graph
The graph to layout.

center
The id of the vertex to put in the center. By default it is the first vertex.

order
Numeric vector, the order of the vertices along the perimeter. The default ordering is given by the vertex ids.

...
Arguments to pass to layout_as_star.

Details

It is possible to choose the vertex that will be in the center, and the order of the vertices can be also given.

Value

A matrix with two columns and as many rows as the number of vertices in the input graph.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

layout and layout.drl for other layout algorithms, plot.igraph and tkplot on how to plot graphs and star on how to create ring graphs.

Other graph layouts: add_layout_, component_wise, layout_as_bipartite, layout_as_tree, layout_in_circle, layout_nicely, layout_on_grid, layout_on_sphere, layout_randomly, layout_with_dh, layout_with_fr, layout_with_gem, layout_with_graphopt, layout_with_kk, layout_with_lgl, layout_with_mds, layout_with_sugiyama, layout_, merge_coords, norm_coords, normalize

Examples

g <- make_star(10)
layout_as_star(g)

## Alternative form
layout_(g, as_star())

[Package igraph version 1.2.2 Index]