layout_on_sphere {igraph}	R Documentation

Graph layout with vertices on the surface of a sphere

Description

Place vertices on a sphere, approximately uniformly, in the order of their vertex ids.

Usage

layout_on_sphere(graph)

on_sphere(...)

Arguments

graph
The input graph.

...
Passed to layout_on_sphere.

Details

layout_on_sphere places the vertices (approximately) uniformly on the surface of a sphere, this is thus a 3d layout. It is not clear however what “uniformly on a sphere” means.

If you want to order the vertices differently, then permute them using the permute function.

Value

A numeric matrix with three columns, and one row for each vertex.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

Other graph layouts: add_layout_; as_bipartite, layout.bipartite, layout_as_bipartite; as_star, layout.star, layout_as_star; as_tree, layout_as_tree; component_wise; in_circle, layout_in_circle; layout.auto, layout_nicely, nicely; layout.davidson.harel, layout_with_dh, with_dh; layout.gem, layout_with_gem, with_gem; layout.graphopt, layout_with_graphopt, with_graphopt; layout.grid, layout.grid.3d, layout.grid.3d, layout_on_grid, on_grid; layout.mds, layout_with_mds, with_mds; layout.merge, layout_components, merge_coords, piecewise.layout, piecewise.layout; layout.norm, norm_coords; layout.sugiyama, layout_with_sugiyama, with_sugiyama; layout_randomly, randomly; layout_with_fr, with_fr; layout_with_kk, with_kk; layout_with_lgl, with_lgl; layout, layout_, print.igraph_layout_modifier, print.igraph_layout_spec; normalize

[Package igraph version 1.0.1 Index]