Large Graph Layout

A layout generator for larger graphs.

Usage

layout_with_lgl(
  graph,
  maxiter = 150,
  maxdelta = vcount(graph),
  area = vcount(graph)^2,
  coolexp = 1.5,
  repulserad = area * vcount(graph),
  cellsize = sqrt(sqrt(area)),
  root = NULL
)

with_lgl(...)

Arguments

graph

The input graph

maxiter

The maximum number of iterations to perform (150).

maxdelta

The maximum change for a vertex during an iteration (the number of vertices).

area

The area of the surface on which the vertices are placed (square of the number of vertices).

coolexp

The cooling exponent of the simulated annealing (1.5).

repulserad

Cancellation radius for the repulsion (the area times the number of vertices).

cellsize

The size of the cells for the grid. When calculating the repulsion forces between vertices only vertices in the same or neighboring grid cells are taken into account (the fourth root of the number of area.

root

The id of the vertex to place at the middle of the layout. The default value is -1 which means that a random vertex is selected.

...

Passed to layout_with_lgl().

Value

A numeric matrix with two columns and as many rows as vertices.

Details

layout_with_lgl() is for large connected graphs, it is similar to the layout generator of the Large Graph Layout software (https://lgl.sourceforge.net/).

See also

Author

Gabor Csardi csardi.gabor@gmail.com