make_chordal_ring {igraph}	R Documentation

Create an extended chordal ring graph

Description

make_chordal_ring creates an extended chordal ring. An extended chordal ring is regular graph, each node has the same degree. It can be obtained from a simple ring by adding some extra edges specified by a matrix. Let p denote the number of columns in the ‘W’ matrix. The extra edges of vertex i are added according to column i mod p in ‘W’. The number of extra edges is the number of rows in ‘W’: for each row j an edge i->i+w[ij] is added if i+w[ij] is less than the number of total nodes. See also Kotsis, G: Interconnection Topologies for Parallel Processing Systems, PARS Mitteilungen 11, 1-6, 1993.

Usage

make_chordal_ring(n, w)

chordal_ring(...)

Arguments

n
The number of vertices.

w
A matrix which specifies the extended chordal ring. See details below.

...
Passed to make_chordal_ring.

Value

An igraph graph.

See Also

Other determimistic constructors: graph_from_atlas, graph_from_edgelist, graph_from_literal, make_empty_graph, make_full_citation_graph, make_full_graph, make_graph, make_lattice, make_ring, make_star, make_tree

Examples

chord <- make_chordal_ring(15,
    matrix(c(3, 12, 4, 7, 8, 11), nr = 2))

[Package igraph version 1.2.2 Index]