make_clusters {igraph}	R Documentation

Creates a communities object.

Description

This is useful to integrate the results of community finding algorithms that are not included in igraph.

Usage

make_clusters(graph, membership = NULL, algorithm = NULL, merges = NULL,
  modularity = TRUE)

Arguments

graph
The graph of the community structure.

membership
The membership vector of the community structure, a numeric vector denoting the id of the community for each vertex. It might be NULL for hierarchical community structures.

algorithm
Character string, the algorithm that generated the community structure, it can be arbitrary.

merges
A merge matrix, for hierarchical community structures (or NULL otherwise.

modularity
Modularity value of the community structure. If this is TRUE and the membership vector is available, then it the modularity values is calculated automatically.

Value

A communities object.

[Package igraph version 1.2.2 Index]