--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.3 --- {% raw %}

make_from_prufer {igraph}	R Documentation

Create an undirected tree graph from its Prufer sequence

Description

make_from_prufer creates an undirected tree graph from its Prufer sequence.

Usage

make_from_prufer(prufer)

from_prufer(...)

Arguments

prufer
The Prufer sequence to convert into a graph

...
Passed to make_from_prufer

Details

The Prufer sequence of a tree graph with n labeled vertices is a sequence of n-2 numbers, constructed as follows. If the graph has more than two vertices, find a vertex with degree one, remove it from the tree and add the label of the vertex that it was connected to to the sequence. Repeat until there are only two vertices in the remaining graph.

Value

A graph object.

See Also

to_prufer to convert a graph into its Prufer sequence

Examples

g <- make_tree(13, 3)
to_prufer(g)

[Package igraph version 1.3.3 Index]
{% endraw %}