--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.4 --- {% raw %}

make_full_bipartite_graph {igraph}	R Documentation

Create a full bipartite graph

Description

Bipartite graphs are also called two-mode by some. This function creates a bipartite graph in which every possible edge is present.

Usage

make_full_bipartite_graph(
  n1,
  n2,
  directed = FALSE,
  mode = c("all", "out", "in")
)

full_bipartite_graph(...)

Arguments

n1
The number of vertices of the first kind.

n2
The number of vertices of the second kind.

directed
Logical scalar, whether the graphs is directed.

mode
Scalar giving the kind of edges to create for directed graphs. If this is ‘out’ then all vertices of the first kind are connected to the others; ‘in’ specifies the opposite direction; ‘all’ creates mutual edges. This argument is ignored for undirected graphs.x

...
Passed to make_full_bipartite_graph.

Details

Bipartite graphs have a ‘type’ vertex attribute in igraph, this is boolean and FALSE for the vertices of the first kind and TRUE for vertices of the second kind.

Value

An igraph graph, with the ‘type’ vertex attribute set.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

make_full_graph for creating one-mode full graphs

Examples

g <- make_full_bipartite_graph(2, 3)
g2 <- make_full_bipartite_graph(2, 3, directed=TRUE)
g3 <- make_full_bipartite_graph(2, 3, directed=TRUE, mode="in")
g4 <- make_full_bipartite_graph(2, 3, directed=TRUE, mode="all")

[Package igraph version 1.3.4 Index]
{% endraw %}