Kautz graphs

Kautz graphs are labeled graphs representing the overlap of strings.

Usage

make_kautz_graph(m, n)

kautz_graph(...)

Arguments

m

Integer scalar, the size of the alphabet. See details below.

n

Integer scalar, the length of the labels. See details below.

...

Passed to make_kautz_graph().

Value

A graph object.

Details

A Kautz graph is a labeled graph, vertices are labeled by strings of length n+1 above an alphabet with m+1 letters, with the restriction that every two consecutive letters in the string must be different. There is a directed edge from a vertex v to another vertex w if it is possible to transform the string of v into the string of w by removing the first letter and appending a letter to it.

Kautz graphs have some interesting properties, see eg. Wikipedia for details.

See also

Author

Gabor Csardi csardi.gabor@gmail.com, the first version in R was written by Vincent Matossian.

Examples

make_line_graph(make_kautz_graph(2, 1))
#> [[1]]
#> [1] 12
#> 
#> [[2]]
#> [1] TRUE
#> 
#> [[3]]
#>  [1]  4  8  4  8  5  9  5  9  0 10  0 10  1 11  1 11  2  6  2  6  3  7  3  7
#> 
#> [[4]]
#>  [1]  0  0  1  1  2  2  3  3  4  4  5  5  6  6  7  7  8  8  9  9 10 10 11 11
#> 
#> [[5]]
#>  [1]  8 10 12 14 16 18 20 22  0  2  4  6 17 19 21 23  1  3  5  7  9 11 13 15
#> 
#> [[6]]
#>  [1]  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23
#> 
#> [[7]]
#>  [1]  0  2  4  6  8 10 12 14 16 18 20 22 24
#> 
#> [[8]]
#>  [1]  0  2  4  6  8 10 12 14 16 18 20 22 24
#> 
#> [[9]]
#> [[9]][[1]]
#> [1] 1 0 1
#> 
#> [[9]][[2]]
#> [[9]][[2]]$name
#> [1] "Line graph"
#> 
#> 
#> [[9]][[3]]
#> list()
#> 
#> [[9]][[4]]
#> list()
#> 
#> 
#> [[10]]
#> <environment: 0x561c8564dd10>
#> 
#> attr(,"class")
#> [1] "igraph"
make_kautz_graph(2, 2)
#> [[1]]
#> [1] 12
#> 
#> [[2]]
#> [1] TRUE
#> 
#> [[3]]
#>  [1]  0  0  1  1  2  2  3  3  4  4  5  5  6  6  7  7  8  8  9  9 10 10 11 11
#> 
#> [[4]]
#>  [1]  4  5  6  7  8  9 10 11  0  1  2  3  8  9 10 11  0  1  2  3  4  5  6  7
#> 
#> [[5]]
#>  [1]  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23
#> 
#> [[6]]
#>  [1]  8 16  9 17 10 18 11 19  0 20  1 21  2 22  3 23  4 12  5 13  6 14  7 15
#> 
#> [[7]]
#>  [1]  0  2  4  6  8 10 12 14 16 18 20 22 24
#> 
#> [[8]]
#>  [1]  0  2  4  6  8 10 12 14 16 18 20 22 24
#> 
#> [[9]]
#> [[9]][[1]]
#> [1] 1 0 1
#> 
#> [[9]][[2]]
#> [[9]][[2]]$name
#> [1] "Kautz graph 2-2"
#> 
#> [[9]][[2]]$m
#> [1] 2
#> 
#> [[9]][[2]]$n
#> [1] 2
#> 
#> 
#> [[9]][[3]]
#> list()
#> 
#> [[9]][[4]]
#> list()
#> 
#> 
#> [[10]]
#> <environment: 0x561c80f51db0>
#> 
#> attr(,"class")
#> [1] "igraph"