--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.2 --- {% raw %}

make_ring {igraph}	R Documentation

Create a ring graph

Description

A ring is a one-dimensional lattice and this function is a special case of make_lattice.

Usage

make_ring(n, directed = FALSE, mutual = FALSE, circular = TRUE)

ring(...)

Arguments

n
Number of vertices.

directed
Whether the graph is directed.

mutual
Whether directed edges are mutual. It is ignored in undirected graphs.

circular
Whether to create a circular ring. A non-circular ring is essentially a “line”: a tree where every non-leaf vertex has one child.

...
Passed to make_ring.

Value

An igraph graph.

See Also

Other deterministic constructors: graph_from_atlas(), graph_from_edgelist(), graph_from_literal(), make_chordal_ring(), make_empty_graph(), make_full_citation_graph(), make_full_graph(), make_graph(), make_lattice(), make_star(), make_tree()

Examples

print_all(make_ring(10))
print_all(make_ring(10, directed = TRUE, mutual = TRUE))

[Package igraph version 1.3.2 Index]
{% endraw %}