make_star {igraph}	R Documentation

Create a star graph, a tree with n vertices and n - 1 leaves

Description

star creates a star graph, in this every single vertex is connected to the center vertex and nobody else.

Usage

make_star(n, mode = c("in", "out", "mutual", "undirected"), center = 1)

star(...)

Arguments

n
Number of vertices.

mode
It defines the direction of the edges, in: the edges point to the center, out: the edges point from the center, mutual: a directed star is created with mutual edges, undirected: the edges are undirected.

center
ID of the center vertex.

...
Passed to make_star.

Value

An igraph graph.

See Also

Other determimistic constructors: graph_from_atlas, graph_from_edgelist, graph_from_literal, make_chordal_ring, make_empty_graph, make_full_citation_graph, make_full_graph, make_graph, make_lattice, make_ring, make_tree

Examples

make_star(10, mode = "out")
make_star(5, mode = "undirected")

[Package igraph version 1.2.2 Index]