measure.dynamics {igraph}	R Documentation

Measuring the driving force in evolving networks

Description

These functions assume a simple evolving network model and measure the functional form of a so-called attractiveness function governing the evolution of the network.

Usage

measure.dynamics.idage (graph, agebins = 300, iterations = 5, 
    error=TRUE, time.window = NULL, number = FALSE, cites=FALSE,
    norm.method="old")
measure.dynamics.id(graph, iterations = 5, error=TRUE, 
    time.window = NULL, number = FALSE, cites=FALSE, norm.method="old",
    debug=FALSE, debugdeg=0, which=2) 
measure.dynamics.d.d(graph, vtime, etime, iterations = 5)
measure.dynamics.citedcat.id.age(graph, categories, agebins = 300,
    iterations = 5, norm = c(1, 1, 1)) 
measure.dynamics.citingcat.id.age(graph, categories, agebins = 300,
    iterations = 5, norm = c(1, 1, 1))
measure.dynamics.lastcit(graph, agebins, iterations=5,
    norm.method="old", number=FALSE)
measure.dynamics.age(graph, agebins, iterations=5, norm.method="old",
    number=FALSE)
measure.dynamics.citedcat(graph, categories, iterations=5,
    number=FALSE, norm.method="old")
measure.dynamics.citingcat.citedcat(graph, categories, iterations=5,
    number=FALSE, norm.method="old", norm=c(1,1))

Arguments

graph 	 The graph of which the evolution is quantified. It is assumed that the vertices were added in increasing order of vertex id.

agebins 	 Numeric constant, the number of bins to use for measuring aging.

iterations 	 Numeric constant, number of iterations to perform while calculating the attractiveness and the total attractiveness function.

time.window

vtime

etime

categories

norm

number

norm.method

error

cites

debug

debugdeg

which

Details

The functions should be considered as experimental, so no detailed documentation yet. Sorry.

Value

TODO

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

[Package igraph version 0.5.2-2 Index]