--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.4 --- {% raw %}

norm_coords {igraph}	R Documentation

Normalize coordinates for plotting graphs

Description

Rescale coordinates linearly to be within given bounds.

Usage

norm_coords(
  layout,
  xmin = -1,
  xmax = 1,
  ymin = -1,
  ymax = 1,
  zmin = -1,
  zmax = 1
)

Arguments

layout
A matrix with two or three columns, the layout to normalize.

xmin, xmax
The limits for the first coordinate, if one of them or both are NULL then no normalization is performed along this direction.

ymin, ymax
The limits for the second coordinate, if one of them or both are NULL then no normalization is performed along this direction.

zmin, zmax
The limits for the third coordinate, if one of them or both are NULL then no normalization is performed along this direction.

Details

norm_coords normalizes a layout, it linearly transforms each coordinate separately to fit into the given limits.

Value

A numeric matrix with at the same dimension as layout.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

Other graph layouts: add_layout_(), component_wise(), layout_as_bipartite(), layout_as_star(), layout_as_tree(), layout_in_circle(), layout_nicely(), layout_on_grid(), layout_on_sphere(), layout_randomly(), layout_with_dh(), layout_with_fr(), layout_with_gem(), layout_with_graphopt(), layout_with_kk(), layout_with_lgl(), layout_with_mds(), layout_with_sugiyama(), layout_(), merge_coords(), normalize()

[Package igraph version 1.3.4 Index]
{% endraw %}