plot.igraph {igraph}	R Documentation

Plotting of graphs

Description

plot.graph is able to plot graphs to any R device. It is the non-interactive companion of the tkplot function.

Usage

plot.igraph(x, axes=FALSE, xlab="", ylab="", add=FALSE,
      xlim=c(-1,1), ylim=c(-1,1), main="", sub="", ...)

Arguments

x 	 The graph to plot.

axes 	 Logical, whether to plot axes, defaults to FALSE.

xlab 	 The label of the horizontal axis. Defaults to the empty string.

ylab 	 The label of the vertical axis. Defaults to the empty string.

add 	 Logical scalar, whether to add the plot to the current device, or delete the device's current contents first.

xlim 	 The limits for the horizontal axis, it is unlikely that you want to modify this.

ylim 	 The limits for the vertical axis, it is unlikely that you want to modify this.

main 	 Main title.

sub 	 Subtitle.

... 	 Additional arguments, passed to plot.

Details

One convenient way to plot graphs is to plot with tkplot first, handtune the placement of the vertices, query the coordinates by the tkplot.getcoords function and use them with plot to plot the graph to any R device.

Value

Returns NULL, invisibly.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

layout for different layouts, igraph.plotting for the detailed description of the plotting parameters and tkplot and rglplot for other graph plotting functions.

Examples

g <- graph.ring(10)
## Not run: plot(g, layout=layout.kamada.kawai, vertex.color="green")

[Package igraph version 0.5.2-2 Index]