print.igraph {igraph}	R Documentation

Print graphs to the terminal

Description

These functions attempt to print a graph to the terminal in a human readable form.

Usage

## S3 method for class 'igraph':
print(x,
  graph.attributes=igraph.par("print.graph.attributes"),
  vertex.attributes=igraph.par("print.vertex.attributes"),
  edge.attributes=igraph.par("print.edge.attributes"),
  names=TRUE, quote.names=TRUE,
  ...)
## S3 method for class 'bgraph':
print(x,
  ...)
## S3 method for class 'igraph':
summary(object, ...)

Arguments

x 	 The graph to print.

graph.attributes 	 Logical constant, whether to print graph attributes.

vertex.attributes 	 Logical constant, whether to print vertex attributes.

edge.attributes 	 Logical constant, whether to print edge attributes.

names 	 Logical constant, whether to print symbolic vertex names (ie. the name vertex attribute) or vertex ids.

quote.names 	 Logical scalar, whether to quote symbolic vertex names.

object 	 The graph of which the summary will be printed.

... 	 Additional agruments, print.bgraph passes these to print.igraph.

Details

summary.igraph prints the number of vertices, edges and whether the graph is directed. print.igraph prints the same information, aand also lists the edges, and optionally graph, vertex and/or edge attributes.

As of igraph 0.4 print.igraph uses the max.print option, see options for details.

print.bgraph prints a bgraph object, a graph together with its cohesive block hierarchy, see cohesive.blocks for details.

Value

All these functions return the graph invisibly.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

Examples

g <- graph.ring(10)
g
summary(g)

[Package igraph version 0.5.2-2 Index]