print.igraph.es {igraph}	R Documentation

Print an edge sequence to the screen

Description

For long edge sequences, the printing is truncated to fit to the screen. Use print explicitly and the codefull argument to see the full sequence.

Usage

## S3 method for class 'igraph.es'
print(x, full = igraph_opt("print.full"), ...)

Arguments

x
An edge sequence.

full
Whether to show the full sequence, or truncate the output to the screen size.

...
Currently ignored.

Details

Edge sequences created with the double bracket operator are printed differently, together with all attributes of the edges in the sequence, as a table.

Value

The edge sequence, invisibly.

See Also

Other vertex and edge sequences: E, V, igraph-es-attributes, igraph-es-indexing2, igraph-es-indexing, igraph-vs-attributes, igraph-vs-indexing2, igraph-vs-indexing, print.igraph.vs

Examples

# Unnamed graphs
g <- make_ring(10)
E(g)

# Named graphs
g2 <- make_ring(10) %>%
  set_vertex_attr("name", value = LETTERS[1:10])
E(g2)

# All edges in a long sequence
g3 <- make_ring(200)
E(g3)
E(g3) %>% print(full = TRUE)

# Metadata
g4 <- make_ring(10) %>%
  set_vertex_attr("name", value = LETTERS[1:10]) %>%
  set_edge_attr("weight", value = 1:10) %>%
  set_edge_attr("color", value = "green")
E(g4)
E(g4)[[]]
E(g4)[[1:5]]

[Package igraph version 1.2.2 Index]