print.igraph.vs {igraph}	R Documentation

Show a vertex sequence on the screen

Description

For long vertex sequences, the printing is truncated to fit to the screen. Use print explicitly and the full argument to see the full sequence.

Usage

## S3 method for class 'igraph.vs'
print(x, full = igraph_opt("print.full"), ...)

Arguments

x
A vertex sequence.

full
Whether to show the full sequence, or truncate the output to the screen size.

...
These arguments are currently ignored.

Details

Vertex sequence created with the double bracket operator are printed differently, together with all attributes of the vertices in the sequence, as a table.

Value

The vertex sequence, invisibly.

See Also

Other vertex and edge sequences: E, V, igraph-es-attributes, igraph-es-indexing2, igraph-es-indexing, igraph-vs-attributes, igraph-vs-indexing2, igraph-vs-indexing, print.igraph.es

Examples

# Unnamed graphs
g <- make_ring(10)
V(g)

# Named graphs
g2 <- make_ring(10) %>%
  set_vertex_attr("name", value = LETTERS[1:10])
V(g2)

# All vertices in the sequence
g3 <- make_ring(1000)
V(g3)
print(V(g3), full = TRUE)

# Metadata
g4 <- make_ring(10) %>%
  set_vertex_attr("name", value = LETTERS[1:10]) %>%
  set_vertex_attr("color", value = "red")
V(g4)[[]]
V(g4)[[2:5, 7:8]]

[Package igraph version 1.2.2 Index]