print.igraphHRGConsensus {igraph}	R Documentation

Print a hierarchical random graph consensus tree to the screen

Description

Consensus dendrograms (igraphHRGConsensus objects) are printed simply by listing the children of each internal node of the dendrogram:

HRG consensus tree:
g1 -> 11 12 13 14 15 16 17 18 19 20
g2 -> 1  2  3  4  5  6  7  8  9  10   
g3 -> g1 g2

The root of the dendrogram is g3 (because it has no incoming edges), and it has two subgroups, g1 and g2.

Usage

## S3 method for class 'igraphHRGConsensus'
print(x, ...)

Arguments

x
igraphHRGConsensus object to print.

...
Ignored.

Value

The input object, invisibly, to allow method chaining.

See Also

Other hierarchical random graph functions: consensus_tree, fit_hrg, hrg-methods, hrg_tree, hrg, predict_edges, print.igraphHRG, sample_hrg

[Package igraph version 1.2.2 Index]