r_pal {igraph}	R Documentation

The default R palette

Description

This is the default R palette, to be able to reproduce the colors of older igraph versions. Its colors are appropriate for categories, but they are not very attractive.

Usage

r_pal(n)

Arguments

n
The number of colors to use, the maximum is eight.

Value

A character vector of color names.

See Also

Other palettes: categorical_pal, diverging_pal, sequential_pal

[Package igraph version 1.2.2 Index]