--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.1 --- {% raw %}

radius {igraph}	R Documentation

Radius of a graph

Description

The eccentricity of a vertex is its shortest path distance from the farthest other node in the graph. The smallest eccentricity in a graph is called its radius

Usage

radius(graph, mode = c("all", "out", "in", "total"))

Arguments

graph
The input graph, it can be directed or undirected.

mode
Character constant, gives whether the shortest paths to or from the given vertices should be calculated for directed graphs. If out then the shortest paths from the vertex, if in then to it will be considered. If all, the default, then the corresponding undirected graph will be used, edge directions will be ignored. This argument is ignored for undirected graphs.

Details

The eccentricity of a vertex is calculated by measuring the shortest distance from (or to) the vertex, to (or from) all vertices in the graph, and taking the maximum.

This implementation ignores vertex pairs that are in different components. Isolate vertices have eccentricity zero.

Value

A numeric scalar, the radius of the graph.

References

Harary, F. Graph Theory. Reading, MA: Addison-Wesley, p. 35, 1994.

See Also

eccentricity for the underlying calculations, codedistances for general shortest path calculations.

Examples

g <- make_star(10, mode="undirected")
eccentricity(g)
radius(g)

[Package igraph version 1.3.1 Index]
{% endraw %}