rewire.edges {igraph}	R Documentation

Rewires the endpoints of the edges of a graph randomly

Description

This function rewires the endpoints of the edges with a constant probability uniformly randomly to a new vertex in a graph.

Usage

rewire.edges(graph, prob)

Arguments

graph 	 The input graph

prob 	 The rewiring probability, a real number between zero and one.

Details

Note that this function might create graphs with multiple and/or loop edges.

Value

A new graph object.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

Examples

# Some random shortcuts shorten the distances on a lattice
g <- graph.lattice( length=100, dim=1, nei=5 )
average.path.length(g)
g <- rewire.edges( g, prob=0.05 )
average.path.length(g)

[Package igraph version 0.5.2-2 Index]