rglplot {igraph}	R Documentation

3D plotting of graphs with OpenGL

Description

Using the rgl package, rglplot plots a graph in 3D. The plot can be zoomed, rotated, shifted, etc. but the coordinates of the vertices is fixed.

Usage

rglplot(x, ...)

Arguments

x 	 The graph to plot.

... 	 Additional arguments, see igraph.plotting fo the details

Details

Note that rglplot is considered to be highly experimental. It is not very useful either. See igraph.plotting for the possible arguments.

Value

NULL, invisibly.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

igraph.plotting, plot.igraph for the 2D version, tkplot for interactive graph drawing in 2D.

Examples

## Not run: 
g <- graph.lattice( c(5,5,5) )
coords <- layout.fruchterman.reingold(g, dim=3)
rglplot(g, layout=coords)
## End(Not run)

[Package igraph version 0.5.2-2 Index]