running.mean {igraph}	R Documentation

Running mean of a time series

Description

running.mean calculates the running mean in a vector with the given bin width.

Usage

running.mean(v, binwidth)

Arguments

v 	 The numeric vector.

binwidth 	 Numeric constant, the size of the bin, should be meaningful, ie. smaller than the length of v.

Details

The running mean of v is a w vector of length length(v)-binwidth+1. The first element of w id the average of the first binwidth elements of v, the second element of w is the average of elements 2:(binwidth+1), etc.

Value

A numeric vector of length length(v)-binwidth+1

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

Examples

running.mean(1:100, 10)

[Package igraph version 0.5.2-2 Index]