Generate random graphs according to the random dot product graph model

In this model, each vertex is represented by a latent position vector. Probability of an edge between two vertices are given by the dot product of their latent position vectors.

Usage

sample_dot_product(vecs, directed = FALSE)

dot_product(...)

Arguments

vecs

A numeric matrix in which each latent position vector is a column.

directed

A logical scalar, TRUE if the generated graph should be directed.

...

Passed to sample_dot_product().

Value

An igraph graph object which is the generated random dot product graph.

Details

The dot product of the latent position vectors should be in the [0,1] interval, otherwise a warning is given. For negative dot products, no edges are added; dot products that are larger than one always add an edge.

References

Christine Leigh Myers Nickel: Random dot product graphs, a model for social networks. Dissertation, Johns Hopkins University, Maryland, USA, 2006.

See also

Author

Gabor Csardi csardi.gabor@gmail.com

Examples

## A randomly generated  graph
lpvs <- matrix(rnorm(200), 20, 10)
lpvs <- apply(lpvs, 2, function(x) {
  return(abs(x) / sqrt(sum(x^2)))
})
g <- sample_dot_product(lpvs)
g
#> [[1]]
#> [1] 10
#> 
#> [[2]]
#> [1] FALSE
#> 
#> [[3]]
#>  [1] 1 3 4 6 7 8 9 2 4 5 6 8 4 5 8 9 4 5 6 7 8 9 6 8 9 6 7 7 8 9 9
#> 
#> [[4]]
#>  [1] 0 0 0 0 0 0 0 1 1 1 1 1 2 2 2 2 3 3 3 3 3 3 4 4 4 5 5 6 6 7 8
#> 
#> [[5]]
#>  [1]  0  7  1  2  8 12 16  9 13 17  3 10 18 22 25  4 19 26 27  5 11 14 20 23 28
#> [26]  6 15 21 24 29 30
#> 
#> [[6]]
#>  [1]  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24
#> [26] 25 26 27 28 29 30
#> 
#> [[7]]
#>  [1]  0  0  1  2  3  7 10 15 19 25 31
#> 
#> [[8]]
#>  [1]  0  7 12 16 22 25 27 29 30 31 31
#> 
#> [[9]]
#> [[9]][[1]]
#> [1] 1 0 1
#> 
#> [[9]][[2]]
#> named list()
#> 
#> [[9]][[3]]
#> list()
#> 
#> [[9]][[4]]
#> list()
#> 
#> 
#> [[10]]
#> <environment: 0x563a47d099f8>
#> 
#> attr(,"class")
#> [1] "igraph"

## Sample latent vectors from the surface of the unit sphere
lpvs2 <- sample_sphere_surface(dim = 5, n = 20)
g2 <- sample_dot_product(lpvs2)
g2
#> [[1]]
#> [1] 20
#> 
#> [[2]]
#> [1] FALSE
#> 
#> [[3]]
#>   [1]  4  5  7  8  9 10 11 12 13 14 15 16 17 19  2  3  4  7  8  9 10 11 12 13 14
#>  [26] 15 16 17 18 19  3  4  5  6  7  9 10 11 12 13 14 15 18 19  4  5  6  7  8  9
#>  [51] 10 13 14 15 16 17 18 19  5  8  9 10 11 12 14 15 16 17  6  7  8 10 12 13 14
#>  [76] 16 17 18 19  8  9 10 11 12 14 17 18 19 10 11 12 13 14 15 16 17 18 19 11 12
#> [101] 13 14 15 16 17 18 19 11 13 16 19 11 12 14 15 16 18 13 14 15 17 19 14 18 16
#> [126] 18 15 16 17 18 16 18 17 18 19 19
#> 
#> [[4]]
#>   [1]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  1  1  1  1  1  1  1  1  1  1  1
#>  [26]  1  1  1  1  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3
#>  [51]  3  3  3  3  3  3  3  3  4  4  4  4  4  4  4  4  4  4  5  5  5  5  5  5  5
#>  [76]  5  5  5  5  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7  7  7  7  8  8
#> [101]  8  8  8  8  8  8  8  9  9  9  9 10 10 10 10 10 10 11 11 11 11 11 12 12 13
#> [126] 13 14 14 14 14 15 15 16 16 17 18
#> 
#> [[5]]
#>   [1]  14  15  30   0  16  31  44   1  32  45  58  33  46  68   2  17  34  47
#>  [19]  69   3  18  48  59  70  79   4  19  35  49  60  80   5  20  36  50  61
#>  [37]  71  81  88   6  21  37  62  82  89  98 107 111   7  22  38  63  72  83
#>  [55]  90  99 112   8  23  39  51  73  91 100 108 117   9  24  40  52  64  74
#>  [73]  84  92 101 113 118 122  10  25  41  53  65  93 102 114 119 126  11  26
#>  [91]  54  66  75  94 103 109 115 124 127 130  12  27  55  67  76  85  95 104
#> [109] 120 128 132  28  42  56  77  86  96 105 116 123 125 129 131 133  13  29
#> [127]  43  57  78  87  97 106 110 121 134 135
#> 
#> [[6]]
#>   [1]   0   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17
#>  [19]  18  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35
#>  [37]  36  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53
#>  [55]  54  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71
#>  [73]  72  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89
#>  [91]  90  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107
#> [109] 108 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125
#> [127] 126 127 128 129 130 131 132 133 134 135
#> 
#> [[7]]
#>  [1]   0   0   0   1   3   7  11  14  19  25  31  39  48  57  66  78  88 100 111
#> [20] 124 136
#> 
#> [[8]]
#>  [1]   0  14  30  44  58  68  79  88  98 107 111 117 122 124 126 130 132 134 135
#> [20] 136 136
#> 
#> [[9]]
#> [[9]][[1]]
#> [1] 1 0 1
#> 
#> [[9]][[2]]
#> named list()
#> 
#> [[9]][[3]]
#> list()
#> 
#> [[9]][[4]]
#> list()
#> 
#> 
#> [[10]]
#> <environment: 0x563a47e1f008>
#> 
#> attr(,"class")
#> [1] "igraph"