sample_gnm {igraph}	R Documentation

Generate random graphs according to the G(n,m) Erdos-Renyi model

Description

This model is very simple, every possible edge is created with the same constant probability.

Usage

sample_gnm(n, m, directed = FALSE, loops = FALSE)

gnm(...)

Arguments

n
The number of vertices in the graph.

m
The number of edges in the graph.

directed
Logical, whether the graph will be directed, defaults to FALSE.

loops
Logical, whether to add loop edges, defaults to FALSE.

...
Passed to sample_app.

Details

The graph has ‘n’ vertices and ‘m’ edges, and the ‘m’ edges are chosen uniformly randomly from the set of all possible edges. This set includes loop edges as well if the loops parameter is TRUE.

Value

A graph object.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

References

Erdos, P. and Renyi, A., On random graphs, Publicationes Mathematicae 6, 290–297 (1959).

See Also

sample_gnp, sample_pa

Examples

g <- sample_gnm(1000, 1000)
degree_distribution(g)

[Package igraph version 1.2.2 Index]