sample_gnp {igraph}	R Documentation

Generate random graphs according to the G(n,p) Erdos-Renyi model

Description

This model is very simple, every possible edge is created with the same constant probability.

Usage

sample_gnp(n, p, directed = FALSE, loops = FALSE)

gnp(...)

Arguments

n
The number of vertices in the graph.

p
The probability for drawing an edge between two arbitrary vertices (G(n,p) graph).

directed
Logical, whether the graph will be directed, defaults to FALSE.

loops
Logical, whether to add loop edges, defaults to FALSE.

...
Passed to sample_app.

Details

The graph has ‘n’ vertices and for each edge the probability that it is present in the graph is ‘p’.

Value

A graph object.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

References

Erdos, P. and Renyi, A., On random graphs, Publicationes Mathematicae 6, 290–297 (1959).

See Also

sample_gnm, sample_pa

Examples

g <- sample_gnp(1000, 1/1000)
degree_distribution(g)

[Package igraph version 1.2.2 Index]