--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.1 --- {% raw %}

sample_growing {igraph}	R Documentation

Growing random graph generation

Description

This function creates a random graph by simulating its stochastic evolution.

Usage

sample_growing(n, m = 1, directed = TRUE, citation = FALSE)

growing(...)

Arguments

n
Numeric constant, number of vertices in the graph.

m
Numeric constant, number of edges added in each time step.

directed
Logical, whether to create a directed graph.

citation
Logical. If TRUE a citation graph is created, ie. in each time step the added edges are originating from the new vertex.

...
Passed to sample_growing.

Details

This is discrete time step model, in each time step a new vertex is added to the graph and m new edges are created. If citation is FALSE these edges are connecting two uniformly randomly chosen vertices, otherwise the edges are connecting new vertex to uniformly randomly chosen old vertices.

Value

A new graph object.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

sample_pa, sample_gnp

Examples

g <- sample_growing(500, citation=FALSE)
g2 <- sample_growing(500, citation=TRUE)

[Package igraph version 1.3.1 Index]
{% endraw %}