--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.5 --- {% raw %}

sample_hrg {igraph}	R Documentation

Sample from a hierarchical random graph model

Description

sample_hrg samples a graph from a given hierarchical random graph model.

Usage

sample_hrg(hrg)

Arguments

hrg
A hierarchical random graph model.

Value

An igraph graph.

See Also

Other hierarchical random graph functions: consensus_tree(), fit_hrg(), hrg-methods, hrg_tree(), hrg(), predict_edges(), print.igraphHRGConsensus(), print.igraphHRG()

[Package igraph version 1.3.5 Index]
{% endraw %}