sample_islands {igraph}	R Documentation

A graph with subgraphs that are each a random graph.

Description

Create a number of Erdos-Renyi random graphs with identical parameters, and connect them with the specified number of edges.

Usage

sample_islands(islands.n, islands.size, islands.pin, n.inter)

Arguments

islands.n
The number of islands in the graph.

islands.size
The size of islands in the graph.

islands.pin
The probability to create each possible edge into each island.

n.inter
The number of edges to create between two islands.

Value

An igraph graph.

Examples

g <- sample_islands(3, 10, 5/10, 1)
oc <- cluster_optimal(g)
oc

Author(s)

Samuel Thiriot

See Also

sample_gnp

[Package igraph version 1.2.2 Index]