--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.5 --- {% raw %}

sample_k_regular {igraph}	R Documentation

Create a random regular graph

Description

Generate a random graph where each vertex has the same degree.

Usage

sample_k_regular(no.of.nodes, k, directed = FALSE, multiple = FALSE)

Arguments

no.of.nodes
Integer scalar, the number of vertices in the generated graph.

k
Integer scalar, the degree of each vertex in the graph, or the out-degree and in-degree in a directed graph.

directed
Logical scalar, whether to create a directed graph.

multiple
Logical scalar, whether multiple edges are allowed.

Details

This game generates a directed or undirected random graph where the degrees of vertices are equal to a predefined constant k. For undirected graphs, at least one of k and the number of vertices must be even.

The game simply uses sample_degseq with appropriately constructed degree sequences.

Value

An igraph graph.

Author(s)

Tamas Nepusz ntamas@gmail.com

See Also

sample_degseq for a generator with prescribed degree sequence.

Examples

## A simple ring
ring <- sample_k_regular(10, 2)
plot(ring)

## k-regular graphs on 10 vertices, with k=1:9
k10 <- lapply(1:9, sample_k_regular, no.of.nodes=10)

layout(matrix(1:9, nrow=3, byrow=TRUE))
sapply(k10, plot, vertex.label=NA)

[Package igraph version 1.3.5 Index]
{% endraw %}