Random citation graphs

sample_last_cit() creates a graph, where vertices age, and gain new connections based on how long ago their last citation happened.

Usage

sample_last_cit(
  n,
  edges = 1,
  agebins = n/7100,
  pref = (1:(agebins + 1))^-3,
  directed = TRUE
)

last_cit(...)

sample_cit_types(
  n,
  edges = 1,
  types = rep(0, n),
  pref = rep(1, length(types)),
  directed = TRUE,
  attr = TRUE
)

cit_types(...)

sample_cit_cit_types(
  n,
  edges = 1,
  types = rep(0, n),
  pref = matrix(1, nrow = length(types), ncol = length(types)),
  directed = TRUE,
  attr = TRUE
)

cit_cit_types(...)

Arguments

n

Number of vertices.

edges

Number of edges per step.

agebins

Number of aging bins.

pref

Vector (sample_last_cit() and sample_cit_types() or matrix (sample_cit_cit_types()) giving the (unnormalized) citation probabilities for the different vertex types.

directed

Logical scalar, whether to generate directed networks.

...

Passed to the actual constructor.

types

Vector of length ‘n’, the types of the vertices. Types are numbered from zero.

attr

Logical scalar, whether to add the vertex types to the generated graph as a vertex attribute called ‘type’.

Value

A new graph.

Details

sample_cit_cit_types() is a stochastic block model where the graph is growing.

sample_cit_types() is similarly a growing stochastic block model, but the probability of an edge depends on the (potentially) cited vertex only.

See also

Author

Gabor Csardi csardi.gabor@gmail.com