sample_motifs {igraph}	R Documentation

Graph motifs

Description

Graph motifs are small connected subgraphs with a well-defined structure. These functions search a graph for various motifs.

Usage

sample_motifs(graph, size = 3, cut.prob = rep(0, size),
  sample.size = vcount(graph)/10, sample = NULL)

Arguments

graph
Graph object, the input graph.

size
The size of the motif, currently 3 and 4 are supported only.

cut.prob
Numeric vector giving the probabilities that the search graph is cut at a certain level. Its length should be the same as the size of the motif (the size argument). By default no cuts are made.

sample.size
The number of vertices to use as a starting point for finding motifs. Only used if the sample argument is NULL.

sample
If not NULL then it specifies the vertices to use as a starting point for finding motifs.

Details

sample_motifs estimates the total number of motifs of a given size in a graph based on a sample.

Value

A numeric scalar, an estimate for the total number of motifs in the graph.

See Also

isomorphism_class

Other graph motifs: count_motifs, motifs

Examples

g <- barabasi.game(100)
motifs(g, 3)
count_motifs(g, 3)
sample_motifs(g, 3)

[Package igraph version 1.2.2 Index]