--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.2 --- {% raw %}

sample_tree {igraph}	R Documentation

Sample trees randomly and uniformly

Description

sample_tree generates a random with a given number of nodes uniform at random from the set of labelled trees.

Usage

sample_tree(n, directed = FALSE, method = c("lerw", "prufer"))

Arguments

n
The number of nodes in the tree

directed
Whether to create a directed tree. The edges of the tree are oriented away from the root.

method
The algorithm to use to generate the tree. ‘prufer’ samples Prufer sequences uniformly and then converts the sampled sequence to a tree. ‘lerw’ performs a loop-erased random walk on the complete graph to uniformly sampleits spanning trees. (This is also known as Wilson's algorithm). The default is ‘lerw’. Note that the method based on Prufer sequences does not support directed trees at the moment.

Details

In other words, the function generates each possible labelled tree with the given number of nodes with the same probability.

Value

A graph object.

Examples

g <- sample_tree(100, method="lerw")

[Package igraph version 1.3.2 Index]
{% endraw %}