set_graph_attr {igraph}	R Documentation

Set a graph attribute

Description

An existing attribute with the same name is overwritten.

Usage

set_graph_attr(graph, name, value)

Arguments

graph
The graph.

name
The name of the attribute to set.

value
New value of the attribute.

Value

The graph with the new graph attribute added or set.

See Also

Other graph attributes: delete_edge_attr, delete_graph_attr, delete_vertex_attr, edge_attr<-, edge_attr_names, edge_attr, graph_attr<-, graph_attr_names, graph_attr, igraph-dollar, igraph-vs-attributes, set_edge_attr, set_vertex_attr, vertex_attr<-, vertex_attr_names, vertex_attr

Examples

g <- make_ring(10) %>%
  set_graph_attr("layout", layout_with_fr)
g
plot(g)

[Package igraph version 1.2.2 Index]