--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.3 --- {% raw %}

set_vertex_attr {igraph}	R Documentation

Set vertex attributes

Description

Set vertex attributes

Usage

set_vertex_attr(graph, name, index = V(graph), value)

Arguments

graph
The graph.

name
The name of the attribute to set.

index
An optional vertex sequence to set the attributes of a subset of vertices.

value
The new value of the attribute for all (or index) vertices.

Value

The graph, with the vertex attribute added or set.

See Also

Other graph attributes: delete_edge_attr(), delete_graph_attr(), delete_vertex_attr(), edge_attr<-(), edge_attr_names(), edge_attr(), graph_attr<-(), graph_attr_names(), graph_attr(), igraph-dollar, igraph-vs-attributes, set_edge_attr(), set_graph_attr(), vertex_attr<-(), vertex_attr_names(), vertex_attr()

Examples

g <- make_ring(10) %>%
  set_vertex_attr("label", value = LETTERS[1:10])
g
plot(g)

[Package igraph version 1.3.3 Index]
{% endraw %}