--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.3 --- {% raw %}

srand {igraph}	R Documentation

Deprecated function, used to set random seed of the C library's RNG

Description

Deprecated function, used to set random seed of the C library's RNG

Usage

srand(seed)

Arguments

seed
Ignored.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

[Package igraph version 1.3.3 Index]
{% endraw %}