In- or out- component of a vertex

Finds all vertices reachable from a given vertex, or the opposite: all vertices from which a given vertex is reachable via a directed path.

Usage

subcomponent(graph, v, mode = c("all", "out", "in"))

Arguments

graph

The graph to analyze.

v

The vertex to start the search from.

mode

Character string, either “in”, “out” or “all”. If “in” all vertices from which v is reachable are listed. If “out” all vertices reachable from v are returned. If “all” returns the union of these. It is ignored for undirected graphs.

Value

Numeric vector, the ids of the vertices in the same component as v.

Details

A breadth-first search is conducted starting from vertex v.

See also

Author

Gabor Csardi csardi.gabor@gmail.com

Examples

g <- sample_gnp(100, 1 / 200)
subcomponent(g, 1, "in")
#> + 1/100 vertex, from f095d18:
#> [1] 1
subcomponent(g, 1, "out")
#> + 1/100 vertex, from f095d18:
#> [1] 1
subcomponent(g, 1, "all")
#> + 1/100 vertex, from f095d18:
#> [1] 1