subgraph {igraph}	R Documentation

Subgraph of a graph

Description

subgraph creates a subgraph of a graph, containing only the specified vertices and all the edges among them.

Usage

subgraph(graph, v)

Arguments

graph 	 The original graph.

v 	 Numeric vector, the vertices of the original graph which will form the subgraph.

Details

The ids of the vertices will change in the subgraph of course since these are always consecutive.

Value

A new graph object.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

Examples

g <- graph.ring(10)
g2 <- subgraph(g, 1:7)

[Package igraph version 0.5.2-2 Index]