tail_of {igraph}	R Documentation

Tails of the edge(s) in a graph

Description

For undirected graphs, head and tail is not defined. In this case tail_of returns vertices incident to the supplied edges, and head_of returns the other end(s) of the edge(s).

Usage

tail_of(graph, es)

Arguments

graph
The input graph.

es
The edges to query.

Value

A vertex sequence with the tail(s) of the edge(s).

See Also

Other structural queries: [.igraph, [[.igraph, adjacent_vertices, are_adjacent, ends, get.edge.ids, gorder, gsize, head_of, incident_edges, incident, is_directed, neighbors

[Package igraph version 1.2.2 Index]