tkigraph {igraph}	R Documentation

Experimental basic igraph GUI

Description

This functions starts an experimental GUI to some igraph functions. The GUI was written in Tcl/Tk, so it is cross platform.

Usage

tkigraph()

Details

tkigraph has its own online help system, please see that for the details about how to use it.

Value

Returns NULL, invisibly.

Author(s)

Gabor Csardi Gabor.Csardi@unil.ch

See Also

tkplot for interactive plotting of graphs.

[Package igraph version 0.5.2-2 Index]