triad.census {igraph}	R Documentation

Triad census, subgraphs with three vertices

Description

This function counts the different subgraphs of three vertices in a graph.

Usage

triad.census(graph)

Arguments

graph 	 The input graph, it should be directed. An undirected graph results a warning, and undefined results.

Details

Triad census was defined by David and Leinhardt (see References below). Every triple of vertices (A, B, C) are classified into the 16 possible states:

003

A,B,C, the empty graph.

012

A->B, C, the graph with a single directed edge.

102

A<->B, C, the graph with a mutual connection between two vertices.

021D

A<-B->C, the out-star.

021U

A->B<-C, the in-star.

021C

A->B->C, directed line.

111D

A<->B<-C.

111U

A<->B->C.

030T

A->B<-C, A->C.

030C

A<-B<-C, A->C.

201

A<->B<->C.

120D

A<-B->C, A<->C.

120U

A->B<-C, A<->C.

120C

A->B->C, A<->C.

210

A->B<->C, A<->C.

300

A<->B<->C, A<->C, the complete graph.

This functions uses the RANDESU motif finder algorithm to find and count the subgraphs, see graph.motifs.

Value

A numeric vector, the subgraph counts, in the order given in the above description.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

References

See also Davis, J.A. and Leinhardt, S. (1972). The Structure of Positive Interpersonal Relations in Small Groups. In J. Berger (Ed.), Sociological Theories in Progress, Volume 2, 218-251. Boston: Houghton Mifflin.

See Also

dyad.census for classifying binary relationships, graph.motifs for the underlying implementation.

Examples

g <- erdos.renyi.game(15, 45, type="gnm", dir=TRUE)
triad.census(g)

[Package igraph version 0.5.2-2 Index]