igraph.undocumented {igraph}	R Documentation

Undocumented and unsupportted igraph functions

Description

These functions are still in the alpha stage or their arguments are expected to change, so they're not documented yet. They are also not very useful for the general audience.

Usage

lastcit.game(n, edges=1, agebins=n/7100, pref=(1:(agebins+1))^-3,
     directed=TRUE)
cited.type.game(n, edges=1, types=rep(0, n),
     pref=rep(1, length(types)),
     directed=TRUE, attr=TRUE)
citing.cited.type.game(n, edges=1, types=rep(0, n),
    pref=matrix(1, nr=length(types), nc=length(types)),
    directed=TRUE, attr=TRUE)

Arguments

n

edges

agebins

pref

directed

types

attr

Value

A new graph.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

[Package igraph version 0.5.2-2 Index]