union {igraph}	R Documentation

Union of two or more sets

Description

This is an S3 generic function. See methods("union") for the actual implementations for various S3 classes. Initially it is implemented for igraph graphs and igraph vertex and edge sequences. See union.igraph, and union.igraph.vs.

Usage

union(...)

Arguments

...
Arguments, their number and interpretation depends on the function that implements union.

Value

Depends on the function that implements this method.

[Package igraph version 1.2.2 Index]