union.igraph.vs {igraph}	R Documentation

Union of vertex sequences

Description

Union of vertex sequences

Usage

## S3 method for class 'igraph.vs'
union(...)

Arguments

...
The vertex sequences to take the union of.

Details

They must belong to the same graph. Note that this function has ‘set’ semantics and the multiplicity of vertices is lost in the result. (This is to match the behavior of the based unique function.)

Value

A vertex sequence that contains all vertices in the given sequences, exactly once.

See Also

Other vertex and edge sequence operations: c.igraph.es, c.igraph.vs, difference.igraph.es, difference.igraph.vs, igraph-es-indexing2, igraph-es-indexing, igraph-vs-indexing2, igraph-vs-indexing, intersection.igraph.es, intersection.igraph.vs, rev.igraph.es, rev.igraph.vs, union.igraph.es, unique.igraph.es, unique.igraph.vs

Examples

g <- make_(ring(10), with_vertex_(name = LETTERS[1:10]))
union(V(g)[1:6], V(g)[5:10])

[Package igraph version 1.2.2 Index]