upgrade_graph {igraph}	R Documentation

Igraph data structure versions

Description

Igraph's internal data representation changes sometimes between versions. This means that it is not possible to use igraph objects that were created (and possibly saved to a file) with an older igraph version.

Usage

upgrade_graph(graph)

Arguments

graph
The input graph.

Details

graph_version queries the current data format, or the data format of a possibly older igraph graph.

upgrade_graph can convert an older data format to the current one.

Value

The graph in the current format.

See Also

graph_version to check the current data format version or the version of a graph.

[Package igraph version 1.2.2 Index]