vertex {igraph}	R Documentation

Helper function for adding and deleting vertices

Description

This is a helper function that simplifies adding and deleting vertices to/from graphs.

Usage

vertex(...)

vertices(...)

Arguments

...
See details below.

Details

vertices is an alias for vertex.

When adding vertices via +, all unnamed arguments are interpreted as vertex names of the new vertices. Named arguments are interpreted as vertex attributes for the new vertices.

When deleting vertices via -, all arguments of vertex (or vertices) are concatenated via c() and passed to delete_vertices.

Value

A special object that can be used with together with igraph graphs and the plus and minus operators.

See Also

Other functions for manipulating graph structure: +.igraph, add_edges, add_vertices, delete_edges, delete_vertices, edge, igraph-minus, path

Examples

g <- make_(ring(10), with_vertex_(name = LETTERS[1:10])) +
  vertices('X', 'Y')
g
plot(g)

[Package igraph version 1.2.2 Index]