vertex_attr<- {igraph}	R Documentation

Set one or more vertex attributes

Description

Set one or more vertex attributes

Usage

vertex_attr(graph, name, index = V(graph)) <- value

Arguments

graph
The graph.

name
The name of the vertex attribute to set. If missing, then value must be a named list, and its entries are set as vertex attributes.

index
An optional vertex sequence to set the attributes of a subset of vertices.

value
The new value of the attribute(s) for all (or index) vertices.

Value

The graph, with the vertex attribute(s) added or set.

See Also

Other graph attributes: delete_edge_attr, delete_graph_attr, delete_vertex_attr, edge_attr<-, edge_attr_names, edge_attr, graph_attr<-, graph_attr_names, graph_attr, igraph-dollar, igraph-vs-attributes, set_edge_attr, set_graph_attr, set_vertex_attr, vertex_attr_names, vertex_attr

Examples

g <- make_ring(10)
vertex_attr(g) <- list(name = LETTERS[1:10],
                        color = rep("yellow", gorder(g)))
vertex_attr(g, "label") <- V(g)$name
g
plot(g)

[Package igraph version 1.2.2 Index]