walktrap.community {igraph}	R Documentation

Community strucure via short random walks

Description

This function tries to find densely connected subgraphs, also called communities in a graph via random walks. The idea is that short random walks tend to stay in the same community.

Usage

walktrap.community(graph, weights = E(graph)$weight, steps = 4, merges =
          TRUE, modularity = FALSE, labels = TRUE)

Arguments

graph 	 The input graph.

weights 	 The edge weights.

steps 	 The length of the random walks to perform.

merges 	 Logical scalar, whether to include the merge matrix in the result.

modularity 	 Logical scalar, whether to include the vector of the modularity scores in the result.

labels 	 Logical scalar, if TRUE and the graph has a vertex attribute called name then it will be included in the result, in the list member called labels.

Details

This function is the implementation of the Walktrap community finding algorithm, see Pascal Pons, Matthieu Latapy: Computing communities in large networks using random walks, http://arxiv.org/abs/physics/0512106

Value

A named list with three members:

merges 	 The merges performed by the algorithm will be stored here. Each merge is a row in a two-column matrix and contains the ids of the merged communities. Communities are numbered from zero and cluster number smaller than the number of nodes in the network belong to the individual vertices as singleton communities. In each step a new community is created from two other communities and its id will be one larger than the largest community id so far. This means that before the first merge we have n communities (the number of vertices in the graph) numbered from zero to n-1. The first merge created community n, the second community n+1, etc.

modularity 	 Numeric vector, the modularity score of the current community structure after each merge operation.

labels 	 The labels of the vertices in the graph. The name vertex attribute will be copied here, if exists.

Author(s)

Pascal Pons google@for.it and Gabor Csardi csardi@rmki.kfki.hu for the R and igraph interface

References

Pascal Pons, Matthieu Latapy: Computing communities in large networks using random walks, http://arxiv.org/abs/physics/0512106

See Also

modularity and fastgreedy.community, spinglass.community, leading.eigenvector.community, edge.betweenness.community for other community detection methods.

Examples

g <- graph.full(5) %du% graph.full(5) %du% graph.full(5)
g <- add.edges(g, c(0,5, 0,10, 5, 10))
walktrap.community(g)

[Package igraph version 0.5.2-2 Index]