without_attr {igraph}	R Documentation

Construtor modifier to remove all attributes from a graph

Description

Construtor modifier to remove all attributes from a graph

Usage

without_attr()

See Also

Other constructor modifiers: simplified, with_edge_, with_graph_, with_vertex_, without_loops, without_multiples

Examples

g1 <- make_ring(10)
g1

g2 <- make_(ring(10), without_attr())
g2

[Package igraph version 1.2.2 Index]