write.pajek.bgraph {igraph}	R Documentation

Write graphs and their cohesive block hierarchy as Pajek files

Description

Creates a series of Pajek-compatible files from a bgraph object as output by cohesive.blocks.

Usage

write.pajek.bgraph(graph, filename, hierarchy = FALSE)

Arguments

graph 	 A bgraph object as output by cohesive.blocks.

filename 	 The filename without file extension to use.

hierarchy 	 Logical. Should a separate .clu file be created for each block in the graph? (see details)

Details

This function writes at least two files to disk. First, the file filename.net contains the basic graph structure—vertices and edges. The second is the file filename.clu, which clusters the vertices according to vertex connectivity of their maxmally cohesive subgraph. This uses the same logic as the default vertex coloring in plot.bgraph.

If hierarchy is TRUE, more detailed and complete information is written to a further series of files. For each block a seperate .clu file is written, each named like “filename_block2(1).clu”, where the first number refers to the block number and the number in parenthesis refers to that block's cohesion. Finally, if hierarchy is TRUE, a tree representing the block hierarchy is written to filename_blocktree.net, allowing the subset-structure of the individual block structures to be determined.

Note

This function has not been thoroughly tested, and may contain bugs

Author(s)

Peter McMahan peter.mcmahan@gmail.com

References

Pajek website: http://vlado.fmf.uni-lj.si/pub/networks/pajek/

See Also

cohesive.blocks, plot.bgraph, write.graph

[Package igraph version 0.5.2-2 Index]