--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.4 --- {% raw %}

add_vertices {igraph}	R Documentation

Add vertices to a graph

Description

If attributes are supplied, and they are not present in the graph, their values for the original vertices of the graph are set to NA.

Usage

add_vertices(graph, nv, ..., attr = list())

Arguments

graph
The input graph.

nv
The number of vertices to add.

...
Additional arguments, they must be named, and they will be added as vertex attributes, for the newly added vertices. See also details below.

attr
A named list, its elements will be added as vertex attributes, for the newly added vertices. See also details below.

Value

The graph, with the vertices (and attributes) added.

See Also

Other functions for manipulating graph structure: +.igraph(), add_edges(), delete_edges(), delete_vertices(), edge(), igraph-minus, path(), vertex()

Examples

g <- make_empty_graph() %>%
  add_vertices(3, color = "red") %>%
  add_vertices(2, color = "green") %>%
  add_edges(c(1,2, 2,3, 3,4, 4,5))
g
V(g)[[]]
plot(g)

[Package igraph version 1.3.4 Index]
{% endraw %}