are_adjacent {igraph}	R Documentation

Are two vertices adjacent?

Description

The order of the vertices only matters in directed graphs, where the existence of a directed (v1, v2) edge is queried.

Usage

are_adjacent(graph, v1, v2)

Arguments

graph
The graph.

v1
The first vertex, tail in directed graphs.

v2
The second vertex, head in directed graphs.

Value

A logical scalar, TRUE is a (v1, v2) exists in the graph.

See Also

Other structural queries: [.igraph(), [[.igraph(), adjacent_vertices(), ends(), get.edge.ids(), gorder(), gsize(), head_of(), incident_edges(), incident(), is_directed(), neighbors(), tail_of()

Examples

ug <- make_ring(10)
ug

## IGRAPH 0bb90bf U--- 10 10 -- Ring graph
## + attr: name (g/c), mutual (g/l), circular (g/l)
## + edges from 0bb90bf:
##  [1] 1-- 2 2-- 3 3-- 4 4-- 5 5-- 6 6-- 7 7-- 8 8-- 9 9--10 1--10

are_adjacent(ug, 1, 2)

## [1] TRUE

are_adjacent(ug, 2, 1)

## [1] TRUE

dg <- make_ring(10, directed = TRUE)
dg

## IGRAPH c0d00b8 D--- 10 10 -- Ring graph
## + attr: name (g/c), mutual (g/l), circular (g/l)
## + edges from c0d00b8:
##  [1]  1-> 2  2-> 3  3-> 4  4-> 5  5-> 6  6-> 7  7-> 8  8-> 9  9->10 10-> 1

are_adjacent(ug, 1, 2)

## [1] TRUE

are_adjacent(ug, 2, 1)

## [1] TRUE

[Package igraph version 1.2.9 Index]