--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.5 --- {% raw %}

are_adjacent {igraph}	R Documentation

Are two vertices adjacent?

Description

The order of the vertices only matters in directed graphs, where the existence of a directed (v1, v2) edge is queried.

Usage

are_adjacent(graph, v1, v2)

Arguments

graph
The graph.

v1
The first vertex, tail in directed graphs.

v2
The second vertex, head in directed graphs.

Value

A logical scalar, TRUE is a (v1, v2) exists in the graph.

See Also

Other structural queries: [.igraph(), [[.igraph(), adjacent_vertices(), ends(), get.edge.ids(), gorder(), gsize(), head_of(), incident_edges(), incident(), is_directed(), neighbors(), tail_of()

Examples

ug <- make_ring(10)
ug
are_adjacent(ug, 1, 2)
are_adjacent(ug, 2, 1)

dg <- make_ring(10, directed = TRUE)
dg
are_adjacent(ug, 1, 2)
are_adjacent(ug, 2, 1)

[Package igraph version 1.3.5 Index]
{% endraw %}