--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.5 --- {% raw %}

articulation_points {igraph}	R Documentation

Articulation points and bridges of a graph

Description

articulation_points finds the articulation points (or cut vertices)

Usage

articulation_points(graph)

bridges(graph)

Arguments

graph
The input graph. It is treated as an undirected graph, even if it is directed.

Details

Articulation points or cut vertices are vertices whose removal increases the number of connected components in a graph. Similarly, bridges or cut-edges are edges whose removal increases the number of connected components in a graph. If the original graph was connected, then the removal of a single articulation point or a single bridge makes it undirected. If a graph contains no articulation points, then its vertex connectivity is at least two.

Value

For articulation_points, a numeric vector giving the vertex IDs of the articulation points of the input graph. For bridges, a numeric vector giving the edge IDs of the bridges of the input graph.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

biconnected_components, components, is_connected, vertex_connectivity, edge_connectivity

Examples

g <- disjoint_union( make_full_graph(5), make_full_graph(5) )
clu <- components(g)$membership
g <- add_edges(g, c(match(1, clu), match(2, clu)) )
articulation_points(g)

g <- make_graph("krackhardt_kite")
bridges(g)

[Package igraph version 1.3.5 Index]
{% endraw %}