as_edgelist {igraph}	R Documentation

Convert a graph to an edge list

Description

Sometimes it is useful to work with a standard representation of a graph, like an edge list.

Usage

as_edgelist(graph, names = TRUE)

Arguments

graph
The graph to convert.

names
Whether to return a character matrix containing vertex names (ie. the name vertex attribute) if they exist or numeric vertex ids.

Details

as_edgelist returns the list of edges in a graph.

Value

A gsize(graph) by 2 numeric matrix.

See Also

graph_from_adjacency_matrix, read_graph

Examples

g <- sample_gnp(10, 2/10)
as_edgelist(g)

##       [,1] [,2]
##  [1,]    3    6
##  [2,]    5    6
##  [3,]    6    7
##  [4,]    2    8
##  [5,]    6    9
##  [6,]    8    9
##  [7,]    2   10
##  [8,]    4   10
##  [9,]    5   10

V(g)$name <- LETTERS[seq_len(gorder(g))]
as_edgelist(g)

##       [,1] [,2]
##  [1,] "C"  "F" 
##  [2,] "E"  "F" 
##  [3,] "F"  "G" 
##  [4,] "B"  "H" 
##  [5,] "F"  "I" 
##  [6,] "H"  "I" 
##  [7,] "B"  "J" 
##  [8,] "D"  "J" 
##  [9,] "E"  "J"

[Package igraph version 1.2.9 Index]