as_graphnel {igraph}	R Documentation

Convert igraph graphs to graphNEL objects from the graph package

Description

The graphNEL class is defined in the graph package, it is another way to represent graphs. These functions are provided to convert between the igraph and the graphNEL objects.

Usage

as_graphnel(graph)

Arguments

graph
An igraph graph object.

Details

as_graphnel converts an igraph graph to a graphNEL graph. It converts all graph/vertex/edge attributes. If the igraph graph has a vertex attribute ‘name’, then it will be used to assign vertex names in the graphNEL graph. Otherwise numeric igraph vertex ids will be used for this purpose.

Value

as_graphnel returns a graphNEL graph object.

See Also

graph_from_graphnel for the other direction, as_adj, graph_from_adjacency_matrix, as_adj_list and graph.adjlist for other graph representations.

Examples

## Undirected
## Not run: 
##D g <- make_ring(10)
##D V(g)$name <- letters[1:10]
##D GNEL <- as_graphnel(g)
##D g2 <- graph_from_graphnel(GNEL)
##D g2
##D 
##D ## Directed
##D g3 <- make_star(10, mode="in")
##D V(g3)$name <- letters[1:10]
##D GNEL2 <- as_graphnel(g3)
##D g4 <- graph_from_graphnel(GNEL2)
##D g4
## End(Not run)

[Package igraph version 1.2.9 Index]