as_membership {igraph}	R Documentation

Declare a numeric vector as a membership vector

Description

This is useful if you want to use functions defined on membership vectors, but your membership vector does not come from an igraph clustering method.

Usage

as_membership(x)

Arguments

x
The input vector.

Value

The input vector, with the membership class added.

Examples

## Compare to the correct clustering
g <- (make_full_graph(10) + make_full_graph(10)) %>%
  rewire(each_edge(p = 0.2))
correct <- rep(1:2, each = 10) %>% as_membership
fc <- cluster_fast_greedy(g)
compare(correct, fc)

## [1] 0

compare(correct, membership(fc))

## [1] 0

[Package igraph version 1.2.9 Index]