closeness {igraph}	R Documentation

Closeness centrality of vertices

Description

Cloness centrality measures how many steps is required to access every other vertex from a given vertex.

Usage

closeness(graph, v=V(graph), mode = c("all", "out", "in"))
closeness.estimate(graph, vids=V(graph), mode = c("out", "in", "all",
       "total"), cutoff) 

Arguments

graph 	 The graph to analyze.

v,vids 	 The vertices for which closeness will be calculated.

mode 	 Character string, defined the types of the paths used for measuring the distance in directed graphs. “in” measures the paths to a vertex, “out” measures paths from a vertex, all uses undirected paths. This argument is ignored for undirected graphs.

cutoff 	 The maximum path length to consider when calculating the betweenness. If zero or negative then there is no such limit.

Details

The closeness centrality of a vertex is defined by the inverse of the average length of the shortest paths to/from all the other vertices in the graph:

(|V|-1)/sum( d(v,i), i != v)

If there is no (directed) path between vertex v and i then the total number of vertices is used in the formula instead of the path length.

closeness.estimate only considers paths of length cutoff or smaller, this can be run for larger graphs, as the running time is not quadratic (if cutoff is small). If cutoff is zero or negative then the function calculates the exact closeness scores.

Value

Numeric vector with the closeness values of all the vertices in v.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

References

Freeman, L.C. (1979). Centrality in Social Networks I: Conceptual Clarification. Social Networks, 1, 215-239.

See Also

betweenness, degree

Examples

g <- graph.ring(10)
g2 <- graph.star(10)
closeness(g)
closeness(g2, mode="in")
closeness(g2, mode="out")
closeness(g2, mode="all")

[Package igraph version 0.5.2-2 Index]