cocitation {igraph}	R Documentation

Cocitation coupling

Description

Two vertices are cocited if there is another vertex citing both of them. cocitation siply counts how many types two vertices are cocited. The bibliographic coupling of two vertices is the number of other vertices they both cite, bibcoupling calculates this.

Usage

cocitation(graph, v=V(graph))
bibcoupling(graph, v=V(graph))

Arguments

graph 	 The graph object to analyze

v 	 Vertex sequence or numeric vector, the vertex ids for which the cocitation or bibliographic coupling values we want to calculate. The default is all vertices.

Details

cocitation calculates the cocitation counts for the vertices in the v argument and all vertices in the graph.

bibcoupling calculates the bibliographic coupling for vertices in v and all vertices in the graph.

Calculating the cocitation or bibliographic coupling for only one vertex costs the same amount of computation as for all vertices. This might change in the future.

Value

A numeric matrix with length(v) lines and vcount(graph) columns. Element (i,j) contains the cocitation or bibliographic coupling for vertices v[i] and j.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

Examples

g <- graph.ring(10)
cocitation(g)
bibcoupling(g)

[Package igraph version 0.5.2-2 Index]