components {igraph}	R Documentation

In- or out- component of a vertex

Description

Finds all vertices reachable from a given vertex, or the opposite: all vertices from which a given vertex is reachable via a directed path.

Usage

subcomponent(graph, v, mode = c("all", "out", "in"))

Arguments

graph 	 The graph to analyze.

v 	 The vertex to start the search from.

mode 	 Character string, either “in”, “out” or “all”. If “in” all vertices from which v is reachable are listed. If “out” all vertices reachable from v are returned. If “all” returns the union of these. It is ignored for undirected graphs.

Details

A breadh-first search is conducted starting from vertex v.

Value

Numeric vector, the ids of the vertices in the same component as v.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

clusters

Examples

g <- erdos.renyi.game(100, 1/200)
subcomponent(g, 1, "in")
subcomponent(g, 1, "out")
subcomponent(g, 1, "all")

[Package igraph version 0.5.2-2 Index]