--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.3 --- {% raw %}

console {igraph}	R Documentation

The igraph console

Description

The igraph console is a GUI windows that shows what the currently running igraph function is doing.

Usage

console()

Details

The console can be started by calling the console function. Then it stays open, until the user closes it.

Another way to start it to set the verbose igraph option to “tkconsole” via igraph_options. Then the console (re)opens each time an igraph function supporting it starts; to close it, set the verbose option to another value.

The console is written in Tcl/Tk and required the tcltk package.

Value

NULL, invisibly.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

igraph_options and the verbose option.

[Package igraph version 1.3.3 Index]
{% endraw %}