constraint {igraph}	R Documentation

Burt's constraint

Description

Given a graph, constraint calculates Burt's constraint for each vertex.

Usage

constraint(graph, nodes=V(graph), weights=NULL)

Arguments

graph 	 A graph object, the input graph.

nodes 	 The vertices for which the constraint will be calculated. Defaults to all vertices.

weights 	 The weights of the edges. If this is NULL and there is a weight edge attribute this is used. If there is no such edge attribute all edges will have the same weight.

Details

Burt's constraint is higher if ego has less, or mutually stronger related (i.e. more redundant) contacts. Burt's measure of constraint, C[i], of vertex i's ego network V[i], is defined for directed and valued graphs,

C[i] = sum( sum( (p[i,q] p[q,j])^2, q in V[i], q != i,j ), j in V[], j != i)

for a graph of order (ie. number of vertices) N, where proportional tie strengths are defined as

p[i,j]=(a[i,j]+a[j,i]) / sum(a[i,k]+a[k,i], k in V[i], k != i),

a[i,j] are elements of A and the latter being the graph adjacency matrix. For isolated vertices, constraint is undefined.

Value

A numeric vector of constraint scores

Author(s)

Jeroen Bruggeman j.p.bruggeman@uva.nl and Gabor Csardi csardi@rmki.kfki.hu

References

Burt, R.S. (2004). Structural holes and good ideas. American Journal of Sociology 110, 349-399.

Examples

g <- erdos.renyi.game(20, 5/20)
constraint(g)

[Package igraph version 0.5.2-2 Index]