diameter {igraph}	R Documentation

Diameter of a graph

Description

The diameter of a graph is the length of the longest geodesic.

Usage

diameter(graph, directed = TRUE, unconnected = TRUE)
get.diameter (graph, directed = TRUE, unconnected = TRUE) 
farthest.nodes (graph, directed = TRUE, unconnected = TRUE) 

Arguments

graph 	 The graph to analyize.

directed 	 Logical, whether directed or undirected paths are to be considered. This is ignored for undirected graphs.

unconnected 	 Logical, what to do if the graph is unconnected. If FALSE, the function will return the largest possible diameter which is the number of vertices. If TRUE, the diameters of the connected components will be calculated and the largest one will be returned.

Details

The diameter is calculated by using a breadth-first search like method.

get.diameter returns a path with the actual diameter. If there are many shortest paths of the length of the diameter, then it returns the first one found.

farthest.points returns two vertex ids, the vertices which are connected by the diameter path.

Value

A numeric constant for diameter, a numeric vector for get.diameter and a numeric vector of length two for farthest.nodes.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

shortest.paths

Examples

g <- graph.ring(10)
g2 <- delete.edges(g, c(0,1,0,9))
diameter(g2, unconnected=TRUE)
diameter(g2, unconnected=FALSE)

[Package igraph version 0.5.2-2 Index]