evcent {igraph}	R Documentation

Find Eigenvector Centrality Scores of Network Positions

Description

evcent takes a graph (graph) and returns the eigenvector centralities of positions v within it

Usage

evcent (graph, scale = TRUE, weights = NULL, options = igraph.arpack.default) 

Arguments

graph 	 Graph to be analyzed.

scale 	 Logical scalar, whether to scale the result to have a maximum score of one. If no scaling is used then the result vector has unit length in the Euclidean norm.

weights 	 A numerical vector or NULL. This argument can be used to give edge weights for calculating the weighted eigenvector centrality of vertices. If this is NULL and the graph has a weight edge attribute then that is used. If weights is a numerical vector then it used, even if the graph has a weights edge attribute. If this is NA, then no edge weights are used (even if the graph has a weight edge attribute.

options 	 A named list, to override some ARPACK options. See arpack for details.

Details

Eigenvector centrality scores correspond to the values of the first eigenvector of the graph adjacency matrix; these scores may, in turn, be interpreted as arising from a reciprocal process in which the centrality of each actor is proportional to the sum of the centralities of those actors to whom he or she is connected. In general, vertices with high eigenvector centralities are those which are connected to many other vertices which are, in turn, connected to many others (and so on). (The perceptive may realize that this implies that the largest values will be obtained by individuals in large cliques (or high-density substructures). This is also intelligible from an algebraic point of view, with the first eigenvector being closely related to the best rank-1 approximation of the adjacency matrix (a relationship which is easy to see in the special case of a diagonalizable symmetric real matrix via the $S Lambda S^{-1}$ decomposition).)

From igraph version 0.5 this function uses ARPACK for the underlying computation, see arpack for more about ARPACK in igraph.

Value

A named list with components:

vector 	 A vector containing the centrality scores.

value 	 The eigenvalue corresponding to the calculated eigenvector, i.e. the centrality scores.

options 	 A named list, information about the underlying ARPACK computation. See arpack for the details.

WARNING

evcent will not symmetrize your data before extracting eigenvectors; don't send this routine asymmetric matrices unless you really mean to do so.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu and Carter T. Butts buttsc@uci.edu for the manual page

References

Bonacich, P. (1987). Power and Centrality: A Family of Measures. American Journal of Sociology, 92, 1170-1182.

Katz, L. (1953). A New Status Index Derived from Sociometric Analysis. Psychometrika, 18, 39-43.

Examples

#Generate some test data
g <- graph.ring(10, directed=FALSE)
#Compute eigenvector centrality scores
evcent(g)

[Package igraph version 0.5.2-2 Index]