graph-motifs {igraph}	R Documentation

Graph motifs

Description

Graph motifs are small subgraphs with a well-defined strucure. These functions search a graph for various motifs.

Usage

graph.motifs(graph, size = 3, cut.prob = rep(0, size)) 
graph.motifs.no(graph, size = 3, cut.prob = rep(0, size)) 
graph.motifs.est(graph, size = 3, cut.prob = rep(0, size), sample.size =
    vcount(graph)/10, sample = NULL) 

Arguments

graph 	 Graph object, the input graph.

size 	 The size of the motif, currently 3 and 4 are supported only.

cut.prob 	 Numeric vector giving the probabilities that the search graph is cut at a certain level. Its length should be the same as the size of the motif (the size argument). By default no cuts are made.

sample.size 	 The number of vertices to use as a starting point for finding motifs. Only used if the sample argument is NULL.

sample 	 If not NULL then it specifies the vertices to use as a starting point for finding motifs.

Details

graph.motifs searches a graph for motifs of a given size and returns a numeric vector containing the number of different motifs. The order of the motifs is defined by their isomorphism class, see graph.isoclass.

graph.motifs.no calculates the total number of motifs of a given size in graph.

graph.motifs.est estimates the total number of motifs of a given size in a graph based on a sample.

Value

graph.motifs returns a numeric vector.
graph.motifs.no and graph.motifs.est return a numeric constant.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

graph.isoclass

Examples

g <- barabasi.game(100)
graph.motifs(g, 3)
graph.motifs.no(g, 3)
graph.motifs.est(g, 3)

[Package igraph version 0.5.2-2 Index]