--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.2 --- {% raw %}

graph_from_atlas {igraph}	R Documentation

Create a graph from the Graph Atlas

Description

graph_from_atlas creates graphs from the book ‘An Atlas of Graphs’ by Roland C. Read and Robin J. Wilson. The atlas contains all undirected graphs with up to seven vertices, numbered from 0 up to 1252. The graphs are listed:

in increasing order of number of nodes;

for a fixed number of nodes, in increasing order of the number of edges;

for fixed numbers of nodes and edges, in increasing order of the degree sequence, for example 111223 < 112222;

for fixed degree sequence, in increasing number of automorphisms.

Usage

graph_from_atlas(n)

atlas(...)

Arguments

n
The id of the graph to create.

...
Passed to graph_from_atlas.

Value

An igraph graph.

See Also

Other deterministic constructors: graph_from_edgelist(), graph_from_literal(), make_chordal_ring(), make_empty_graph(), make_full_citation_graph(), make_full_graph(), make_graph(), make_lattice(), make_ring(), make_star(), make_tree()

Examples

## Some randomly picked graphs from the atlas
graph_from_atlas(sample(0:1252, 1))
graph_from_atlas(sample(0:1252, 1))

[Package igraph version 1.3.2 Index]
{% endraw %}