Create a graph from the Graph Atlas

graph_from_atlas() creates graphs from the book ‘An Atlas of Graphs’ by Roland C. Read and Robin J. Wilson. The atlas contains all undirected graphs with up to seven vertices, numbered from 0 up to 1252. The graphs are listed:

in increasing order of number of nodes;

for a fixed number of nodes, in increasing order of the number of edges;

for fixed numbers of nodes and edges, in increasing order of the degree sequence, for example 111223 < 112222;

for fixed degree sequence, in increasing number of automorphisms.

Usage

graph_from_atlas(n)

atlas(...)

Arguments

n

The id of the graph to create.

...

Passed to graph_from_atlas().

Value

An igraph graph.

See also

Examples

## Some randomly picked graphs from the atlas
graph_from_atlas(sample(0:1252, 1))
#> IGRAPH a539ba0 U--- 7 11 -- Graph from the Atlas #766
#> + attr: name (g/c), n (g/n)
#> + edges from a539ba0:
#>  [1] 4--5 1--5 1--2 2--3 3--4 1--6 4--6 1--4 2--6 3--6 6--7
graph_from_atlas(sample(0:1252, 1))
#> IGRAPH b6f33a6 U--- 7 12 -- Graph from the Atlas #893
#> + attr: name (g/c), n (g/n)
#> + edges from b6f33a6:
#>  [1] 1--2 2--3 3--4 5--6 1--5 2--4 2--5 3--5 1--4 4--6 1--3 2--7