--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.4 --- {% raw %}

graph_from_edgelist {igraph}	R Documentation

Create a graph from an edge list matrix

Description

graph_from_edgelist creates a graph from an edge list. Its argument is a two-column matrix, each row defines one edge. If it is a numeric matrix then its elements are interpreted as vertex ids. If it is a character matrix then it is interpreted as symbolic vertex names and a vertex id will be assigned to each name, and also a name vertex attribute will be added.

Usage

graph_from_edgelist(el, directed = TRUE)

from_edgelist(...)

Arguments

el
The edge list, a two column matrix, character or numeric.

directed
Whether to create a directed graph.

...
Passed to graph_from_edgelist.

Value

An igraph graph.

See Also

Other deterministic constructors: graph_from_atlas(), graph_from_literal(), make_chordal_ring(), make_empty_graph(), make_full_citation_graph(), make_full_graph(), make_graph(), make_lattice(), make_ring(), make_star(), make_tree()

Examples

el <- matrix( c("foo", "bar", "bar", "foobar"), nc = 2, byrow = TRUE)
graph_from_edgelist(el)

# Create a ring by hand
graph_from_edgelist(cbind(1:10, c(2:10, 1)))

[Package igraph version 1.3.4 Index]
{% endraw %}