--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.2 --- {% raw %}

graph_from_lcf {igraph}	R Documentation

Creating a graph from LCF notation

Description

LCF is short for Lederberg-Coxeter-Frucht, it is a concise notation for 3-regular Hamiltonian graphs. It constists of three parameters, the number of vertices in the graph, a list of shifts giving additional edges to a cycle backbone and another integer giving how many times the shifts should be performed. See http://mathworld.wolfram.com/LCFNotation.html for details.

Usage

graph_from_lcf(n, shifts, repeats = 1)

Arguments

n
Integer, the number of vertices in the graph.

shifts
Integer vector, the shifts.

repeats
Integer constant, how many times to repeat the shifts.

Value

A graph object.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

graph can create arbitrary graphs, see also the other functions on the its manual page for creating special graphs.

Examples

# This is the Franklin graph:
g1 <- graph_from_lcf(12, c(5,-5), 6)
g2 <- make_graph("Franklin")
isomorphic(g1, g2)

[Package igraph version 1.3.2 Index]
{% endraw %}