--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.3 --- {% raw %}

hrg_tree {igraph}	R Documentation

Create an igraph graph from a hierarchical random graph model

Description

hrg_tree creates the corresponsing igraph tree of a hierarchical random graph model.

Usage

hrg_tree(hrg)

Arguments

hrg
A hierarchical random graph model.

Value

An igraph graph.

See Also

Other hierarchical random graph functions: consensus_tree(), fit_hrg(), hrg-methods, hrg(), predict_edges(), print.igraphHRGConsensus(), print.igraphHRG(), sample_hrg()

[Package igraph version 1.3.3 Index]
{% endraw %}