Kleinberg's authority centrality scores.

Kleinberg's authority centrality scores.

Kleinberg's hub centrality scores.

Usage

authority_score(
  graph,
  scale = TRUE,
  weights = NULL,
  options = arpack_defaults()
)

hub_score(graph, scale = TRUE, weights = NULL, options = arpack_defaults())

Arguments

graph

The input graph.

scale

Logical scalar, whether to scale the result to have a maximum score of one. If no scaling is used then the result vector has unit length in the Euclidean norm.

weights

Optional positive weight vector for calculating weighted scores. If the graph has a weight edge attribute, then this is used by default. This function interprets edge weights as connection strengths. In the random surfer model, an edge with a larger weight is more likely to be selected by the surfer.

options

A named list, to override some ARPACK options. See arpack() for details.

See also