igraph-parameters {igraph}	R Documentation

Parameters for the igraph package

Description

igraph has some parameters which (usually) affect the behavior of many functions. These can be set for the whole session via igraph.par.

Usage

igraph.par(parid, parvalue = NULL)

Arguments

parid 	 The name of the parameter. See the currently used parameters below.

parvalue 	 The new value of the parameter. If NULL then the current value of the parameter is listed.

Details

The currently used parameters in alphabetical order:

print.edge.attributes

Logical constant, whether to print edge attributes when printing graphs. Defaults to FALSE.

print.graph.attributes

Logical constant, whether to print graph attributes when printing graphs. Defaults to FALSE.

print.vertex.attriubutes

Logical constant, whether to print vertex attributes when printing graphs. Defaults to FALSE.

verbose

Logical constant, whether igraph functions should talk more than minimal. Eg. if TRUE thne some functions will use progress bars while computing. Defaults to FALSE.

Value

If parvalue is NULL then the current value of the parameter is returned. Otherwise the new value of the parameter is returned invisibly.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

Examples

igraph.par("verbose", FALSE)
layout.kamada.kawai( graph.ring(10) )
igraph.par("verbose", TRUE)

[Package igraph version 0.5.2-2 Index]