--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.4 --- {% raw %}

igraph_demo {igraph}	R Documentation

Run igraph demos, step by step

Description

Run one of the accompanying igraph demos, somewhat interactively, using a Tk window.

Usage

igraph_demo(which)

Arguments

which
If not given, then the names of the available demos are listed. Otherwise it should be either a filename or the name of an igraph demo.

Details

This function provides a somewhat nicer interface to igraph demos that come with the package, than the standard demo function. igraph demos are divided into chunks and igraph_demo runs them chunk by chunk, with the possibility of inspecting the workspace between two chunks.

The tcltk package is needed for igraph_demo.

Value

Returns NULL, invisibly.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

demo

Examples

igraph_demo()
if (interactive()) {
  igraph_demo("centrality")
}

[Package igraph version 1.3.4 Index]
{% endraw %}