Incident edges of a vertex in a graph

Incident edges of a vertex in a graph

Usage

incident(graph, v, mode = c("all", "out", "in", "total"))

Arguments

graph

The input graph.

v

The vertex of which the incident edges are queried.

mode

Whether to query outgoing (‘out’), incoming (‘in’) edges, or both types (‘all’). This is ignored for undirected graphs.

Value

An edge sequence containing the incident edges of the input vertex.

See also

Examples

g <- make_graph("Zachary")
incident(g, 1)
#> + 16/78 edges from 23b3cbb:
#>  [1] 1-- 2 1-- 3 1-- 4 1-- 5 1-- 6 1-- 7 1-- 8 1-- 9 1--11 1--12 1--13 1--14
#> [13] 1--18 1--20 1--22 1--32
incident(g, 34)
#> + 17/78 edges from 23b3cbb:
#>  [1]  9--34 10--34 14--34 15--34 16--34 19--34 20--34 21--34 23--34 24--34
#> [11] 27--34 28--34 29--34 30--34 31--34 32--34 33--34