Intersection of two or more sets

This is an S3 generic function. See methods("intersection") for the actual implementations for various S3 classes. Initially it is implemented for igraph graphs and igraph vertex and edge sequences. See intersection.igraph(), and intersection.igraph.vs().

Usage

intersection(...)

Arguments

...

Arguments, their number and interpretation depends on the function that implements intersection().

Value

Depends on the function that implements this method.

See also