--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.2 --- {% raw %}

is_dag {igraph}	R Documentation

Directed acyclic graphs

Description

This function tests whether the given graph is a DAG, a directed acyclic graph.

Usage

is_dag(graph)

Arguments

graph
The input graph. It may be undirected, in which case FALSE is reported.

Details

is_dag checks whether there is a directed cycle in the graph. If not, the graph is a DAG.

Value

A logical vector of length one.

Author(s)

Tamas Nepusz ntamas@gmail.com for the C code, Gabor Csardi csardi.gabor@gmail.com for the R interface.

Examples

g <- make_tree(10)
is_dag(g)
g2 <- g + edge(5,1)
is_dag(g2)

[Package igraph version 1.3.2 Index]
{% endraw %}