--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.2 --- {% raw %}

is_separator {igraph}	R Documentation

Vertex separators

Description

Check whether a given set of vertices is a vertex separator.

Usage

is_separator(graph, candidate)

Arguments

graph
The input graph. It may be directed, but edge directions are ignored.

candidate
A numeric vector giving the vertex ids of the candidate separator.

Details

is_separator decides whether the supplied vertex set is a vertex separator. A vertex set is a vertex separator if its removal results a disconnected graph.

In the special case of a fully connected graph with n vertices, each set of n-1 vertices is considered to be a vertex separator.

Value

A logical scalar, whether the supplied vertex set is a (minimal) vertex separator or not.

See Also

is_min_separator, min_separators lists all vertex separator of minimum size.

[Package igraph version 1.3.2 Index]
{% endraw %}