--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.1 --- {% raw %}

isomorphism_class {igraph}	R Documentation

Isomorphism class of a graph

Description

The isomorphism class is a non-negative integer number. Graphs (with the same number of vertices) having the same isomorphism class are isomorphic and isomorphic graphs always have the same isomorphism class. Currently it can handle directed graphs with 3 or 4 vertices and undirected graphs with 3 to 6 vertices.

Usage

isomorphism_class(graph, v)

Arguments

graph
The input graph.

v
Optionally a vertex sequence. If not missing, then an induced subgraph of the input graph, consisting of this vertices, is used.

Value

An integer number.

See Also

Other graph isomorphism: count_isomorphisms(), count_subgraph_isomorphisms(), graph_from_isomorphism_class(), isomorphic(), isomorphisms(), subgraph_isomorphic(), subgraph_isomorphisms()

Examples

# create some non-isomorphic graphs
g1 <- graph_from_isomorphism_class(3, 10)
g2 <- graph_from_isomorphism_class(3, 11)
isomorphism_class(g1)
isomorphism_class(g2)
isomorphic(g1, g2)

[Package igraph version 1.3.1 Index]
{% endraw %}