isomorphisms {igraph}	R Documentation

Calculate all isomorphic mappings between the vertices of two graphs

Description

Calculate all isomorphic mappings between the vertices of two graphs

Usage

isomorphisms(graph1, graph2, method = "vf2", ...)

Arguments

graph1
The first graph.

graph2
The second graph.

method
Currently only ‘vf2’ is supported, see isomorphic for details about it and extra arguments.

...
Extra arguments, passed to the various methods.

Value

A list of vertex sequences, corresponding to all mappings from the first graph to the second.

See Also

Other graph isomorphism: count_isomorphisms, graph.count.isomorphisms.vf2; count_subgraph_isomorphisms, graph.count.subisomorphisms.vf2; graph.get.subisomorphisms.vf2, subgraph_isomorphisms; graph.isoclass, graph.isoclass.subgraph, isomorphism_class; graph.isocreate, graph_from_isomorphism_class; graph.isomorphic, graph.isomorphic.34, graph.isomorphic.bliss, graph.isomorphic.vf2, is_isomorphic_to, isomorphic; graph.subisomorphic.lad, graph.subisomorphic.vf2, is_subgraph_isomorphic_to, subgraph_isomorphic

[Package igraph version 1.0.1 Index]