Average nearest neighbor degree

Calculate the average nearest neighbor degree of the given vertices and the same quantity in the function of vertex degree

Usage

knn(
  graph,
  vids = V(graph),
  mode = c("all", "out", "in", "total"),
  neighbor.degree.mode = c("all", "out", "in", "total"),
  weights = NULL
)

Arguments

graph

The input graph. It may be directed.

vids

The vertices for which the calculation is performed. Normally it includes all vertices. Note, that if not all vertices are given here, then both ‘knn’ and ‘knnk’ will be calculated based on the given vertices only.

mode

Character constant to indicate the type of neighbors to consider in directed graphs. out considers out-neighbors, in considers in-neighbors and all ignores edge directions.

neighbor.degree.mode

The type of degree to average in directed graphs. out averages out-degrees, in averages in-degrees and all ignores edge directions for the degree calculation.

weights

Weight vector. If the graph has a weight edge attribute, then this is used by default. If this argument is given, then vertex strength (see strength()) is used instead of vertex degree. But note that knnk is still given in the function of the normal vertex degree. Weights are are used to calculate a weighted degree (also called strength()) instead of the degree.

Value

A list with two members:

knn

A numeric vector giving the average nearest neighbor degree for all vertices in vids.

knnk

A numeric vector, its length is the maximum (total) vertex degree in the graph. The first element is the average nearest neighbor degree of vertices with degree one, etc.

Details

Note that for zero degree vertices the answer in ‘knn’ is NaN (zero divided by zero), the same is true for ‘knnk’ if a given degree never appears in the network.

The weighted version computes a weighted average of the neighbor degrees as

k_nn_u = 1/s_u sum_v w_uv k_v,

where s_u = sum_v w_uv is the sum of the incident edge weights of vertex u, i.e. its strength. The sum runs over the neighbors v of vertex u as indicated by mode. w_uv denotes the weighted adjacency matrix and k_v is the neighbors' degree, specified by neighbor_degree_mode.

References

Alain Barrat, Marc Barthelemy, Romualdo Pastor-Satorras, Alessandro Vespignani: The architecture of complex weighted networks, Proc. Natl. Acad. Sci. USA 101, 3747 (2004)

See also

Author

Gabor Csardi csardi.gabor@gmail.com

Examples

# Some trivial ones
g <- make_ring(10)
knn(g)
#> $knn
#>  [1] 2 2 2 2 2 2 2 2 2 2
#> 
#> $knnk
#> [1] NaN   2
#> 
g2 <- make_star(10)
knn(g2)
#> $knn
#>  [1] 1 9 9 9 9 9 9 9 9 9
#> 
#> $knnk
#> [1]   9 NaN NaN NaN NaN NaN NaN NaN   1
#> 

# A scale-free one, try to plot 'knnk'
g3 <- sample_pa(1000, m = 5)
knn(g3)
#> $knn
#>    [1]  14.560241  14.943038  13.635359  14.597183  16.182759  71.538462
#>    [7]  18.277487  16.269841  85.095238  66.800000  22.737864  19.278912
#>   [13]  32.855072  27.060976  50.675676  48.923077  30.457627  27.850000
#>   [19]  22.421053  96.733333  30.268657 115.583333  34.142857  19.081395
#>   [25]  48.473684  61.923077  54.028571  47.676471  53.241379  86.687500
#>   [31]  59.375000  33.931818  64.625000  18.161765 117.545455  56.407407
#>   [37]  28.416667  40.576923  46.523810  35.156250  33.464286  84.300000
#>   [43]  43.580645  65.411765  45.285714 203.333333  34.970588  86.611111
#>   [49]  65.470588  59.000000 214.000000  38.000000 171.285714  70.476190
#>   [55]  54.906250  28.959184  83.111111 142.545455  42.382353 157.375000
#>   [61]  72.142857  59.040000  45.400000  91.666667 118.888889 275.200000
#>   [67]  50.300000  66.562500 143.200000  30.833333 135.333333  82.250000
#>   [73]  68.333333  28.242424  48.550000  60.894737 118.333333  62.000000
#>   [79] 224.166667  28.152174  58.210526 243.400000  39.121951 142.714286
#>   [85]  43.222222 199.200000  84.307692 138.333333  79.090909   9.758621
#>   [91] 174.000000 100.636364  78.666667 101.250000  88.000000  39.571429
#>   [97]  36.052632  54.588235  79.583333 154.333333 170.500000  85.600000
#>  [103] 109.714286 126.777778  49.625000  75.181818 274.400000  72.363636
#>  [109]  79.142857  46.500000 158.166667 206.500000  58.958333 120.600000
#>  [115] 252.200000 124.000000 117.750000 211.333333  86.600000  27.562500
#>  [121]  92.428571  43.388889  94.166667 113.428571 119.714286 168.800000
#>  [127]  39.142857  98.000000  66.285714 192.200000 135.222222  44.166667
#>  [133]  99.250000 100.833333  19.294118 120.571429 124.571429  82.636364
#>  [139] 146.500000 152.000000 122.666667  31.611111  38.444444 155.000000
#>  [145]  65.200000 102.000000 112.428571 168.000000 196.285714 117.250000
#>  [151] 174.200000  89.857143  53.500000 176.666667 161.600000 157.666667
#>  [157]  86.454545 237.400000 193.833333 168.666667  62.714286 107.636364
#>  [163]  48.625000 180.400000 170.000000 182.285714  56.250000 165.000000
#>  [169]  89.615385  69.000000 117.111111 103.800000 138.375000 112.375000
#>  [175]  85.250000 222.000000 174.200000  91.166667 147.800000 105.571429
#>  [181]  81.400000  83.615385 110.833333 127.400000  91.777778 115.625000
#>  [187] 156.444444 135.500000 117.666667  21.200000  49.214286  80.100000
#>  [193] 124.400000  63.076923 183.666667 138.222222  63.076923 101.222222
#>  [199]  48.800000  91.625000  44.166667 123.222222  26.333333 152.833333
#>  [205]  86.200000 114.000000 158.200000  67.500000  99.142857  46.285714
#>  [211]  97.600000  71.500000  99.750000  70.200000 135.666667 184.571429
#>  [217] 203.000000  49.000000 217.600000 153.000000 188.400000 176.666667
#>  [223]  29.500000 143.875000 162.000000 229.000000  77.714286  59.166667
#>  [229] 115.125000 136.500000 117.428571  97.200000 177.600000 230.000000
#>  [235] 196.000000 117.400000  51.000000  99.454545  39.583333  77.538462
#>  [241]  95.400000 100.555556 191.400000 223.400000  93.000000 118.600000
#>  [247] 116.125000 178.000000 167.625000 128.600000  61.400000  95.428571
#>  [253] 198.000000  87.000000  97.400000  25.461538 225.600000  95.400000
#>  [259] 172.200000 158.714286  77.000000 133.500000 143.400000  73.000000
#>  [265] 186.833333 144.666667 122.333333 117.800000 112.600000  93.545455
#>  [271]  83.888889  91.615385 121.600000 108.200000 119.666667  96.200000
#>  [277] 194.000000 167.333333 165.600000 147.200000  99.400000 162.285714
#>  [283]  62.230769 134.400000 119.200000  30.666667  59.714286 125.200000
#>  [289] 139.000000  53.625000  20.714286  88.666667 146.800000  68.125000
#>  [295] 190.500000 109.428571  70.000000 195.500000  86.800000  48.800000
#>  [301]  24.166667  77.818182  77.375000  83.125000 118.400000  18.125000
#>  [307]  52.000000  53.222222 114.142857 159.800000  68.500000 179.600000
#>  [313]  90.800000 174.000000  95.600000  99.375000 171.400000  79.800000
#>  [319]  74.833333  77.666667  69.000000 113.333333 108.800000 298.200000
#>  [325] 155.500000 233.200000 125.666667 107.285714  79.428571  34.888889
#>  [331]  53.700000 165.400000  88.000000 105.285714  57.222222  87.600000
#>  [337] 160.800000 152.000000 121.833333 105.571429 155.857143 122.000000
#>  [343] 179.142857 217.200000  75.333333  65.500000 129.000000 184.333333
#>  [349]  73.900000  70.111111 122.666667  18.000000  69.181818  84.200000
#>  [355]  41.833333  86.250000  59.500000  72.166667 122.800000 104.333333
#>  [361] 116.600000 123.142857  89.400000  95.400000 119.000000  69.250000
#>  [367] 107.000000  98.000000 109.000000 109.444444  96.666667 214.400000
#>  [373] 125.666667 171.200000  44.000000  26.800000  80.400000 116.666667
#>  [379] 100.800000  74.375000 182.000000 231.800000 134.800000 163.333333
#>  [385] 161.833333 205.000000  87.833333 150.000000  64.142857 131.000000
#>  [391] 139.000000 146.000000 196.833333 117.000000 151.200000  73.200000
#>  [397] 174.857143 114.000000  30.714286  42.833333  82.800000 135.200000
#>  [403]  36.000000  81.250000  54.700000 222.000000 102.142857  90.833333
#>  [409]  79.375000 123.400000  92.000000  77.571429 197.333333  73.428571
#>  [415] 106.000000 108.285714 116.800000  69.333333  32.200000  67.428571
#>  [421] 152.800000 156.800000 114.166667 113.714286 111.142857 205.000000
#>  [427] 140.000000 183.000000 139.333333 129.500000 272.800000 132.500000
#>  [433] 156.800000 121.857143 135.600000 215.400000  76.833333 131.400000
#>  [439] 109.428571 117.000000 128.833333 138.000000 156.000000  93.625000
#>  [445]  84.625000  63.000000  77.600000 184.666667 157.000000  47.800000
#>  [451] 126.500000  81.166667 144.833333 191.400000  29.833333  73.750000
#>  [457] 105.111111 126.000000  79.600000 109.833333  85.666667 166.000000
#>  [463] 180.200000  53.200000  30.400000 244.000000  69.333333 173.142857
#>  [469] 163.857143  47.333333  88.000000  89.000000 214.400000  92.200000
#>  [475]  88.428571 171.600000 179.400000  75.714286 137.800000 151.600000
#>  [481] 184.333333 174.800000 109.400000  87.600000  93.600000 187.000000
#>  [487] 122.500000 122.833333 123.142857  20.000000 130.666667  87.000000
#>  [493] 152.000000 100.800000 204.400000  59.142857 103.833333 139.750000
#>  [499]  76.666667  78.800000  54.000000 171.000000 121.714286 171.200000
#>  [505]  70.666667 262.000000  58.000000 118.400000 161.400000 159.800000
#>  [511] 101.200000 178.600000 142.800000 146.000000 188.200000 103.857143
#>  [517] 109.200000  91.666667 131.000000  60.500000  67.000000  86.250000
#>  [523] 130.833333  39.400000 167.666667  31.250000 116.333333 165.400000
#>  [529]  67.400000 105.200000 190.000000 156.800000 210.400000 110.428571
#>  [535]  44.000000 101.600000 269.000000  86.571429 208.400000  83.333333
#>  [541] 222.400000 123.166667 159.000000 143.600000 177.200000  80.666667
#>  [547] 110.200000 207.200000 193.800000 187.800000 130.600000 133.166667
#>  [553]  21.500000 101.000000 147.500000  44.166667  77.166667 102.400000
#>  [559] 153.400000 140.000000 113.000000  81.200000 129.200000 166.857143
#>  [565] 101.600000  50.600000  20.333333 154.500000  90.800000 135.800000
#>  [571] 197.800000  53.750000  69.666667 172.600000  94.600000 108.625000
#>  [577] 228.400000  54.500000 145.800000 143.000000  73.200000  97.500000
#>  [583]  40.000000  64.400000 164.600000  83.000000 121.833333 284.200000
#>  [589] 128.600000 110.571429 168.333333 103.200000 110.142857  97.142857
#>  [595]  83.500000 138.000000 233.600000  10.400000  75.800000 278.400000
#>  [601] 207.166667 101.200000 109.500000  49.300000 175.600000 115.600000
#>  [607] 172.500000  27.600000 146.800000 212.600000  92.571429  73.800000
#>  [613] 156.000000 149.800000 183.000000 218.800000 107.800000 142.800000
#>  [619] 158.600000 282.600000 123.600000 178.166667 100.666667  89.142857
#>  [625] 163.000000 154.800000 121.800000 102.800000 152.000000 168.285714
#>  [631] 176.000000 268.600000  88.666667 114.400000 127.333333 102.400000
#>  [637] 184.800000  89.800000  96.400000 144.400000 110.600000  72.428571
#>  [643] 102.166667 138.000000 144.600000 157.200000  65.833333 135.000000
#>  [649]  62.857143 162.000000  36.600000 115.200000 128.600000 149.000000
#>  [655] 116.571429 202.000000 228.600000 254.800000 158.000000  99.600000
#>  [661] 107.400000 246.600000  47.200000 200.833333  75.800000  57.400000
#>  [667] 148.000000 232.400000 124.400000 165.400000 242.000000 127.000000
#>  [673]  55.800000 206.000000 230.400000 160.600000  77.333333  46.142857
#>  [679]  74.200000 229.000000 113.333333 122.400000 188.666667 220.800000
#>  [685] 181.400000 146.200000  56.800000 148.400000 198.400000  64.857143
#>  [691] 147.600000  60.285714 112.800000  83.000000 281.600000 160.000000
#>  [697] 100.142857  82.800000 170.000000  88.333333 145.000000 178.200000
#>  [703]  68.600000 170.600000 135.400000 190.400000  38.500000 201.200000
#>  [709] 171.200000  48.400000 138.400000  99.500000 189.200000  96.000000
#>  [715] 176.400000 123.166667 130.333333 157.200000 170.333333  13.800000
#>  [721] 249.200000  82.000000 138.800000  59.500000  49.800000  68.857143
#>  [727] 120.600000  85.200000 175.600000 136.800000  19.285714 174.600000
#>  [733]  28.333333 154.000000  87.400000 150.400000 179.333333 100.600000
#>  [739] 189.142857  87.600000 173.166667 263.800000  74.400000 198.200000
#>  [745]  22.000000  55.600000 181.166667 150.800000  74.000000 131.333333
#>  [751]  86.285714 185.400000 137.333333  77.600000  40.333333  50.400000
#>  [757]  33.000000 116.400000  90.000000  71.666667 100.800000 209.600000
#>  [763] 253.800000 102.000000  82.833333 142.600000  17.142857  82.400000
#>  [769] 136.000000 113.800000  83.800000 156.600000  89.400000 201.000000
#>  [775] 270.400000 150.000000 161.800000 157.600000 170.600000   8.000000
#>  [781] 127.000000 132.800000  37.000000 142.800000 150.200000  83.600000
#>  [787] 134.600000  26.000000 162.200000 116.666667 148.400000  11.000000
#>  [793] 163.400000 200.600000  69.400000 137.200000  27.833333 189.400000
#>  [799]  39.400000 123.600000 140.000000 117.800000  42.000000 137.200000
#>  [805] 128.400000 153.400000 208.600000  28.600000  24.600000 122.166667
#>  [811] 171.400000 102.333333 127.000000 140.600000  98.000000  86.800000
#>  [817]  91.600000  91.600000 125.500000 274.200000  68.800000 164.600000
#>  [823]  28.000000  51.400000 123.600000 209.400000  40.400000 117.500000
#>  [829]  67.166667  31.200000 130.166667 137.000000 105.600000 231.200000
#>  [835]  32.800000 129.200000 181.200000  10.800000 145.000000 191.000000
#>  [841] 268.600000  74.400000 169.000000 181.600000  40.200000  97.200000
#>  [847]  17.800000 179.200000  74.200000  84.000000 175.800000 130.000000
#>  [853] 102.800000 207.600000  15.600000  63.800000  90.200000  83.571429
#>  [859]  49.600000  17.200000  84.000000  76.000000 153.000000 103.800000
#>  [865]  60.600000 166.400000 220.200000 275.600000  31.000000  78.200000
#>  [871] 114.600000 200.200000 104.000000 209.600000 214.800000 221.000000
#>  [877] 131.200000 149.600000 168.000000  89.000000  72.000000  48.400000
#>  [883] 198.800000 184.000000 163.000000 143.400000 144.400000 153.000000
#>  [889] 204.000000  88.000000  83.600000  29.000000 117.400000 127.800000
#>  [895] 111.400000 126.800000  43.400000 162.200000  30.200000  18.400000
#>  [901] 169.000000  34.800000 127.000000 175.600000  70.600000 139.600000
#>  [907]  76.600000 161.000000  90.800000 152.600000 124.800000 146.400000
#>  [913] 165.833333  66.800000 167.800000 156.800000 170.200000 131.600000
#>  [919] 108.400000 146.400000 199.200000  97.400000 147.400000  78.400000
#>  [925] 249.000000  87.000000 115.800000  18.000000 120.200000  80.200000
#>  [931]  77.800000  56.400000  70.333333  16.000000  83.400000  58.600000
#>  [937] 209.200000 139.200000  84.600000 106.800000 139.600000  94.800000
#>  [943] 143.400000 137.200000  24.600000  78.200000 146.800000 102.200000
#>  [949]  77.600000 171.600000  39.600000 116.200000 128.600000 185.800000
#>  [955]  77.600000  77.600000 108.200000  85.200000  96.800000  34.000000
#>  [961]  80.000000 133.600000  66.800000 133.000000  93.800000 188.400000
#>  [967] 197.600000  70.600000  90.400000 133.200000  88.200000  45.200000
#>  [973] 113.600000 146.800000  92.400000 180.000000 140.800000 196.800000
#>  [979] 215.200000 105.166667  93.200000 198.400000  85.600000 146.600000
#>  [985]  82.600000 219.000000 195.800000  24.800000  59.600000 172.800000
#>  [991] 130.400000 179.200000 185.000000 178.000000 141.200000 160.600000
#>  [997] 152.800000 135.400000  82.400000  91.000000
#> 
#> $knnk
#>   [1]       NaN       NaN       NaN       NaN 134.57579 118.48750 103.71586
#>   [8]  95.50581  97.70370  83.88182  89.86364  70.55303  70.95385  48.08333
#>  [15]  76.11111  62.00893  51.19118  49.46296  50.90789  49.42500  67.36508
#>  [22]       NaN       NaN  50.15625  62.92000  58.01282  56.40741       NaN
#>  [29]  31.50000       NaN  43.58065  45.03125  33.12121  41.67647  44.08571
#>  [36]  30.83333  50.67568       NaN  48.92308       NaN  39.12195       NaN
#>  [43]       NaN  33.93182       NaN  28.15217       NaN       NaN  28.95918
#>  [50]       NaN       NaN       NaN       NaN       NaN       NaN  33.46429
#>  [57]       NaN       NaN  30.45763  27.85000       NaN       NaN       NaN
#>  [64]       NaN       NaN       NaN  30.26866  18.16176  32.85507       NaN
#>  [71]       NaN       NaN       NaN       NaN       NaN  22.42105       NaN
#>  [78]       NaN       NaN       NaN       NaN  27.06098       NaN       NaN
#>  [85]       NaN  19.08140       NaN       NaN       NaN       NaN       NaN
#>  [92]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#>  [99]       NaN       NaN       NaN       NaN  22.73786       NaN       NaN
#> [106]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [113]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [120]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [127]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [134]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [141]       NaN       NaN       NaN       NaN       NaN       NaN  19.27891
#> [148]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [155]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [162]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [169]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [176]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [183]       NaN       NaN       NaN       NaN       NaN       NaN  16.26984
#> [190]       NaN  18.27749       NaN       NaN       NaN       NaN       NaN
#> [197]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [204]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [211]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [218]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [225]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [232]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [239]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [246]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [253]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [260]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [267]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [274]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [281]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [288]       NaN       NaN  16.18276       NaN       NaN       NaN       NaN
#> [295]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [302]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [309]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [316]  14.94304       NaN       NaN       NaN       NaN       NaN       NaN
#> [323]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [330]       NaN       NaN  14.56024       NaN       NaN       NaN       NaN
#> [337]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [344]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [351]       NaN       NaN       NaN       NaN  14.59718       NaN       NaN
#> [358]       NaN       NaN       NaN       NaN  13.63536
#> 

# A random graph
g4 <- sample_gnp(1000, p = 5 / 1000)
knn(g4)
#> $knn
#>    [1]  5.571429  5.285714  5.800000  6.800000  7.000000  6.400000  5.600000
#>    [8]  6.400000  5.272727  5.000000  6.230769  8.750000  6.600000  6.333333
#>   [15]  7.250000  6.333333  6.750000  5.500000  5.000000  6.363636  8.200000
#>   [22]  8.000000  5.500000  5.500000  5.333333  6.111111  5.500000  6.250000
#>   [29]  6.000000  5.625000  5.500000  5.000000  5.375000  5.666667  6.833333
#>   [36]  6.400000  5.714286  5.000000  5.800000  8.000000  5.000000  5.000000
#>   [43]  4.800000  5.833333  5.250000  6.285714  7.500000  5.428571  6.125000
#>   [50]  7.500000  6.666667  4.000000  6.666667  7.750000  6.666667  6.000000
#>   [57]  5.666667  7.000000  4.714286  5.750000  4.000000  6.555556  6.000000
#>   [64]  4.500000  5.000000  4.000000  5.000000  3.500000  7.000000  7.666667
#>   [71]  7.142857  5.750000  9.333333  6.142857  5.333333  4.800000  4.250000
#>   [78]  8.000000  5.666667  3.875000  5.833333  7.000000  5.857143  5.625000
#>   [85]  6.666667  6.333333  7.428571  4.000000  5.714286  6.500000  5.166667
#>   [92]  4.333333  8.000000  6.166667  8.000000  6.250000  5.666667  8.333333
#>   [99]  4.750000  5.800000  7.000000  4.666667  6.666667  6.400000  5.666667
#>  [106]  6.600000  5.625000  7.800000  7.200000  6.500000  7.333333  7.500000
#>  [113]  8.750000  6.000000  7.000000  7.200000  6.571429  7.000000  6.600000
#>  [120]  6.400000  4.777778  6.500000  5.200000  6.200000  4.500000  7.500000
#>  [127]  6.000000  6.666667  9.500000  7.000000  4.333333  7.666667  6.250000
#>  [134]  6.500000  6.000000  6.250000  5.555556  6.000000  6.000000  8.333333
#>  [141]  6.500000  5.571429  5.333333  4.000000  5.000000  5.833333       NaN
#>  [148]  4.750000  6.250000  6.000000  5.142857  6.500000  5.200000  5.571429
#>  [155]  6.666667  5.800000  5.333333  6.666667  7.333333  4.500000  8.200000
#>  [162]  8.200000  5.000000  5.600000  6.200000  4.000000  6.000000  6.200000
#>  [169]  5.833333  5.000000  5.600000  6.000000  4.333333  5.600000  5.400000
#>  [176]  6.333333  5.200000  6.428571  6.333333  7.000000  7.666667  6.000000
#>  [183]  5.000000  7.000000       NaN  4.333333  5.250000  6.600000  4.750000
#>  [190]  4.625000  6.800000       NaN  6.750000  5.400000  4.000000  8.000000
#>  [197]  7.500000  7.000000  5.500000  7.833333  5.600000  6.666667  6.200000
#>  [204]  5.750000  5.000000  6.000000  5.250000  6.750000  5.714286  9.500000
#>  [211]  4.800000  5.333333  4.600000  7.000000  5.714286  6.666667  6.111111
#>  [218]  6.000000  6.000000  8.250000  4.400000  5.200000  8.000000  5.333333
#>  [225]  4.600000  5.500000  4.750000  4.000000  5.000000  6.333333  6.363636
#>  [232]  5.750000  7.857143  7.125000  5.833333  5.000000  6.500000  5.800000
#>  [239]  5.500000  5.857143  7.000000  6.250000  6.200000  6.200000  6.250000
#>  [246]  6.000000  7.500000  6.111111  6.166667  5.777778  6.250000  6.200000
#>  [253]  6.000000  5.714286  2.500000  5.800000  6.250000  8.571429  3.000000
#>  [260]  6.400000  6.500000  4.500000  5.250000  6.000000  5.666667  7.800000
#>  [267]  5.666667  5.750000  6.750000  5.750000  6.200000  6.400000  6.000000
#>  [274]  4.333333  4.500000  6.166667  5.333333  6.000000  6.700000  6.400000
#>  [281]  7.800000  5.142857  5.000000  6.666667  8.800000  5.000000  4.000000
#>  [288]  8.200000  6.000000  6.400000  7.200000  5.600000  6.800000  5.250000
#>  [295]  6.500000  6.666667  5.777778  6.142857  4.800000  6.000000  4.500000
#>  [302]  6.625000  4.500000  6.000000  6.666667  7.400000  5.444444  7.636364
#>  [309]  7.000000  5.500000  6.857143  5.400000  6.000000  6.333333  7.166667
#>  [316]  5.500000  4.333333  7.666667  7.000000  4.000000  6.166667  4.833333
#>  [323]  6.000000  7.222222  5.800000  4.600000  5.600000  6.500000  5.333333
#>  [330]  9.000000  6.142857  6.600000  7.400000  7.500000  6.200000  6.000000
#>  [337]  6.714286  5.333333  6.500000  6.333333  5.200000  8.000000  7.000000
#>  [344]  7.400000  3.500000  8.666667  4.833333  6.000000  5.200000  5.571429
#>  [351]  7.666667  7.000000  5.833333  6.100000  5.333333  6.666667  9.000000
#>  [358]  6.500000  7.833333  6.500000  5.166667  5.400000  6.222222  7.428571
#>  [365]  7.500000  7.666667  6.250000  6.333333  6.571429  6.250000  8.833333
#>  [372]       NaN  7.000000  6.250000  7.666667  4.333333  6.833333  6.285714
#>  [379]  8.500000  5.625000  5.400000  7.666667  8.000000  6.500000  6.285714
#>  [386]  4.000000  5.000000  5.700000  6.000000  6.714286  8.166667  4.333333
#>  [393]  6.666667  4.333333  7.000000  6.800000  7.000000  6.500000  6.750000
#>  [400]  7.400000  5.777778  6.250000  7.000000  6.400000  6.600000  5.500000
#>  [407]  6.500000  4.800000  7.666667  6.750000  7.750000  7.000000  5.166667
#>  [414]  8.142857  3.500000  6.875000  5.200000  8.000000  7.000000  5.600000
#>  [421]  5.857143  6.750000  7.666667  6.500000  6.571429  7.400000  6.833333
#>  [428]  4.666667  6.500000  6.875000  6.444444  8.500000  4.750000  7.750000
#>  [435]  6.250000  5.600000  5.166667  7.750000  7.500000  7.000000  6.250000
#>  [442]  4.833333  6.571429  5.666667  6.166667  5.000000  5.750000  6.000000
#>  [449]  5.400000  5.625000  5.200000  7.000000  8.000000  6.714286  6.250000
#>  [456]  6.333333  5.333333  8.000000  5.000000  7.000000  6.500000  3.500000
#>  [463]  6.666667  5.666667  6.333333  5.666667  4.500000  6.833333  5.500000
#>  [470]  5.555556  4.000000  7.100000  5.400000  4.250000  6.200000  5.600000
#>  [477]  5.000000  5.250000  7.000000  5.000000  6.333333  7.166667  5.285714
#>  [484]  7.000000  7.000000  9.666667  6.666667  4.750000  6.714286  6.800000
#>  [491]  6.166667  5.200000  9.000000  8.500000  7.500000  5.375000  4.500000
#>  [498]  6.500000  5.857143  4.285714  5.833333  5.857143  5.833333  4.571429
#>  [505]  7.500000  5.000000  8.000000  6.625000  6.000000  7.400000  8.000000
#>  [512]  4.333333  5.250000  4.500000  6.500000  6.600000  6.666667  6.666667
#>  [519]  4.000000  5.000000  5.400000  6.076923  9.333333  6.714286  7.333333
#>  [526]  4.500000  4.200000  6.666667  7.000000  5.555556  6.700000  5.500000
#>  [533]       NaN  7.333333  6.142857  6.166667  7.000000  5.500000  5.000000
#>  [540]  5.000000  7.000000  6.333333  4.714286  6.000000  6.428571  5.600000
#>  [547]  7.200000  6.750000  6.142857  6.111111  6.000000  6.500000  5.600000
#>  [554]  3.000000  4.833333  6.166667  5.666667  4.571429  4.888889  9.333333
#>  [561]  6.300000  3.000000  7.500000  4.600000  7.333333  5.250000  5.875000
#>  [568]  7.000000  6.400000  6.333333  5.666667  4.000000  7.750000  4.000000
#>  [575]  6.000000  5.000000  4.800000  6.250000  8.000000  7.500000  7.000000
#>  [582]  6.571429  7.333333  6.500000  6.333333  7.250000  8.000000  5.400000
#>  [589]  5.333333  8.666667  8.750000  6.400000  6.333333  4.666667  7.000000
#>  [596]  6.000000  9.000000       NaN  7.333333  5.000000  7.000000  4.375000
#>  [603]  5.333333  7.600000  7.500000  8.000000  7.333333  5.714286  6.250000
#>  [610]  5.333333  5.125000  5.111111  6.000000  5.333333  6.833333  7.000000
#>  [617]  5.727273  6.000000  6.600000  4.333333  7.600000  6.000000  6.166667
#>  [624]  6.000000  7.250000  6.375000  5.500000  5.000000  4.250000  4.000000
#>  [631]  7.250000  5.666667  5.500000  6.666667  5.333333  6.800000  7.000000
#>  [638]  5.500000  6.600000  5.333333  4.800000  5.285714  6.222222  9.400000
#>  [645]  7.000000  5.600000  7.250000  3.714286  6.800000  6.750000  5.800000
#>  [652]  6.000000  5.666667  7.250000  5.857143  4.666667  6.500000  6.666667
#>  [659]  5.800000  6.833333  4.428571  5.571429  6.285714  5.285714 10.000000
#>  [666]  4.666667  4.200000  6.500000  5.142857  6.666667  5.444444  7.000000
#>  [673]  4.400000  6.200000  7.000000  5.333333  5.000000  8.000000  6.200000
#>  [680]  7.000000  7.000000  5.500000  5.500000  5.625000  7.166667  6.142857
#>  [687]  6.285714  5.166667  6.500000  3.833333  6.400000  2.666667  5.000000
#>  [694]  6.000000  6.166667  5.800000  6.500000  7.000000  6.400000  6.857143
#>  [701]  4.750000  6.000000  6.000000  5.600000  6.333333  4.500000  5.000000
#>  [708]  7.800000  7.750000  5.666667  6.600000  5.000000  4.000000  7.000000
#>  [715]  7.000000  5.500000  6.800000  6.500000  6.833333  6.833333  6.800000
#>  [722]  5.000000  5.875000  8.500000  6.100000  5.166667  4.500000  4.333333
#>  [729]  6.625000  7.000000  3.857143  5.833333  7.000000  5.333333  5.700000
#>  [736]  9.000000  6.666667  6.000000  6.000000  6.333333  6.444444  5.875000
#>  [743]  4.600000  7.125000  6.666667  6.000000  6.833333  7.400000  7.000000
#>  [750]  6.600000  6.333333  4.000000  5.500000  6.400000  4.800000  5.333333
#>  [757]  6.000000  5.800000  5.666667  6.200000  6.250000  6.333333  6.444444
#>  [764]  5.666667  6.666667  6.250000  5.500000  5.888889  5.750000  4.333333
#>  [771]  5.250000  4.666667  5.800000  4.600000  6.000000  6.250000  6.666667
#>  [778]  5.500000  5.666667  4.200000  5.750000  6.400000  5.000000  7.333333
#>  [785]  6.000000  8.000000  6.000000  5.400000  6.400000  6.285714  6.500000
#>  [792]  5.000000  6.000000  7.000000  6.800000  6.700000  5.285714  5.750000
#>  [799]  6.333333  5.444444  6.000000  6.000000  6.333333  6.500000  5.250000
#>  [806]  8.000000  7.500000  6.000000  5.333333  4.800000  7.875000  6.250000
#>  [813]  6.000000  5.875000  8.000000  6.285714  4.750000  6.166667  7.000000
#>  [820]  6.200000  5.400000  5.600000  7.600000  7.000000  6.000000  6.000000
#>  [827]  8.000000  5.200000  5.333333  5.600000  5.666667  4.800000  5.750000
#>  [834]  5.000000  7.444444  5.500000  7.666667  6.800000  5.500000  8.666667
#>  [841]  7.000000  6.166667  5.000000  6.375000  4.428571  6.000000  6.333333
#>  [848]  5.333333  7.500000  5.833333  7.000000       NaN  5.666667  5.000000
#>  [855]  5.800000  6.000000  4.666667  6.666667  5.250000  4.000000  7.000000
#>  [862]  5.500000  5.600000  7.375000  7.800000  5.000000  6.200000  5.142857
#>  [869]  6.500000  7.000000  5.166667  6.500000  5.333333       NaN  7.600000
#>  [876]  3.333333  4.000000  6.000000  5.750000  5.333333  6.000000  7.000000
#>  [883]  4.500000  6.000000  8.000000  7.000000  6.444444  5.333333  7.400000
#>  [890]  6.600000  5.000000  5.000000  6.714286  6.222222  5.600000  5.333333
#>  [897]  6.333333  6.600000  7.250000  2.500000  7.166667  5.750000  3.500000
#>  [904]  5.000000  6.300000  4.500000  5.571429  5.000000  8.000000  6.400000
#>  [911]  4.857143  5.714286  4.666667  4.500000  6.000000  5.750000       NaN
#>  [918]  5.571429  6.000000  4.600000  5.000000  6.333333  4.714286  3.250000
#>  [925]  4.333333  5.250000  8.000000  6.800000  5.250000  6.100000  6.500000
#>  [932]  4.250000  5.250000  6.500000  7.500000       NaN  5.428571  6.000000
#>  [939]  7.000000  4.500000  7.000000  6.000000  7.000000  6.000000  5.250000
#>  [946]  6.000000  5.818182  7.200000  6.000000  5.500000  6.750000  5.500000
#>  [953]  5.000000  6.000000  6.000000  5.600000  5.600000  7.000000  7.000000
#>  [960]  5.000000  6.500000  5.500000  5.900000  6.250000  6.750000  6.500000
#>  [967]  4.333333  7.800000  6.000000  7.200000  7.333333  5.166667  4.857143
#>  [974]       NaN  8.666667  4.166667  5.000000  5.666667  6.600000  5.600000
#>  [981]  5.200000  5.428571  8.666667  6.500000  5.375000  6.000000  4.333333
#>  [988]  6.555556  5.333333  7.400000  7.333333  2.000000  5.000000  5.833333
#>  [995]  6.250000  6.000000  5.888889  6.200000  4.833333       NaN
#> 
#> $knnk
#>  [1] 5.964286 5.782895 6.258547 6.132353 6.155080 6.105145 5.913323 6.034420
#>  [9] 6.012346 6.130435 6.025974 5.854167 6.153846
#> 

# A weighted graph
g5 <- make_star(10)
E(g5)$weight <- seq(ecount(g5))
knn(g5)
#> $knn
#>  [1] 1 9 9 9 9 9 9 9 9 9
#> 
#> $knnk
#> [1]   9 NaN NaN NaN NaN NaN NaN NaN   1
#>