Average nearest neighbor degree

Calculate the average nearest neighbor degree of the given vertices and the same quantity in the function of vertex degree

Usage

knn(
  graph,
  vids = V(graph),
  mode = c("all", "out", "in", "total"),
  neighbor.degree.mode = c("all", "out", "in", "total"),
  weights = NULL
)

Arguments

graph

The input graph. It may be directed.

vids

The vertices for which the calculation is performed. Normally it includes all vertices. Note, that if not all vertices are given here, then both ‘knn’ and ‘knnk’ will be calculated based on the given vertices only.

mode

Character constant to indicate the type of neighbors to consider in directed graphs. out considers out-neighbors, in considers in-neighbors and all ignores edge directions.

neighbor.degree.mode

The type of degree to average in directed graphs. out averages out-degrees, in averages in-degrees and all ignores edge directions for the degree calculation.

weights

Weight vector. If the graph has a weight edge attribute, then this is used by default. If this argument is given, then vertex strength (see strength()) is used instead of vertex degree. But note that knnk is still given in the function of the normal vertex degree. Weights are are used to calculate a weighted degree (also called strength()) instead of the degree.

Value

A list with two members:

knn

A numeric vector giving the average nearest neighbor degree for all vertices in vids.

knnk

A numeric vector, its length is the maximum (total) vertex degree in the graph. The first element is the average nearest neighbor degree of vertices with degree one, etc.

Details

Note that for zero degree vertices the answer in ‘knn’ is NaN (zero divided by zero), the same is true for ‘knnk’ if a given degree never appears in the network.

The weighted version computes a weighted average of the neighbor degrees as

$$k_{nn,u} = \frac{1}{s_u} \sum_v w_{uv} k_v,$$

where \(s_u = \sum_v w_{uv}\) is the sum of the incident edge weights of vertex u, i.e. its strength. The sum runs over the neighbors v of vertex u as indicated by mode. \(w_{uv}\) denotes the weighted adjacency matrix and \(k_v\) is the neighbors' degree, specified by neighbor_degree_mode.

References

Alain Barrat, Marc Barthelemy, Romualdo Pastor-Satorras, Alessandro Vespignani: The architecture of complex weighted networks, Proc. Natl. Acad. Sci. USA 101, 3747 (2004)

See also

Author

Gabor Csardi csardi.gabor@gmail.com

Related documentation in the C library

igraph_avg_nearest_neighbor_degree().

Examples

# Some trivial ones
g <- make_ring(10)
knn(g)
#> $knn
#>  [1] 2 2 2 2 2 2 2 2 2 2
#> 
#> $knnk
#> [1] NaN   2
#> 
g2 <- make_star(10)
knn(g2)
#> $knn
#>  [1] 1 9 9 9 9 9 9 9 9 9
#> 
#> $knnk
#> [1]   9 NaN NaN NaN NaN NaN NaN NaN   1
#> 

# A scale-free one, try to plot 'knnk'
g3 <- sample_pa(1000, m = 5)
knn(g3)
#> $knn
#>    [1]  14.926154  15.804348  13.437500  15.336700  16.507092  15.849817
#>    [7]  25.950000  66.230769  16.280423  19.182432  24.094737  49.942857
#>   [13] 190.285714  32.551724  25.250000  16.939394  63.368421  20.520000
#>   [19]  69.045455  52.518519 263.400000  49.160000  63.428571  19.065217
#>   [25]  88.266667  91.800000  44.638889 151.400000  21.871429  32.865385
#>   [31]  57.400000 146.222222 110.666667  16.471429  78.933333 102.166667
#>   [37]  29.595745  37.750000  43.633333  47.323529  60.550000 216.500000
#>   [43]  76.200000  90.166667  32.964286  39.032258  56.954545  43.655172
#>   [49]  40.628571  59.642857  48.965517  54.176471  68.066667  82.875000
#>   [55] 107.600000  67.666667  68.428571  33.888889  52.166667  96.400000
#>   [61] 101.937500 131.000000 255.400000 206.500000 136.200000  50.258065
#>   [67]  51.125000  31.181818  41.526316  30.888889 138.250000 143.333333
#>   [73]  42.354839  46.090909  57.800000  55.600000  29.416667 156.875000
#>   [79] 187.333333 114.833333 119.714286  80.200000  94.000000  36.250000
#>   [85]  59.200000 163.200000 115.583333 237.200000  92.777778  35.037037
#>   [91]  63.600000  87.636364  48.900000 111.571429  86.500000  51.071429
#>   [97] 251.400000  81.375000 184.000000 144.600000  73.000000  86.153846
#>  [103]  45.818182  76.142857 127.888889 204.000000 268.000000 245.500000
#>  [109] 109.800000 180.666667 111.000000 202.600000 221.333333 197.600000
#>  [115]  88.000000  99.888889 186.666667 177.428571 182.400000 172.800000
#>  [121] 172.000000 184.166667 106.153846 101.600000 120.142857 125.666667
#>  [127] 170.000000  46.105263  79.666667  28.756757 123.166667 133.800000
#>  [133]  77.062500  80.071429 148.750000 224.000000 163.875000  96.833333
#>  [139]  64.733333  96.250000  74.214286  80.333333 213.500000  66.647059
#>  [145]  30.875000 136.444444  34.444444  67.272727  23.307692 233.000000
#>  [151] 113.200000 143.625000  89.272727  64.761905  84.000000  35.066667
#>  [157]  94.500000  74.750000  84.000000 107.600000  98.375000  59.421053
#>  [163] 185.200000 170.571429 207.666667  45.200000 171.571429 175.500000
#>  [169]  60.714286 139.142857  75.714286 120.000000 139.777778 262.800000
#>  [175]  59.090909  77.142857 175.800000 155.600000  83.400000  75.111111
#>  [181] 116.300000 135.750000 156.125000 134.800000  61.833333 163.833333
#>  [187]  82.888889  60.384615  59.700000 163.142857 137.285714 125.000000
#>  [193] 105.142857  35.952381 171.400000 165.000000  65.947368 135.333333
#>  [199]  57.428571  67.500000 248.200000  19.285714 161.000000 120.600000
#>  [205] 151.857143 102.000000 235.000000 117.000000 116.714286 213.666667
#>  [211] 127.142857 117.200000 224.400000 214.000000  97.000000 129.666667
#>  [217]  40.777778 140.428571 113.222222 258.000000 124.200000  39.222222
#>  [223] 156.000000 107.000000  83.100000 244.600000 102.800000  71.000000
#>  [229] 176.400000 201.200000 192.400000 182.200000 125.200000  54.600000
#>  [235]  44.800000  80.000000 117.000000  61.692308  32.800000  44.307692
#>  [241] 135.666667  44.125000 171.000000 112.555556  83.000000  86.222222
#>  [247] 135.000000  38.750000  52.600000 156.800000  65.000000 107.500000
#>  [253]  84.090909  34.000000 113.833333  59.454545 125.500000  91.750000
#>  [259]  90.800000 145.600000 128.000000 108.125000  57.400000  94.250000
#>  [265]  99.875000  57.909091  54.142857  62.500000  55.555556  72.333333
#>  [271]  95.285714 213.000000 128.625000 257.600000 125.666667 198.400000
#>  [277]  76.625000  48.428571 105.200000 118.875000  78.400000 114.428571
#>  [283] 121.000000 116.400000  96.142857 189.400000 170.000000  60.375000
#>  [289] 254.000000 150.600000 162.000000 194.800000 122.666667 134.200000
#>  [295] 115.833333 110.714286 217.800000 100.000000 134.714286  12.562500
#>  [301] 159.000000  81.090909  66.285714 142.600000 113.400000  43.272727
#>  [307] 143.000000 161.000000  64.166667 202.000000  60.166667  63.125000
#>  [313] 123.500000  52.000000 177.000000  69.200000 138.800000 102.666667
#>  [319] 153.333333 259.000000  64.750000 144.000000  52.000000  25.571429
#>  [325]  91.000000  58.133333 110.800000  68.888889 158.400000 175.800000
#>  [331] 104.875000 171.600000 155.857143 202.200000 151.200000  86.625000
#>  [337]  99.285714 131.833333 190.800000  74.111111 136.166667 135.000000
#>  [343]  88.857143  67.500000 168.285714 108.166667  70.857143  34.000000
#>  [349] 103.600000  58.100000  91.833333 111.000000  57.000000 207.600000
#>  [355] 140.200000 164.000000  90.166667 149.600000  87.500000 160.400000
#>  [361] 186.000000 137.000000 204.000000 104.571429 121.800000  49.111111
#>  [367]  36.714286 129.400000 116.800000 116.714286  51.571429  64.833333
#>  [373] 129.400000 259.800000 110.714286  62.545455 248.400000 155.200000
#>  [379] 204.666667 175.333333 145.000000  91.500000 151.000000 196.800000
#>  [385] 101.400000 148.600000 124.857143  91.142857 200.400000 123.400000
#>  [391] 150.200000 107.571429  97.333333  69.428571 105.000000 112.875000
#>  [397] 187.500000  47.285714  97.666667 154.200000 143.166667 116.666667
#>  [403] 106.400000  61.083333  67.571429 144.142857 101.400000  55.333333
#>  [409]  99.000000  95.600000 167.166667 116.800000  43.200000 173.571429
#>  [415] 102.200000 127.500000 186.833333 138.000000  87.000000 139.285714
#>  [421] 172.800000 167.200000 136.000000 180.200000 131.333333 142.600000
#>  [427] 131.800000  77.000000 165.428571 122.000000 138.333333  68.111111
#>  [433]  59.600000 244.600000 181.600000 157.800000  75.571429  91.333333
#>  [439]  65.875000 122.500000 140.000000 153.600000 150.800000 105.875000
#>  [445]  66.000000 202.000000  86.750000  18.166667 129.000000 222.400000
#>  [451] 196.600000  87.166667 120.200000 187.714286  87.714286  51.000000
#>  [457] 214.000000 170.400000 132.200000 142.200000 121.200000 111.428571
#>  [463] 192.000000  62.000000 107.400000 179.000000 118.750000  95.000000
#>  [469] 116.400000 107.000000 124.500000  95.400000 204.800000  59.375000
#>  [475] 121.800000 125.000000  86.250000  76.000000 144.800000 123.333333
#>  [481]  70.615385  89.000000 112.200000  81.000000 232.600000 168.800000
#>  [487] 144.285714 110.714286 103.285714 163.333333 179.800000 135.200000
#>  [493] 111.333333 129.400000 109.714286 129.285714 226.800000 208.000000
#>  [499] 127.833333 200.000000  44.833333  89.666667 157.600000 114.400000
#>  [505]  50.333333 105.333333 161.500000 122.833333 143.200000 126.800000
#>  [511] 169.800000 176.142857 123.600000  42.000000  75.400000  83.500000
#>  [517] 150.800000 155.800000 192.000000  79.000000  92.200000 146.600000
#>  [523]  77.200000  88.000000 218.800000  73.000000 165.666667 126.800000
#>  [529]  87.800000  92.800000  60.250000  62.000000 151.200000  50.222222
#>  [535]  99.857143 177.666667 127.800000 123.800000 192.600000 142.800000
#>  [541] 102.400000 194.400000  89.000000 130.800000 150.000000  75.200000
#>  [547] 100.600000 142.285714 243.800000 141.000000 143.500000 137.000000
#>  [553] 121.600000 162.400000 129.500000  29.500000  78.666667 239.800000
#>  [559]  41.428571 103.285714 126.600000 134.714286 178.800000 213.600000
#>  [565]  52.888889 147.400000 102.600000 206.714286 124.000000 125.166667
#>  [571] 134.200000 108.833333 205.800000 151.000000 139.500000 148.600000
#>  [577] 158.000000 152.166667 237.400000 145.428571 110.142857  91.428571
#>  [583] 147.000000  53.600000 164.200000 107.800000 147.400000  84.000000
#>  [589] 147.666667 104.200000  74.625000  91.333333 183.800000 114.200000
#>  [595] 163.500000  87.166667  50.600000 194.800000  31.800000  87.600000
#>  [601] 144.000000 177.200000 170.800000 171.166667 252.800000 261.400000
#>  [607] 196.600000  94.200000 122.600000 212.800000 113.200000 187.200000
#>  [613] 140.000000  43.800000 114.500000 272.400000 160.800000 129.285714
#>  [619] 192.166667 118.200000 141.800000 130.000000 217.800000 210.600000
#>  [625]  79.333333  54.000000 137.600000 167.000000 122.000000 121.285714
#>  [631] 155.000000 175.800000 238.200000  76.200000  98.800000 121.400000
#>  [637] 187.600000 137.600000 130.200000 104.166667 106.400000  15.400000
#>  [643] 271.800000 217.600000 146.000000  64.142857  79.400000 157.000000
#>  [649] 117.400000 234.200000  56.600000 147.200000 115.800000 199.000000
#>  [655]  43.800000  95.571429 183.285714  18.833333 147.000000  38.800000
#>  [661] 249.400000  75.800000 177.600000 157.666667 159.600000 137.125000
#>  [667] 109.285714  74.600000 237.333333  30.000000 147.000000  63.600000
#>  [673]  65.800000 118.800000 128.500000 139.600000 215.000000 116.666667
#>  [679]  17.800000 204.800000  43.333333 135.500000 118.600000 146.800000
#>  [685] 122.625000 102.000000 194.000000 141.400000 220.400000 205.800000
#>  [691] 122.000000 201.200000  28.571429  89.800000 207.200000  93.777778
#>  [697] 140.000000 118.833333  77.428571 119.800000 189.600000 146.666667
#>  [703]  62.000000  45.000000  40.200000  67.600000 148.800000  68.000000
#>  [709]  60.500000 201.000000 128.166667 158.000000  87.166667 168.000000
#>  [715]  27.400000  13.800000 126.714286  72.833333  94.600000 158.000000
#>  [721] 140.800000 205.200000 202.600000 216.200000 113.333333 184.333333
#>  [727] 140.200000  64.000000  69.833333 112.500000  74.400000 157.400000
#>  [733] 167.000000  85.500000  84.714286  28.000000 106.600000 166.833333
#>  [739] 105.000000  14.400000 155.400000 215.000000  72.200000 133.200000
#>  [745]  81.800000 117.200000  91.400000  80.000000 117.666667 119.800000
#>  [751] 103.200000 145.400000 118.800000 144.200000 245.400000  52.000000
#>  [757]  13.600000 143.800000 138.000000  94.600000 158.200000 207.200000
#>  [763] 167.000000 104.200000  75.833333  99.800000 150.600000 242.600000
#>  [769] 114.000000 114.714286  65.400000  53.000000 111.200000 173.333333
#>  [775]  52.000000  74.800000 109.000000  73.500000 145.000000  62.000000
#>  [781] 150.600000 245.000000  59.000000  98.000000 169.200000  64.200000
#>  [787]  78.333333 167.400000 170.666667 170.200000 174.200000 176.400000
#>  [793] 104.000000 126.200000  41.400000 148.800000 147.400000  93.000000
#>  [799] 188.800000 105.600000  80.200000 194.200000  85.200000 126.200000
#>  [805]  28.800000 160.800000  74.000000  50.800000  54.333333  98.166667
#>  [811] 178.000000  42.000000  70.666667 141.200000 252.000000 311.000000
#>  [817]  19.800000  11.333333 143.600000 201.000000 175.600000 143.166667
#>  [823] 219.800000 248.800000 129.200000 111.200000 219.400000 113.800000
#>  [829]  34.200000  85.333333 149.800000 118.200000 186.400000 177.600000
#>  [835] 124.200000 172.800000 184.000000 101.200000 114.000000  15.000000
#>  [841] 120.600000  93.400000 170.200000  82.200000  99.800000 186.800000
#>  [847]  29.000000  53.000000 174.000000  48.600000  87.800000 185.166667
#>  [853]   9.857143 174.000000  83.800000 110.800000 166.400000 123.333333
#>  [859] 141.600000  71.200000 174.800000 133.600000 149.000000 215.600000
#>  [865] 136.166667 128.600000 124.200000 100.800000 254.200000  50.600000
#>  [871] 158.600000  97.000000 174.000000 174.200000 143.000000  20.400000
#>  [877] 124.400000  32.000000  92.800000 105.400000  80.600000  34.166667
#>  [883]  99.400000  61.400000 146.400000 204.000000  71.000000 101.000000
#>  [889] 121.000000  20.800000 191.400000 119.600000  35.800000  70.800000
#>  [895] 122.200000 126.000000  70.800000 120.200000  85.000000 122.600000
#>  [901]  79.800000 221.600000  34.600000  68.666667 157.600000  70.000000
#>  [907]  84.400000  28.600000 109.800000 146.800000  64.800000 141.800000
#>  [913]  61.000000 246.200000 174.600000 112.200000 103.200000 138.800000
#>  [919] 109.600000  83.600000  71.000000 121.000000 156.600000 112.400000
#>  [925] 196.600000 136.600000 202.600000 104.833333 131.200000 220.400000
#>  [931]  97.400000 151.600000  87.400000 194.000000 174.600000  84.000000
#>  [937]  51.800000 194.200000  93.400000 184.600000 117.666667 222.600000
#>  [943] 138.800000  99.400000 221.400000  87.000000 134.400000 144.200000
#>  [949]  52.000000 189.200000 185.800000 122.200000 109.285714 127.600000
#>  [955]  78.600000  27.400000 218.400000  22.400000 119.200000 175.400000
#>  [961]  93.800000  80.200000 120.000000 135.400000  89.400000  84.400000
#>  [967]  63.600000  80.200000 105.833333  25.400000 212.600000 140.000000
#>  [973] 184.000000 270.600000 163.000000 233.600000 183.200000 201.000000
#>  [979] 156.200000  16.800000 104.000000 102.800000 158.400000  39.600000
#>  [985]  89.600000  65.000000 142.600000  22.166667  65.400000 157.600000
#>  [991] 209.800000 135.800000 118.200000  71.200000 122.800000   8.200000
#>  [997]  81.200000  53.000000  74.000000 104.600000
#> 
#> $knnk
#>   [1]       NaN       NaN       NaN       NaN 137.53050 123.76667 108.61516
#>   [8]  99.74219  85.68199  85.30000  65.78788  73.07292  65.70192  69.49107
#>  [15]  70.68889  60.76250  60.41176  59.91667  55.27368  55.01667  54.71429
#>  [22]  50.81818       NaN  51.12500  49.16000  66.23077  38.08333  35.65476
#>  [29]  46.31034  50.51667  43.88172       NaN       NaN  47.32353  45.28571
#>  [36]  44.63889  28.75676       NaN       NaN       NaN       NaN       NaN
#>  [43]       NaN       NaN       NaN       NaN  29.59574       NaN       NaN
#>  [50]       NaN       NaN  32.86538       NaN       NaN       NaN       NaN
#>  [57]       NaN  32.55172       NaN       NaN       NaN       NaN       NaN
#>  [64]       NaN       NaN       NaN       NaN       NaN       NaN  21.87143
#>  [71]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#>  [78]       NaN       NaN  25.95000       NaN       NaN       NaN  25.25000
#>  [85]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#>  [92]  19.06522       NaN       NaN  24.09474       NaN       NaN       NaN
#>  [99]       NaN  20.52000       NaN       NaN       NaN       NaN       NaN
#> [106]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [113]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [120]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [127]       NaN       NaN       NaN       NaN       NaN  16.93939       NaN
#> [134]       NaN       NaN       NaN       NaN       NaN       NaN  16.47143
#> [141]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [148]  19.18243       NaN       NaN       NaN       NaN       NaN       NaN
#> [155]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [162]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [169]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [176]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [183]       NaN       NaN       NaN       NaN       NaN       NaN  16.28042
#> [190]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [197]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [204]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [211]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [218]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [225]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [232]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [239]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [246]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [253]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [260]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [267]       NaN       NaN       NaN       NaN       NaN       NaN  15.84982
#> [274]       NaN       NaN  15.80435       NaN       NaN       NaN       NaN
#> [281]       NaN  16.50709       NaN       NaN       NaN       NaN       NaN
#> [288]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [295]       NaN       NaN  15.33670       NaN       NaN       NaN       NaN
#> [302]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [309]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [316]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [323]       NaN       NaN  14.92615       NaN       NaN       NaN       NaN
#> [330]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [337]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [344]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [351]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [358]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [365]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [372]       NaN       NaN       NaN       NaN       NaN       NaN       NaN
#> [379]       NaN       NaN       NaN       NaN       NaN  13.43750
#> 

# A random graph
g4 <- sample_gnp(1000, p = 5 / 1000)
knn(g4)
#> $knn
#>    [1] 10.000000  4.750000  6.250000  4.600000  5.666667  6.800000  5.888889
#>    [8]  4.500000  4.600000  5.166667  6.428571  8.500000  8.000000  6.200000
#>   [15]  8.000000  4.000000  5.250000  3.000000  6.500000  6.333333  9.666667
#>   [22]  7.833333  6.000000  6.000000  6.444444  6.500000  5.285714  4.250000
#>   [29]  9.000000  5.500000  5.454545  5.625000  5.500000  5.000000  7.000000
#>   [36]  6.500000  6.666667  6.200000  5.900000  6.000000  6.500000  6.000000
#>   [43]  5.000000  7.500000  6.500000  5.625000  7.000000  5.142857  4.333333
#>   [50]  3.500000  6.375000  5.500000  7.000000  7.600000  7.750000  5.666667
#>   [57]  6.000000  5.833333  7.250000  6.666667  6.000000  6.666667  6.750000
#>   [64]  5.700000  6.555556  6.000000  6.333333  6.666667  5.500000  6.222222
#>   [71]  7.000000  6.500000  5.500000  5.000000  5.800000  4.625000  6.600000
#>   [78]  6.400000  7.400000  7.200000  4.600000  5.571429  7.000000  5.857143
#>   [85]  6.250000  4.400000  7.000000  5.500000  6.000000  5.500000  7.000000
#>   [92]  7.000000  3.750000  5.000000  6.000000  5.000000  6.714286  7.000000
#>   [99]  9.400000  5.857143  7.000000  6.333333  6.000000  6.750000  5.000000
#>  [106]  6.833333  6.800000  4.750000  5.500000  5.666667  9.000000  6.428571
#>  [113]  5.750000  5.000000  6.000000  5.500000  8.833333  6.000000  3.833333
#>  [120]  5.750000  4.250000  6.500000  7.000000  6.250000  5.000000  5.400000
#>  [127]  5.800000  9.500000  7.333333  5.714286  6.200000  6.400000  6.333333
#>  [134]  5.857143  7.142857  6.666667  5.400000  6.000000  6.714286  4.333333
#>  [141]  7.000000  6.142857  5.555556  5.818182  4.600000  6.833333  2.000000
#>  [148]  4.600000  3.000000  5.625000  6.000000  4.500000  4.000000  4.750000
#>  [155]  6.200000  7.000000  5.000000  5.500000  5.500000  5.400000  6.571429
#>  [162]  8.800000  6.000000  6.666667  7.000000  5.000000  9.500000  6.000000
#>  [169]  6.500000  7.000000  8.000000  7.200000  5.666667  6.714286  4.666667
#>  [176]  4.000000  5.600000  5.666667  4.000000  6.600000  7.600000  5.857143
#>  [183]  4.000000  7.000000  6.000000  6.500000  4.857143 10.000000  7.500000
#>  [190]  7.500000  6.000000  5.375000  7.000000  7.000000  6.000000  5.666667
#>  [197]  5.333333  7.166667  7.000000  5.800000  5.000000  6.000000  5.000000
#>  [204]  6.857143  3.666667  5.500000  5.750000  4.250000  6.750000  7.333333
#>  [211]  5.000000  4.000000  7.000000  3.500000  5.333333  6.750000  6.500000
#>  [218]  5.800000  6.500000  5.714286  5.333333  8.000000  6.272727  6.750000
#>  [225]  4.666667  6.833333  6.400000  6.800000  4.500000  6.000000  8.333333
#>  [232]  6.777778  5.000000  6.400000  6.000000  5.000000  7.400000  4.500000
#>  [239]  5.500000  7.000000  6.666667  6.250000  7.250000  4.833333  6.666667
#>  [246]  7.250000  4.500000  6.000000  6.428571  7.000000  6.600000  4.333333
#>  [253]  6.750000  6.750000  6.636364  6.600000  6.900000  6.142857  4.500000
#>  [260]  5.200000  6.000000  5.142857  5.666667  8.333333  8.250000  6.250000
#>  [267]  6.333333  5.857143  5.500000  6.500000  6.000000  6.833333  7.833333
#>  [274]  5.100000  7.000000  5.000000  6.250000  6.800000  5.375000  5.500000
#>  [281]  6.000000  7.750000  7.500000  6.400000  6.000000  6.666667  6.000000
#>  [288]  4.400000  4.800000  4.666667  4.250000  6.500000  5.875000  5.750000
#>  [295]  6.666667  8.600000  6.200000  6.333333  5.500000  7.000000  7.000000
#>  [302]  6.000000  6.750000  4.666667  6.000000  6.000000  8.000000  7.500000
#>  [309]  6.666667  6.000000  7.400000  6.800000  5.166667  5.000000  6.636364
#>  [316]  5.714286  6.000000  7.400000  5.800000  7.750000  5.285714  6.900000
#>  [323]  5.571429  5.500000  5.600000  4.500000  6.142857  7.833333  7.400000
#>  [330]  6.000000  6.375000  4.909091  7.000000  7.285714  7.166667  6.000000
#>  [337]  6.142857  5.166667  8.333333  6.400000  5.857143  6.250000  7.000000
#>  [344]  6.857143  6.200000  7.400000  6.500000  3.500000  6.200000  5.250000
#>  [351]  5.000000  4.166667  4.500000  6.000000  5.333333  4.000000  6.333333
#>  [358]  6.200000  6.555556  6.400000  5.166667  6.400000  6.500000  6.200000
#>  [365]  5.000000  6.166667  5.750000  9.333333  7.000000  5.666667  6.142857
#>  [372]  7.250000  7.166667  6.111111  5.888889  5.857143  3.666667  5.555556
#>  [379]  4.000000  5.111111  7.000000  5.375000  6.750000  5.875000  6.285714
#>  [386]  6.250000  6.166667  6.888889  5.500000  5.400000  5.333333  4.000000
#>  [393]  6.666667  4.600000  3.250000  5.600000  9.000000  3.800000  7.000000
#>  [400]  5.333333  6.333333  7.000000  5.857143  6.333333  6.428571  9.000000
#>  [407]  6.500000  4.428571  7.111111  5.000000  5.500000  5.875000  5.666667
#>  [414]  7.250000  6.000000  5.714286  6.111111  2.666667  5.727273  3.500000
#>  [421]  5.833333  6.333333  4.000000  6.666667  7.000000  5.000000  6.500000
#>  [428]  4.333333  6.545455  7.000000  3.000000  6.285714  6.333333  7.500000
#>  [435]  5.500000  4.200000  4.500000  4.500000  4.400000  6.250000  5.166667
#>  [442]  6.000000  5.714286  7.333333  8.500000  7.200000  6.500000  5.666667
#>  [449]  5.666667  6.000000  7.500000  5.500000  6.285714  5.666667  7.200000
#>  [456]  4.000000  4.000000  5.375000  6.000000  8.333333  5.200000  6.000000
#>  [463]  7.666667  5.000000  7.000000  5.125000  5.000000  7.571429  7.000000
#>  [470]  5.000000  4.500000  5.250000  6.333333  6.250000  5.875000  6.750000
#>  [477]  6.000000  7.333333  7.600000  9.000000  6.142857  5.200000  6.600000
#>  [484]  5.000000  6.666667  7.250000  5.500000  4.000000  7.000000  5.600000
#>  [491]  7.000000  7.500000  6.555556  4.250000  6.300000  6.750000  5.375000
#>  [498]  5.833333  5.142857  5.400000  5.400000  6.571429  4.800000  5.800000
#>  [505]  6.800000  6.000000  6.000000  9.000000  7.500000  4.250000  7.000000
#>  [512]  6.000000  6.000000  5.400000  4.750000  4.333333  4.666667  4.000000
#>  [519]  5.833333  6.500000  6.000000  6.500000  5.833333  6.666667  6.333333
#>  [526]  5.333333  6.000000  5.250000  5.636364  6.250000  8.500000  7.500000
#>  [533]  6.000000  5.666667  5.800000  3.500000  4.333333  5.900000  6.666667
#>  [540]  4.666667  6.666667  8.250000  7.666667  5.200000  4.333333  5.666667
#>  [547]  5.500000  5.000000  7.100000  5.083333  3.000000  6.000000  8.250000
#>  [554]  7.400000  5.666667  6.250000  5.666667  6.166667  5.900000       NaN
#>  [561]  4.000000  5.500000  6.000000  6.750000  4.857143  7.500000       NaN
#>  [568]  5.000000  6.666667  6.900000  6.000000  4.500000  5.750000  5.500000
#>  [575]  7.285714  5.200000  6.000000  5.666667  5.500000  7.333333  6.333333
#>  [582]  6.500000  8.714286  5.600000  7.250000  5.400000  4.500000  5.666667
#>  [589]  6.000000  7.500000  7.666667  5.800000  7.222222  5.166667  6.000000
#>  [596]  5.625000  6.125000  5.500000  7.545455  9.000000  4.200000  4.500000
#>  [603]  7.200000  4.400000  6.285714  4.500000  5.400000  5.000000  6.333333
#>  [610]  6.888889  6.000000  6.250000  4.000000  4.833333  5.500000  4.666667
#>  [617]  4.750000  6.500000  6.333333  4.250000  6.200000  7.166667  5.500000
#>  [624]  6.333333  4.875000  5.500000  5.285714  5.833333  7.250000  5.833333
#>  [631]  5.900000  5.750000  5.083333  5.500000  6.750000  4.666667  4.500000
#>  [638]  6.750000  6.000000  8.200000  6.200000  5.200000  7.833333  4.333333
#>  [645]  6.714286  6.125000  6.400000  6.000000  6.666667  6.800000  4.857143
#>  [652]  6.800000  5.333333  6.600000  5.166667  5.666667  4.000000  8.000000
#>  [659]  6.250000  3.000000  6.333333  5.000000  5.500000  5.166667  4.833333
#>  [666] 12.000000  6.125000  5.818182  4.500000  7.166667  5.000000  6.500000
#>  [673]  6.600000  6.142857  4.000000  4.000000  8.333333  6.250000  6.125000
#>  [680]  7.166667  7.000000  5.400000  5.800000  5.800000  6.000000  5.333333
#>  [687]  7.750000  4.500000  8.500000  5.666667  5.000000  4.800000  5.800000
#>  [694]  7.500000  8.333333  5.500000  6.500000  7.666667  5.000000  6.125000
#>  [701]  7.500000  5.800000  4.800000  5.833333  7.000000  5.600000  6.600000
#>  [708]  6.571429  7.142857  7.375000  8.000000  5.666667  6.000000  5.666667
#>  [715]  5.166667  6.875000  6.500000  6.571429  6.600000  5.500000  5.000000
#>  [722]  7.333333  5.375000  5.500000  5.600000  5.857143  7.142857  7.200000
#>  [729]  6.500000  5.600000  6.600000  3.666667  6.250000  5.333333  5.928571
#>  [736]  6.571429  8.200000  6.400000  4.500000  8.000000  5.111111  6.400000
#>  [743]  7.166667  9.000000  4.250000  5.166667  4.500000  6.600000  5.625000
#>  [750]  5.500000  5.333333  5.500000  5.428571  5.100000  5.600000  7.000000
#>  [757]  7.750000  4.000000  7.500000  6.200000  8.000000  5.000000  6.000000
#>  [764]  6.000000  6.750000  3.600000  4.571429  5.500000  5.000000  6.333333
#>  [771]  4.333333  6.666667  5.600000  4.250000  6.000000  6.777778  7.000000
#>  [778]  6.000000  5.000000  6.200000  6.400000  5.111111  6.857143  7.750000
#>  [785]  5.571429  6.142857  7.666667  6.400000  5.833333  6.500000  5.571429
#>  [792]  4.333333  7.750000  8.000000  7.000000  5.666667  5.285714  6.200000
#>  [799]  5.800000  5.600000  7.600000  6.600000  8.250000  5.666667  7.666667
#>  [806]  8.000000  5.250000  6.000000  5.750000  7.000000  5.750000  4.000000
#>  [813]  6.666667  7.000000  5.000000  8.333333  5.250000  5.000000  5.833333
#>  [820]  5.666667  6.500000  5.250000  7.666667  4.600000  5.500000  6.250000
#>  [827]  5.666667  6.500000  7.000000  5.750000  6.285714  7.000000  5.000000
#>  [834]  8.333333  6.166667  8.714286  6.375000  4.375000  5.500000  6.714286
#>  [841]  4.333333  7.250000  6.750000  6.750000  7.222222  6.833333  8.000000
#>  [848]  8.500000  6.333333  4.500000  5.200000  5.200000  5.666667  5.666667
#>  [855]  7.000000  4.200000  6.111111  6.000000  7.750000  5.400000  4.000000
#>  [862]  6.750000  9.000000  4.000000  5.000000  6.000000  5.833333  4.375000
#>  [869]  6.000000  4.500000  5.857143  4.200000  6.500000  5.666667  5.857143
#>  [876]  6.000000  5.500000  5.400000  7.000000  5.500000  6.857143  6.142857
#>  [883]  7.285714  4.750000  5.600000  5.250000  5.750000  6.000000  6.333333
#>  [890]  5.200000  5.500000  8.400000  5.800000  5.400000  5.333333  8.400000
#>  [897]  4.500000  6.166667  6.750000  6.111111  6.750000  6.500000  6.333333
#>  [904]  6.500000  9.000000  5.600000  7.500000  5.833333  6.000000  5.000000
#>  [911]  6.125000  6.250000  6.333333  6.750000  6.333333  6.166667  7.000000
#>  [918]  5.625000  5.400000  6.400000  5.500000  6.000000  5.857143  6.000000
#>  [925]  7.333333  4.500000  6.000000  5.000000  6.428571  6.000000  6.500000
#>  [932]  5.222222  7.600000  6.875000  6.875000  6.333333  7.750000  3.333333
#>  [939]  6.666667  4.750000  4.800000  5.250000  8.500000  6.125000  6.285714
#>  [946]  7.000000  5.000000  3.000000  8.333333  6.500000  5.750000  6.333333
#>  [953]  4.600000  5.333333  4.000000  5.333333  6.000000  6.166667  3.000000
#>  [960]  8.400000  4.571429  6.200000  5.000000  5.800000  6.500000  5.000000
#>  [967]  5.600000  5.800000  7.000000  6.166667  7.750000  5.833333  5.444444
#>  [974]  4.600000  7.600000  5.833333  5.857143  7.250000  8.000000  4.750000
#>  [981]  7.000000  6.333333  6.200000  6.571429  4.500000  6.000000  5.571429
#>  [988]  5.000000  4.500000  5.666667  5.000000  2.500000  6.400000  6.500000
#>  [995]  6.800000  3.000000  7.250000  6.625000  6.500000  7.500000
#> 
#> $knnk
#>  [1] 5.625000 6.064706 5.994792 6.056452 6.114754 6.103571 6.128571 5.971429
#>  [9] 6.033951 5.977273 6.000000 5.666667      NaN 5.928571
#> 

# A weighted graph
g5 <- make_star(10)
E(g5)$weight <- seq(ecount(g5))
knn(g5)
#> $knn
#>  [1] 1 9 9 9 9 9 9 9 9 9
#> 
#> $knnk
#> [1]   9 NaN NaN NaN NaN NaN NaN NaN   1
#>