R igraph manual pages

Use this if you are using igraph from R

layout_with_lgl {igraph}	R Documentation

Large Graph Layout

Description

A layout generator for larger graphs.

Usage

layout_with_lgl(
  graph,
  maxiter = 150,
  maxdelta = vcount(graph),
  area = vcount(graph)^2,
  coolexp = 1.5,
  repulserad = area * vcount(graph),
  cellsize = sqrt(sqrt(area)),
  root = NULL
)

with_lgl(...)

Arguments

graph
The input graph

maxiter
The maximum number of iterations to perform (150).

maxdelta
The maximum change for a vertex during an iteration (the number of vertices).

area
The area of the surface on which the vertices are placed (square of the number of vertices).

coolexp
The cooling exponent of the simulated annealing (1.5).

repulserad
Cancellation radius for the repulsion (the area times the number of vertices).

cellsize
The size of the cells for the grid. When calculating the repulsion forces between vertices only vertices in the same or neighboring grid cells are taken into account (the fourth root of the number of area.

root
The id of the vertex to place at the middle of the layout. The default value is -1 which means that a random vertex is selected.

...
Passed to layout_with_lgl.

Details

layout_with_lgl is for large connected graphs, it is similar to the layout generator of the Large Graph Layout software (http://lgl.sourceforge.net/).

Value

A numeric matrix with two columns and as many rows as vertices.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

Other graph layouts: add_layout_(), component_wise(), layout_as_bipartite(), layout_as_star(), layout_as_tree(), layout_in_circle(), layout_nicely(), layout_on_grid(), layout_on_sphere(), layout_randomly(), layout_with_dh(), layout_with_fr(), layout_with_gem(), layout_with_graphopt(), layout_with_kk(), layout_with_mds(), layout_with_sugiyama(), layout_(), merge_coords(), norm_coords(), normalize()

[Package igraph version 1.3.2 Index]