Create a complete (full) citation graph

make_full_citation_graph() creates a full citation graph. This is a directed graph, where every i->j edge is present if and only if \(j<i\). If directed=FALSE then the graph is just a full graph.

Usage

make_full_citation_graph(n, directed = TRUE)

full_citation_graph(...)

Arguments

n

The number of vertices.

directed

Whether to create a directed graph.

...

Passed to make_full_citation_graph().

Value

An igraph graph.

See also

Examples

print_all(make_full_citation_graph(10))
#> IGRAPH 1b12dd6 D--- 10 45 -- Full citation graph
#> + attr: name (g/c)
#> + graph attributes:
#> | + name:
#> |   [1] "Full citation graph"
#> + edges:
#>  1 ->                      2 -> 1                    3 -> 1 2              
#>  4 -> 1 2 3                5 -> 1 2 3 4              6 -> 1 2 3 4 5        
#>  7 -> 1 2 3 4 5 6          8 -> 1 2 3 4 5 6 7        9 -> 1 2 3 4 5 6 7 8  
#> 10 -> 1 2 3 4 5 6 7 8 9