make_kautz_graph {igraph}	R Documentation

Kautz graphs

Description

Kautz graphs are labeled graphs representing the overlap of strings.

Usage

make_kautz_graph(m, n)

kautz_graph(...)

Arguments

m
Integer scalar, the size of the alphabet. See details below.

n
Integer scalar, the length of the labels. See details below.

...
Passed to make_kautz_graph.

Details

A Kautz graph is a labeled graph, vertices are labeled by strings of length n+1 above an alphabet with m+1 letters, with the restriction that every two consecutive letters in the string must be different. There is a directed edge from a vertex v to another vertex w if it is possible to transform the string of v into the string of w by removing the first letter and appending a letter to it.

Kautz graphs have some interesting properties, see eg. Wikipedia for details.

Value

A graph object.

Author(s)

Gabor Csardi <csardi.gabor@gmail.com>, the first version in R was written by Vincent Matossian.

See Also

make_de_bruijn_graph, make_line_graph

Examples

make_line_graph(make_kautz_graph(2,1))
make_kautz_graph(2,2)

[Package igraph version 1.2.2 Index]