Create a star graph, a tree with n vertices and n - 1 leaves

star() creates a star graph, in this every single vertex is connected to the center vertex and nobody else.

Usage

make_star(n, mode = c("in", "out", "mutual", "undirected"), center = 1)

star(...)

Arguments

n

Number of vertices.

mode

It defines the direction of the edges, in: the edges point to the center, out: the edges point from the center, mutual: a directed star is created with mutual edges, undirected: the edges are undirected.

center

ID of the center vertex.

...

Passed to make_star().

Value

An igraph graph.

See also

Examples

make_star(10, mode = "out")
#> IGRAPH ad7dd16 D--- 10 9 -- Out-star
#> + attr: name (g/c), mode (g/c), center (g/n)
#> + edges from ad7dd16:
#> [1] 1-> 2 1-> 3 1-> 4 1-> 5 1-> 6 1-> 7 1-> 8 1-> 9 1->10
make_star(5, mode = "undirected")
#> IGRAPH ad68626 U--- 5 4 -- Star
#> + attr: name (g/c), mode (g/c), center (g/n)
#> + edges from ad68626:
#> [1] 1--2 1--3 1--4 1--5