--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.5 --- {% raw %}

motifs {igraph}	R Documentation

Graph motifs

Description

Graph motifs are small connected subgraphs with a well-defined structure. These functions search a graph for various motifs.

Usage

motifs(graph, size = 3, cut.prob = rep(0, size))

Arguments

graph
Graph object, the input graph.

size
The size of the motif, currently sizes 3 and 4 are supported in directed graphs and sizes 3-6 in undirected graphs.

cut.prob
Numeric vector giving the probabilities that the search graph is cut at a certain level. Its length should be the same as the size of the motif (the size argument). By default no cuts are made.

Details

motifs searches a graph for motifs of a given size and returns a numeric vector containing the number of different motifs. The order of the motifs is defined by their isomorphism class, see isomorphism_class.

Value

motifs returns a numeric vector, the number of occurrences of each motif in the graph. The motifs are ordered by their isomorphism classes. Note that for unconnected subgraphs, which are not considered to be motifs, the result will be NA.

See Also

isomorphism_class

Other graph motifs: count_motifs(), sample_motifs()

Examples

g <- barabasi.game(100)
motifs(g, 3)
count_motifs(g, 3)
sample_motifs(g, 3)

[Package igraph version 1.3.5 Index]
{% endraw %}