--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.5 --- {% raw %}

path {igraph}	R Documentation

Helper function to add or delete edges along a path

Description

This function can be used to add or delete edges that form a path.

Usage

path(...)

Arguments

...
See details below.

Details

When adding edges via +, all unnamed arguments are concatenated, and each element of a final vector is interpreted as a vertex in the graph. For a vector of length n+1, n edges are then added, from vertex 1 to vertex 2, from vertex 2 to vertex 3, etc. Named arguments will be used as edge attributes for the new edges.

When deleting edges, all attributes are concatenated and then passed to delete_edges.

Value

A special object that can be used together with igraph graphs and the plus and minus operators.

See Also

Other functions for manipulating graph structure: +.igraph(), add_edges(), add_vertices(), delete_edges(), delete_vertices(), edge(), igraph-minus, vertex()

Examples

# Create a (directed) wheel
g <- make_star(11, center = 1) + path(2:11, 2)
plot(g)

g <- make_empty_graph(directed = FALSE, n = 10) %>%
  set_vertex_attr("name", value = letters[1:10])

g2 <- g + path("a", "b", "c", "d")
plot(g2)

g3 <- g2 + path("e", "f", "g", weight=1:2, color="red")
E(g3)[[]]

g4 <- g3 + path(c("f", "c", "j", "d"), width=1:3, color="green")
E(g4)[[]]

[Package igraph version 1.3.5 Index]
{% endraw %}