reciprocity {igraph}	R Documentation

Reciprocity of graphs

Description

Calculates the reciprocity of a directed graph.

Usage

reciprocity(graph, ignore.loops = TRUE) 

Arguments

graph 	 The graph object.

ignore.loops 	 Logical constant, whether to ignore loop edges.

Details

A vertex pair (A, B) is said to be reciprocal if there are edges between them in both directions. The reciprocity of a directed graph is the proportion of all possible (A, B) pairs which are reciprocal, provided there is at least one edge between A and B. The reciprocity of an empty graph is undefined (results in an error code). Undirected graphs always have a reciprocity of 1.0 unless they are empty.

Value

A numeric constaint between zero and one.

Author(s)

Tamas Nepusz ntamas@rmki.kfki.hu and Gabor Csardi csardi@rmki.kfki.hu

Examples

g <- random.graph.game(20, 5/20, directed=TRUE)
reciprocity(g)

[Package igraph version 0.5.2-2 Index]