--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.4 --- {% raw %}

rep.igraph {igraph}	R Documentation

Replicate a graph multiple times

Description

The new graph will contain the input graph the given number of times, as unconnected components.

Usage

## S3 method for class 'igraph'
rep(x, n, mark = TRUE, ...)

## S3 method for class 'igraph'
x * n

Arguments

x
The input graph.

n
Number of times to replicate it.

mark
Whether to mark the vertices with a which attribute, an integer number denoting which replication the vertex is coming from.

...
Additional arguments to satisfy S3 requirements, currently ignored.

Examples

rings <- make_ring(5) * 5

[Package igraph version 1.3.4 Index]
{% endraw %}