--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.4 --- {% raw %}

reverse_edges {igraph}	R Documentation

Reverse edges in a graph

Description

The new graph will contain the same vertices, edges and attributes as the original graph, except that the direction of the edges selected by their edge IDs in the eids argument will be reversed. When reversing all edges, this operation is also known as graph transpose.

Usage

reverse_edges(graph, eids = E(graph))

## S3 method for class 'igraph'
t(x)

Arguments

graph
The input graph.

eids
The edge IDs of the edges to reverse.

x
The input graph.

Value

The result graph where the direction of the edges with the given IDs are reversed

Examples

g <- make_graph( ~ 1-+2, 2-+3, 3-+4 )
reverse_edges(g, 2)

[Package igraph version 1.3.4 Index]
{% endraw %}