--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.2 --- {% raw %}

rewire {igraph}	R Documentation

Rewiring edges of a graph

Description

See the links below for the implemented rewiring methods.

Usage

rewire(graph, with)

Arguments

graph
The graph to rewire

with
A function call to one of the rewiring methods, see details below.

Value

The rewired graph.

See Also

Other rewiring functions: each_edge(), keeping_degseq()

Examples

g <- make_ring(10)
g %>%
  rewire(each_edge(p = .1, loops = FALSE)) %>%
  plot(layout=layout_in_circle)
print_all(rewire(g, with = keeping_degseq(niter = vcount(g) * 10)))

[Package igraph version 1.3.2 Index]
{% endraw %}