--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.4 --- {% raw %}

running_mean {igraph}	R Documentation

Running mean of a time series

Description

running_mean calculates the running mean in a vector with the given bin width.

Usage

running_mean(v, binwidth)

Arguments

v
The numeric vector.

binwidth
Numeric constant, the size of the bin, should be meaningful, ie. smaller than the length of v.

Details

The running mean of v is a w vector of length length(v)-binwidth+1. The first element of w id the average of the first binwidth elements of v, the second element of w is the average of elements 2:(binwidth+1), etc.

Value

A numeric vector of length length(v)-binwidth+1

Author(s)

Gabor Csardi csardi.gabor@gmail.com

Examples

running_mean(1:100, 10)

[Package igraph version 1.3.4 Index]
{% endraw %}