Generate a new random graph from a given graph by randomly adding/removing edges

Sample a new graph by perturbing the adjacency matrix of a given graph and shuffling its vertices.

Usage

sample_correlated_gnp(
  old.graph,
  corr,
  p = edge_density(old.graph),
  permutation = NULL
)

Arguments

old.graph

The original graph.

corr

A scalar in the unit interval, the target Pearson correlation between the adjacency matrices of the original and the generated graph (the adjacency matrix being used as a vector).

p

A numeric scalar, the probability of an edge between two vertices, it must in the open (0,1) interval. The default is the empirical edge density of the graph. If you are resampling an Erdos-Renyi graph and you know the original edge probability of the Erdos-Renyi model, you should supply that explicitly.

permutation

A numeric vector, a permutation vector that is applied on the vertices of the first graph, to get the second graph. If NULL, the vertices are not permuted.

Value

An unweighted graph of the same size as old.graph such that the correlation coefficient between the entries of the two adjacency matrices is corr. Note each pair of corresponding matrix entries is a pair of correlated Bernoulli random variables.

Details

Please see the reference given below.

References

Lyzinski, V., Fishkind, D. E., Priebe, C. E. (2013). Seeded graph matching for correlated Erdos-Renyi graphs. https://arxiv.org/abs/1304.7844

See also

Examples

g <- sample_gnp(1000, .1)
g2 <- sample_correlated_gnp(g, corr = 0.5)
cor(as.vector(g[]), as.vector(g2[]))
#> [1] 0.50259
g
#> IGRAPH 4f98d0c U--- 1000 49962 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 4f98d0c:
#>  [1]  4-- 7  5-- 8  7-- 8  1--10  5--10  2--12  4--12  6--14  4--15  6--15
#> [11] 10--15  3--16  6--16  1--17  7--17  6--20  8--21  1--22  8--22  2--23
#> [21] 15--23 14--24  9--25 12--25 14--25 23--25 12--26 20--26  6--27 20--27
#> [31]  6--28 16--28 26--28  3--29 17--29 18--29 23--29 15--30 17--30  4--31
#> [41] 25--31  4--32  5--32 20--32 25--32  9--33 15--33 30--33  3--34  6--34
#> [51] 11--34 33--34 19--35 20--35 24--35 16--36 23--36 27--36  4--37 10--37
#> [61]  1--38 13--38 19--38 26--38 29--38  5--39  6--39 13--39 32--39 24--40
#> [71] 25--40 32--40  3--41  6--41 10--41 14--41 16--41  3--42 11--42 33--42
#> + ... omitted several edges
g2
#> IGRAPH 71dd525 U--- 1000 49817 -- Correlated random graph
#> + attr: name (g/c), corr (g/n), p (g/n)
#> + edges from 71dd525:
#>  [1]  5-- 7  7-- 8  1--10  5--10  8--10  2--12  4--12  7--12  8--13  6--14
#> [11]  4--15  6--15  6--16  7--17 12--17 16--17  6--20  9--21  1--22  8--22
#> [21] 13--22 16--22  1--23  2--23 15--23  5--24  9--25 12--25 14--25 23--25
#> [31] 10--26 12--26  6--27  7--28  9--28 26--28 11--29 14--29 18--29 23--29
#> [41] 13--30 15--30 22--30 10--31 14--31 16--31 25--31  4--32 16--32 20--32
#> [51] 18--33 30--33  6--34 10--34 11--34 18--34  1--35 20--35 22--35 24--35
#> [61] 27--35 16--36 27--36 28--36  3--37  6--37  8--37 10--37 16--37  1--38
#> [71] 13--38 19--38 29--38 13--39 14--39 25--39  5--40 21--40 29--40 32--40
#> + ... omitted several edges