Generate a new random graph from a given graph by randomly adding/removing edges

Sample a new graph by perturbing the adjacency matrix of a given graph and shuffling its vertices.

Usage

sample_correlated_gnp(
  old.graph,
  corr,
  p = edge_density(old.graph),
  permutation = NULL
)

Arguments

old.graph

The original graph.

corr

A scalar in the unit interval, the target Pearson correlation between the adjacency matrices of the original and the generated graph (the adjacency matrix being used as a vector).

p

A numeric scalar, the probability of an edge between two vertices, it must in the open (0,1) interval. The default is the empirical edge density of the graph. If you are resampling an Erdős-Rényi graph and you know the original edge probability of the Erdős-Rényi model, you should supply that explicitly.

permutation

A numeric vector, a permutation vector that is applied on the vertices of the first graph, to get the second graph. If NULL, the vertices are not permuted.

Value

An unweighted graph of the same size as old.graph such that the correlation coefficient between the entries of the two adjacency matrices is corr. Note each pair of corresponding matrix entries is a pair of correlated Bernoulli random variables.

Details

Please see the reference given below.

References

Lyzinski, V., Fishkind, D. E., Priebe, C. E. (2013). Seeded graph matching for correlated Erdős-Rényi graphs. https://arxiv.org/abs/1304.7844

See also

Examples

g <- sample_gnp(1000, .1)
g2 <- sample_correlated_gnp(g, corr = 0.5)
cor(as.vector(g[]), as.vector(g2[]))
#> [1] 0.5004541
g
#> IGRAPH 80b7aa5 U--- 1000 49982 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 80b7aa5:
#>  [1]  5-- 6  6-- 9  9--11  5--12 11--12  2--14  3--14 10--17 14--17 15--17
#> [11]  4--18  9--18 11--19  9--20  5--21  6--21 17--21 18--21  6--22 18--22
#> [21] 21--22 11--23 14--23  2--24  4--25 11--25 22--25  5--26 12--26  7--27
#> [31] 11--27 15--27 23--27  5--28  6--28 14--28 15--28 18--28  9--29 12--29
#> [41]  5--30  8--30 17--30 29--30  9--31 31--32  5--33  8--33 13--33  1--34
#> [51]  8--34 13--34 11--35 14--35 27--35 32--35  3--36 10--36 11--36 22--36
#> [61] 26--36 33--36 12--38 25--38 16--39 17--39 27--39  4--40 22--40 23--40
#> [71] 24--40 32--40 39--40  2--41  3--41  9--41 20--41 10--42 12--42 41--42
#> + ... omitted several edges
g2
#> IGRAPH 9c7a3aa U--- 1000 49551 -- Correlated random graph
#> + attr: name (g/c), corr (g/n), p (g/n)
#> + edges from 9c7a3aa:
#>  [1]  5-- 6  4-- 7  4-- 9  6-- 9  4--11  9--11  5--12  3--14  1--17 10--17
#> [11]  9--18  6--19 11--19 15--19  8--20  5--21  6--21  5--22 13--22 18--22
#> [21] 21--22 14--23 15--25  5--26 12--26 13--26 17--26 20--26 10--27 11--27
#> [31] 23--27  5--28 14--28 15--28 18--28 19--28 26--28  6--29  9--29 10--29
#> [41] 12--29 17--30  6--31  9--31  9--32 31--32  5--33  6--33 11--33 13--33
#> [51] 17--33 22--33 31--33  1--34  7--34 13--34 11--35 27--35  3--36 11--36
#> [61] 21--36 22--36 24--36  5--37 19--37 29--38 17--39 27--39  4--40 22--40
#> [71] 24--40 28--40 32--40 39--40  3--41  9--41 20--41 10--42 41--42  3--43
#> + ... omitted several edges