Generate a new random graph from a given graph by randomly adding/removing edges

Sample a new graph by perturbing the adjacency matrix of a given graph and shuffling its vertices.

Usage

sample_correlated_gnp(
  old.graph,
  corr,
  p = edge_density(old.graph),
  permutation = NULL
)

Arguments

old.graph

The original graph.

corr

A scalar in the unit interval, the target Pearson correlation between the adjacency matrices of the original and the generated graph (the adjacency matrix being used as a vector).

p

A numeric scalar, the probability of an edge between two vertices, it must in the open (0,1) interval. The default is the empirical edge density of the graph. If you are resampling an Erdos-Renyi graph and you know the original edge probability of the Erdos-Renyi model, you should supply that explicitly.

permutation

A numeric vector, a permutation vector that is applied on the vertices of the first graph, to get the second graph. If NULL, the vertices are not permuted.

Value

An unweighted graph of the same size as old.graph such that the correlation coefficient between the entries of the two adjacency matrices is corr. Note each pair of corresponding matrix entries is a pair of correlated Bernoulli random variables.

Details

Please see the reference given below.

References

Lyzinski, V., Fishkind, D. E., Priebe, C. E. (2013). Seeded graph matching for correlated Erdos-Renyi graphs. https://arxiv.org/abs/1304.7844

See also

Examples

g <- sample_gnp(1000, .1)
g2 <- sample_correlated_gnp(g, corr = 0.5)
cor(as.vector(g[]), as.vector(g2[]))
#> [1] 0.4990424
g
#> IGRAPH 9645135 U--- 1000 49932 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 9645135:
#>  [1]  1-- 2  3-- 4  1-- 5  2-- 6  7--11  1--12  9--13 12--13 10--14  7--16
#> [11]  8--16 11--16  5--17  6--17  8--17  9--17 14--17  9--18  4--19 14--19
#> [21] 13--20  2--21  9--21 15--21  7--22 11--22  3--23 18--23 16--24 14--25
#> [31] 22--26  5--27 11--27 16--27 27--28 22--29 15--30 29--30  5--31 26--31
#> [41]  2--32  4--32  8--32 14--32  5--33  8--33  1--34  9--34 32--34 22--35
#> [51] 25--35 26--35  4--36 18--36  5--37  8--37 12--37  2--39 15--39 23--39
#> [61] 33--39 34--39  6--40 10--40 19--40 21--40 31--40  8--41 12--41 16--41
#> [71] 23--41 24--41 27--41 40--41 13--42 32--42  1--43  7--43  8--43 15--43
#> + ... omitted several edges
g2
#> IGRAPH f9cccd1 U--- 1000 49818 -- Correlated random graph
#> + attr: name (g/c), corr (g/n), p (g/n)
#> + edges from f9cccd1:
#>  [1]  3-- 6  4-- 7  4--10  2--13 12--13 10--14  8--15  3--16  6--16  7--16
#> [11]  8--16 15--16  6--17 14--17  9--18 14--19  2--21  9--21 15--21 11--22
#> [21]  1--23  3--23  5--23 18--23 19--23 12--25 14--25  2--26 19--26 22--26
#> [31] 16--27 13--29 12--30  8--31 15--31 26--31  2--32  4--32  8--32 25--32
#> [41] 31--32  5--33  8--33  1--34  9--34 33--34 24--35 25--35 33--35  4--36
#> [51] 18--36  8--37 11--37 12--37 20--37 25--37 34--37  1--38 11--38 18--38
#> [61]  2--39  5--39 23--39  9--40 10--40 19--40 21--40 27--40 29--40 31--40
#> [71] 34--40 23--41 24--41 36--41 40--41  1--42 13--42 32--42 38--42  1--43
#> + ... omitted several edges