--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.3 --- {% raw %}

sample_correlated_gnp_pair {igraph}	R Documentation

Sample a pair of correlated G(n,p) random graphs

Description

Sample a new graph by perturbing the adjacency matrix of a given graph and shuffling its vertices.

Usage

sample_correlated_gnp_pair(n, corr, p, directed = FALSE, permutation = NULL)

Arguments

n
Numeric scalar, the number of vertices for the sampled graphs.

corr
A scalar in the unit interval, the target Pearson correlation between the adjacency matrices of the original the generated graph (the adjacency matrix being used as a vector).

p
A numeric scalar, the probability of an edge between two vertices, it must in the open (0,1) interval.

directed
Logical scalar, whether to generate directed graphs.

permutation
A numeric vector, a permutation vector that is applied on the vertices of the first graph, to get the second graph. If NULL, the vertices are not permuted.

Details

Please see the reference given below.

Value

A list of two igraph objects, named graph1 and graph2, which are two graphs whose adjacency matrix entries are correlated with corr.

References

Lyzinski, V., Fishkind, D. E., Priebe, C. E. (2013). Seeded graph matching for correlated Erdos-Renyi graphs. https://arxiv.org/abs/1304.7844

See Also

sample_correlated_gnp, sample_gnp.

Examples

gg <- sample_correlated_gnp_pair(n = 10, corr = .8, p = .5,
           directed = FALSE)
gg
cor(as.vector(gg[[1]][]), as.vector(gg[[2]][]))

[Package igraph version 1.3.3 Index]
{% endraw %}