A graph with subgraphs that are each a random graph.

Create a number of Erdős-Rényi random graphs with identical parameters, and connect them with the specified number of edges.

Usage

sample_islands(islands.n, islands.size, islands.pin, n.inter)

Arguments

islands.n

The number of islands in the graph.

islands.size

The size of islands in the graph.

islands.pin

The probability to create each possible edge into each island.

n.inter

The number of edges to create between two islands.

Value

An igraph graph.

Examples

g <- sample_islands(3, 10, 5/10, 1)
oc <- cluster_optimal(g)
oc

See also

Author

Samuel Thiriot

Related documentation in the C library

igraph_simple_interconnected_islands_game().