Sample stochastic block model

Sampling from the stochastic block model of networks

Usage

sample_sbm(n, pref.matrix, block.sizes, directed = FALSE, loops = FALSE)

sbm(...)

Arguments

n

Number of vertices in the graph.

pref.matrix

The matrix giving the Bernoulli rates. This is a \(K\times K\) matrix, where \(K\) is the number of groups. The probability of creating an edge between vertices from groups \(i\) and \(j\) is given by element \((i,j)\). For undirected graphs, this matrix must be symmetric.

block.sizes

Numeric vector giving the number of vertices in each group. The sum of the vector must match the number of vertices.

directed

Logical scalar, whether to generate a directed graph.

loops

Logical scalar, whether self-loops are allowed in the graph.

...

Passed to sample_sbm().

Value

An igraph graph.

Details

This function samples graphs from a stochastic block model by (doing the equivalent of) Bernoulli trials for each potential edge with the probabilities given by the Bernoulli rate matrix, pref.matrix. The order of the vertices in the generated graph corresponds to the block.sizes argument.

References

Faust, K., & Wasserman, S. (1992a). Blockmodels: Interpretation and evaluation. Social Networks, 14, 5--61.

See also

Author

Gabor Csardi csardi.gabor@gmail.com

Examples

## Two groups with not only few connection between groups
pm <- cbind(c(.1, .001), c(.001, .05))
g <- sample_sbm(1000, pref.matrix = pm, block.sizes = c(300, 700))
g
#> IGRAPH 2477d6a U--- 1000 16799 -- Stochastic block model
#> + attr: name (g/c), loops (g/l)
#> + edges from 2477d6a:
#>  [1]  1-- 7  1-- 8  5-- 8  6-- 9  5--10  1--11 10--11  5--12  1--14  4--14
#> [11] 10--14  7--15 12--15  3--17  7--17 13--17 14--17 16--17  2--18  3--18
#> [21]  8--18 11--20 13--20 19--20  3--21  3--22 11--22 14--22 17--22 21--22
#> [31] 15--23 17--23 11--24 21--24 19--25 19--26 25--26  2--27 12--27 10--28
#> [41]  2--29  5--29 16--29  4--31 11--31 20--31  7--32 25--32 13--33 19--33
#> [51]  3--34 24--34 28--34  6--35 10--35 12--35 15--35 26--35  1--36 18--36
#> [61]  2--37  7--37 12--37 20--37 26--37 35--37  3--38 34--38 12--39 16--39
#> [71] 21--39 35--39 26--40 36--40  4--41  8--41  9--41 16--41 39--41 10--42
#> + ... omitted several edges