Sample vectors uniformly from the volume of a sphere

Sample finite-dimensional vectors to use as latent position vectors in random dot product graphs

Usage

sample_sphere_volume(dim, n = 1, radius = 1, positive = TRUE)

Arguments

dim

Integer scalar, the dimension of the random vectors.

n

Integer scalar, the sample size.

radius

Numeric scalar, the radius of the sphere to sample.

positive

Logical scalar, whether to sample from the positive orthant of the sphere.

Value

A dim (length of the alpha vector for sample_dirichlet()) times n matrix, whose columns are the sample vectors.

Details

sample_sphere_volume() generates uniform samples from \(S^{dim-1}\) (the (dim-1)-sphere) i.e. the Euclidean norm of the samples is smaller or equal to radius.

See also

Examples

lpvs.sph.vol <- sample_sphere_volume(dim = 10, n = 20, radius = 1)
RDP.graph.4 <- sample_dot_product(lpvs.sph.vol)
vec.norm <- apply(lpvs.sph.vol, 2, function(x) {
  sum(x^2)
})
vec.norm
#>  [1] 0.8933667 0.9330896 0.4992353 0.8503588 0.9729300 0.8623218 0.9949072
#>  [8] 0.4118385 0.7996143 0.7803815 0.9119889 0.9199697 0.7551018 0.7104507
#> [15] 0.8701593 0.8874671 0.9594074 0.8095282 0.5046112 0.9943260