Scan statistics on a time series of graphs

Calculate scan statistics on a time series of graphs. This is done by calculating the local scan statistics for each graph and each vertex, and then normalizing across the vertices and across the time steps.

Usage

scan_stat(graphs, tau = 1, ell = 0, locality = c("us", "them"), ...)

Arguments

graphs

A list of igraph graph objects. They must be all directed or all undirected and they must have the same number of vertices.

tau

The number of previous time steps to consider for the time-dependent normalization for individual vertices. In other words, the current locality statistics of each vertex will be compared to this many previous time steps of the same vertex to decide whether it is significantly larger.

ell

The number of previous time steps to consider for the aggregated scan statistics. This is essentially a smoothing parameter.

locality

Whether to calculate the ‘us’ or ‘them’ statistics.

...

Extra arguments are passed to local_scan().

Value

A list with entries:

stat

The scan statistics in each time step. It is NA for the initial tau + ell time steps.

arg_max_v

The (numeric) vertex ids for the vertex with the largest locality statistics, at each time step. It is NA for the initial tau + ell time steps.

See also

Examples

## Generate a bunch of SBMs, with the last one being different
num_t <- 20
block_sizes <- c(10, 5, 5)
p_ij <- list(p = 0.1, h = 0.9, q = 0.9)

P0 <- matrix(p_ij$p, 3, 3)
P0[2, 2] <- p_ij$h
PA <- P0
PA[3, 3] <- p_ij$q
num_v <- sum(block_sizes)

tsg <- replicate(num_t - 1, P0, simplify = FALSE) %>%
  append(list(PA)) %>%
  lapply(sample_sbm, n = num_v, block.sizes = block_sizes, directed = TRUE)

scan_stat(graphs = tsg, k = 1, tau = 4, ell = 2)
#> $stat
#>  [1]         NA         NA         NA         NA         NA         NA
#>  [7] -1.3349739  0.3334680 -0.3218619  3.0318257 -0.6578531 -0.4941383
#> [13]  9.6805554 -0.9740239 -0.4819492  0.8392918  0.6188672  0.4822330
#> [19]  7.2846250  1.0826482
#> 
#> $arg_max_v
#>  [1] NA NA NA NA NA NA  6 10  6  9  1 14 20  5  9  8  7  1 19 18
#> 
scan_stat(graphs = tsg, locality = "them", k = 1, tau = 4, ell = 2)
#> $stat
#>  [1]         NA         NA         NA         NA         NA         NA
#>  [7] -0.3299832  0.0000000  1.1722769  2.5244679  0.1267890 -0.6250000
#> [13] -0.8750000 -2.4395216  0.7071068  1.0496107 -1.0000000 -0.3750000
#> [19]  3.3750000  1.0117066
#> 
#> $arg_max_v
#>  [1] NA NA NA NA NA NA  1  8 15  9  1 16 18  5 16 15 16  1 14 19
#>