--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.1 --- {% raw %}

scg_eps {igraph}	R Documentation

Error of the spectral coarse graining (SCG) approximation

Description

scg_eps computes \Vert v_i-Pv_i\Vert, where v_i is the ith eigenvector in V and P is the projector corresponding to the mtype argument.

Usage

scg_eps(
  V,
  groups,
  mtype = c("symmetric", "laplacian", "stochastic"),
  p = NULL,
  norm = c("row", "col")
)

Arguments

V
A numeric matrix of (eigen)vectors assumed normalized. The vectors are to be stored column-wise in V).

groups
A vector of nrow(V) integers labeling each group vertex in the partition.

mtype
The type of semi-projector used for the SCG. For now “symmetric”, “laplacian” and “stochastic” are available.

p
A probability vector of length nrow(V). p is the stationary probability distribution of a Markov chain when mtype = “stochastic”. This parameter is ignored otherwise.

norm
Either “row” or “col”. If set to “row” the rows of the Laplacian matrix sum to zero and the rows of the stochastic matrix sum to one; otherwise it is the columns.

Value

scg_eps returns with a numeric vector whose ith component is \Vert v_i-Pv_i\Vert (see Details).

Author(s)

David Morton de Lachapelle, http://people.epfl.ch/david.morton.

References

D. Morton de Lachapelle, D. Gfeller, and P. De Los Rios, Shrinking Matrices while Preserving their Eigenpairs with Application to the Spectral Coarse Graining of Graphs. Submitted to SIAM Journal on Matrix Analysis and Applications, 2008. http://people.epfl.ch/david.morton

See Also

scg-method and scg.

Examples

v <- rexp(20)
km <- kmeans(v,5)
sum(km$withinss)
scg_eps(cbind(v), km$cluster)^2

[Package igraph version 1.3.1 Index]
{% endraw %}