similarity {igraph}	R Documentation

Similarity measures of two vertices

Description

These functions calculates similarity scores for vertices based on their connection patterns.

Usage

similarity.jaccard(graph, vids = V(graph), mode = c("all", "out", "in",
      "total"), loops = FALSE) 
similarity.dice(graph, vids = V(graph), mode = c("all", "out", "in",
      "total"), loops = FALSE)
similarity.invlogweighted(graph, vids = V(graph),
       mode = c("all", "out", "in", "total"))

Arguments

graph 	 The input graph.

vids 	 The vertex ids for which the similarity is calculated.

mode 	 The type of neighboring vertices to use for the calculation, possible values: ‘out’, ‘in’, ‘all’.

loops 	 Whether to include vertices themselves in the neighbor sets.

Details

The Jaccard similarity coefficient of two vertices is the number of common neighbors divided by the number of vertices that are neighbors of at least one of the two vertices being considered. similarity.jaccard calculates the pairwise Jaccard similarities for some (or all) of the vertices.

The Dice similarity coefficient of two vertices is twice the number of common neighbors divided by the sum of the degrees of the vertices. similarity.dice calculates the pairwise Dice similarities for some (or all) of the vertices.

The inverse log-weighted similarity of two vertices is the number of their common neighbors, weighted by the inverse logarithm of their degrees. It is based on the assumption that two vertices should be considered more similar if they share a low-degree common neighbor, since high-degree common neighbors are more likely to appear even by pure chance. Isolated vertices will have zero similarity to any other vertex. Self-similarities are not calculated. See the following paper for more details: Lada A. Adamic and Eytan Adar: Friends and neighbors on the Web. Social Networks, 25(3):211-230, 2003.

Value

A length(vids) by length(vids) numeric matrix containing the similarity scores.

Author(s)

Tamas Nepusz ntamas@rmki.kfki.hu and Gabor Csardi csardi@rmki.kfki.hu for the manual page.

References

Lada A. Adamic and Eytan Adar: Friends and neighbors on the Web. Social Networks, 25(3):211-230, 2003.

See Also

cocitation and bibcoupling

Examples

g <- graph.ring(5)
similarity.dice(g)
similarity.jaccard(g)

[Package igraph version 0.5.2-2 Index]