simplify {igraph}	R Documentation

Simple graphs

Description

Simple graphs are graphs which do not contain loop and multiple edges.

Usage

simplify(graph, remove.multiple = TRUE, remove.loops = TRUE)
is.simple(graph)

Arguments

graph 	 The graph to work on.

remove.loops 	 Logical, whether the loop edges are to be removed.

remove.multiple 	 Logical, whether the multiple edges are to be removed.

Details

A loop edge is an edge for which the two endpoints are the same vertex. Two edges are multiple edges if they have exactly the same two endpoints (for directed graphs order does matter). A graph is simple is it does not contain loop edges and multiple edges.

is.simple checks whether a graph is simple.

simplify removes the loop and/or multiple edges from a graph. If both remove.loops and remove.multiple are TRUE the function returns a simple graph.

Value

A new graph object with the edges deleted.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

is.loop, is.multiple and count.multiple, delete.edges, delete.vertices

Examples

g <- graph( c(1,2,1,2,3,3) )
is.simple(g)
is.simple(simplify(g, remove.loops=FALSE))
is.simple(simplify(g, remove.multiple=FALSE))
is.simple(simplify(g))

[Package igraph version 0.5.2-2 Index]