--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.3 --- {% raw %}

simplify {igraph}	R Documentation

Simple graphs

Description

Simple graphs are graphs which do not contain loop and multiple edges.

Usage

simplify(
  graph,
  remove.multiple = TRUE,
  remove.loops = TRUE,
  edge.attr.comb = igraph_opt("edge.attr.comb")
)

is_simple(graph)

simplify_and_colorize(graph)

Arguments

graph
The graph to work on.

remove.multiple
Logical, whether the multiple edges are to be removed.

remove.loops
Logical, whether the loop edges are to be removed.

edge.attr.comb
Specifies what to do with edge attributes, if remove.multiple=TRUE. In this case many edges might be mapped to a single one in the new graph, and their attributes are combined. Please see attribute.combination for details on this.

Details

A loop edge is an edge for which the two endpoints are the same vertex. Two edges are multiple edges if they have exactly the same two endpoints (for directed graphs order does matter). A graph is simple is it does not contain loop edges and multiple edges.

is_simple checks whether a graph is simple.

simplify removes the loop and/or multiple edges from a graph. If both remove.loops and remove.multiple are TRUE the function returns a simple graph.

simplify_and_colorize constructs a new, simple graph from a graph and also sets a color attribute on both the vertices and the edges. The colors of the vertices represent the number of self-loops that were originally incident on them, while the colors of the edges represent the multiplicities of the same edges in the original graph. This allows one to take into account the edge multiplicities and the number of loop edges in the VF2 isomorphism algorithm. Other graph, vertex and edge attributes from the original graph are discarded as the primary purpose of this function is to facilitate the usage of multigraphs with the VF2 algorithm.

Value

a new graph object with the edges deleted.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

which_loop, which_multiple and count_multiple, delete_edges, delete_vertices

Examples

g <- graph( c(1,2,1,2,3,3) )
is_simple(g)
is_simple(simplify(g, remove.loops=FALSE))
is_simple(simplify(g, remove.multiple=FALSE))
is_simple(simplify(g))

[Package igraph version 1.3.3 Index]
{% endraw %}