tkplot {igraph}	R Documentation

Interactive plotting of graphs

Description

tkplot and its companion functions serve as an interactive graph drawing facility. Not all parameters of the plot can be changed interactively right now though, eg. the colors of vertices, edges, and also others have to be pre-defined.

Usage

tkplot(graph, canvas.width=450, canvas.height=450, ...)

tkplot.close(tkp.id, window.close = TRUE) 
tkplot.off()
tkplot.fit.to.screen(tkp.id, width = NULL, height = NULL)
tkplot.reshape(tkp.id, newlayout, ...)
tkplot.export.postscript(tkp.id)
tkplot.getcoords(tkp.id, norm = FALSE)
tkplot.center(tkp.id)
tkplot.rotate(tkp.id, degree = NULL, rad = NULL)

Arguments

graph 	 The graph to plot.

canvas.width,canvas.height 	 The size of the tkplot drawing area.

tkp.id 	 The id of the tkplot window to close/reshape/etc.

window.close 	 Leave this on the default value.

width 	 The width of the rectangle for generating new coordinates.

height 	 The height of the rectangle for generating new coordinates.

newlayout 	 The new layout, see the layout parameter of tkplot.

norm 	 Logical, should we norm the coordinates.

degree 	 The degree to rotate the plot.

rad 	 The degree to rotate the plot, in radian.

... 	 Not used right now.

Details

tkplot is an interactive graph drawing facility. It is not very well developed at this stage, but it should be still useful.

It's handling should be quite straightforward most of the time, here are some remarks and hints.

There are different popup menus, activated by the right mouse button, for vertices and edges. Both operate on the current selection if the vertex/edge under the cursor is part of the selection and operate on the vertex/edge under the cursor if it is not.

One selection can be active at a time, either a vertex or an edge selection. A vertex/edge can be added to a selection by holding the control key while clicking on it with the left mouse button. Doing this again deselect the vertex/edge.

Selections can be made also from the Select menu. The `Select some vertices' dialog allows to give an expression for the vertices to be selected: this can be a list of numeric R expessions separated by commas, like `1,2:10,12,14,15' for example. Similarly in the `Select some edges' dialog two such lists can be given and all edges connecting a vertex in the first list to one in the second list will be selected.

In the color dialog a color name like 'orange' or RGB notation can also be used.

The tkplot command creates a new Tk window with the graphical representation of graph. The command returns an integer number, the tkplot id. The other commands utilize this id to be able to query or manipulate the plot.

tkplot.close closes the Tk plot with id tkp.id.

tkplot.off closes all Tk plots.

tkplot.fit.to.screen fits the plot to the given rectange (width and height), if some of these are NULL the actual phisical width od height of the plot window is used.

tkplot.reshape applies a new layout to the plot, its optional parameters will be collected to a list analogous to layout.par.

tkplot.export.postscript creates a dialog window for saving the plot in postscript format.

tkplot.getcoords returns the coordinates of the vertices in a matrix. Each row corresponds to one vertex.

tkplot.center shifts the figure to the center of its plot window.

tkplot.rotate rotates the figure, its parameter can be given either in degrees or in radians.

Value

tkplot returns an integer, the id of the plot, this can be used to manipulate it from the command line.
tkplot.getcoords returns a matrix with the coordinates.
tkplot.close, tkplot.off, tkplot.fit.to.screen, tkplot.reshape, tkplot.export.postscript, tkplot.center and tkplot.rotate return NULL invisibly.

Author(s)

Gabor Csardi csardi@rmki.kfki.hu

See Also

plot.igraph, layout

Examples

g <- graph.ring(10)
## Not run: tkplot(g)

[Package igraph version 0.5.2-2 Index]