--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.4 --- {% raw %}

union.igraph {igraph}	R Documentation

Union of graphs

Description

The union of two or more graphs are created. The graphs may have identical or overlapping vertex sets.

Usage

## S3 method for class 'igraph'
union(..., byname = "auto")

Arguments

...
Graph objects or lists of graph objects.

byname
A logical scalar, or the character scalar auto. Whether to perform the operation based on symbolic vertex names. If it is auto, that means TRUE if all graphs are named and FALSE otherwise. A warning is generated if auto and some (but not all) graphs are named.

Details

union creates the union of two or more graphs. Edges which are included in at least one graph will be part of the new graph. This function can be also used via the %u% operator.

If the byname argument is TRUE (or auto and all graphs are named), then the operation is performed on symbolic vertex names instead of the internal numeric vertex ids.

union keeps the attributes of all graphs. All graph, vertex and edge attributes are copied to the result. If an attribute is present in multiple graphs and would result a name clash, then this attribute is renamed by adding suffixes: _1, _2, etc.

The name vertex attribute is treated specially if the operation is performed based on symbolic vertex names. In this case name must be present in all graphs, and it is not renamed in the result graph.

An error is generated if some input graphs are directed and others are undirected.

Value

A new graph object.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

Examples

## Union of two social networks with overlapping sets of actors
net1 <- graph_from_literal(D-A:B:F:G, A-C-F-A, B-E-G-B, A-B, F-G,
                  H-F:G, H-I-J)
net2 <- graph_from_literal(D-A:F:Y, B-A-X-F-H-Z, F-Y)
print_all(net1 %u% net2)

[Package igraph version 1.3.4 Index]
{% endraw %}