--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.2 --- {% raw %}

Pie charts as vertices {igraph}	R Documentation

Using pie charts as vertices in graph plots

Description

More complex vertex images can be used to express addtional information about vertices. E.g. pie charts can be used as vertices, to denote vertex classes, fuzzy classification of vertices, etc.

Details

The vertex shape ‘pie’ makes igraph draw a pie chart for every vertex. There are some extra graphical vertex parameters that specify how the pie charts will look like:

pie

Numeric vector, gives the sizes of the pie slices.

pie.color

A list of color vectors to use for the pies. If it is a list of a single vector, then this is used for all pies. It the color vector is shorter than the number of areas in a pie, then it is recycled.

pie.border

The color of the border line of the pie charts, in the same format as pie.color.

pie.angle

The slope of shading lines, given as an angle in degrees (counter-clockwise).

pie.density

The density of the shading lines, in lines per inch. Non-positive values inhibit the drawing of shading lines.

pie.lty

The line type of the border of the slices.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

igraph.plotting, plot.igraph

Examples

g <- make_ring(10)
values <- lapply(1:10, function(x) sample(1:10,3))
if (interactive()) {
  plot(g, vertex.shape="pie", vertex.pie=values,
       vertex.pie.color=list(heat.colors(5)),
       vertex.size=seq(10,30,length.out=10), vertex.label=NA)
}

[Package igraph version 1.3.2 Index]
{% endraw %}