--- layout: r-manual title: igraph R manual pages mainheader: R igraph manual pages lead: Use this if you are using igraph from R vmenu: true doctype: html/ langversion: 1.3.3 --- {% raw %}

which_mutual {igraph}	R Documentation

Find mutual edges in a directed graph

Description

This function checks the reciprocal pair of the supplied edges.

Usage

which_mutual(graph, eids = E(graph))

Arguments

graph
The input graph.

eids
Edge sequence, the edges that will be probed. By default is includes all edges in the order of their ids.

Details

In a directed graph an (A,B) edge is mutual if the graph also includes a (B,A) directed edge.

Note that multi-graphs are not handled properly, i.e. if the graph contains two copies of (A,B) and one copy of (B,A), then these three edges are considered to be mutual.

Undirected graphs contain only mutual edges by definition.

Value

A logical vector of the same length as the number of edges supplied.

Author(s)

Gabor Csardi csardi.gabor@gmail.com

See Also

reciprocity, dyad_census if you just want some statistics about mutual edges.

Examples

g <- sample_gnm(10, 50, directed=TRUE)
reciprocity(g)
dyad_census(g)
which_mutual(g)
sum(which_mutual(g))/2 == dyad_census(g)$mut

[Package igraph version 1.3.3 Index]
{% endraw %}